### 2024-11-02 #667 @adamhsparks
read.abares - Provides simple downloading, parsing and importing of Australian Bureau of Agricultural and Resource Economics and Sciences (ABARES) data sources
Submitting Author Name: Adam Sparks
Submitting Author Github Handle: @adamhsparks
Repository:  https://github.com/adamhsparks/read.abares
Version submitted: 0.1.0
Submission type: Standard
Version accepted: TBD

---

```
Package: read.abares
Title: Provides simple downloading, parsing and importing of Australian
    Bureau of Agricultural and Resource Economics and Sciences (ABARES)
    data sources
Version: 0.1.0
    person("Adam H.", "Sparks", , "adamhsparks@gmail.com", role = c("cre", "aut"),
Description: Download and import data from the Australian Bureau of
    Agricultural and Resource Economics and Sciences (ABARES)
    <https://www.agriculture.gov.au/abares>.
    https://adamhsparks.github.io/read.abares/
    cli,
    curl,
    openxlsx2,
    readtext,
    sf,
    stars,
    terra,
    tidync,
    withr
    pander,
    roxyglobals,
    knitr
Config/roxyglobals/filename: globals.R
Config/roxyglobals/unique: FALSE
Roxygen: list(markdown = TRUE, roclets = c("collate", "namespace", "rd",
    "roxyglobals::global_roclet"))
```

Provides workflow assistance for fetching freely available data from the Australian Bureau of Agricultural and Resource Economics and Sciences (ABARES), the science and economics research division of the Department of Agriculture, Fisheries and Forestry. Data include trade information, soil depth, agricultural survey data on production and model outputs of agricultural production useful for agricultural researchers.

    - Ag economists
    - Biometricians working in agriculture
    - Anyone else who is interested in Australian agricultural and economic data

None that I'm aware of.

NA
