+++ README.Rmd
@@ -25,7 +25,7 @@ knitr::opts_chunk$set(
 An R package for automated downloading and ingestion of data from the Australian Bureau of Agricultural and Resource Economics and Sciences.
 Not all ABARES data are serviced.
 The list is hand-picked to be reasonably useful and maintainable, _i.e._, frequently updated values are not included in this, _e.g._, [Australian crop reports](https://daff.ent.sirsidynix.net.au/client/en_AU/ABARES/search/results?te=ASSET&st=PD#).
-However, if there is a data set that you feel would be useful to be serviced by {read.abares}, please feel free to [open an issue](https://codeberg.org/adamhsparks/read.abares/issues/new) with details about the data set or better yet, open a pull request!
+However, if there is a data set that you feel would be useful to be serviced by {read.abares}, please feel free to [open an issue](https://github.com/adamhsparks/read.abares/issues/new) with details about the data set or better yet, open a pull request!
 
 Data serviced include:
 
@@ -56,7 +56,7 @@ But you can install it like so:
 ```r
 if (!require("remotes"))
   install.packages("remotes")
-remotes::install_git("https://codeberg.org/adamhsparks/read.abares")
+remotes::install_git("https://github.com/adamhsparks/read.abares")
 ```
 
 ## Features
@@ -92,7 +92,7 @@ I am considering including some pared down examples for tests as the data are re
 
 ## Metadata
 
-Please report any [issues or bugs](https://codeberg.org/adamhsparks/read.abares/issues).
+Please report any [issues or bugs](https://github.com/adamhsparks/read.abares/issues).
 
 License: [MIT](LICENSE.md)
 
@@ -109,4 +109,4 @@ citation("read.abares")
 
 ## Code of Conduct
 
-Please note that the {read.abares} project is released with a [Contributor Code of Conduct](https://adamhsparks.codeberg.page/read.abares/CODE_OF_CONDUCT.html). By contributing to this project, you agree to abide by its terms.
+Please note that the {read.abares} project is released with a [Contributor Code of Conduct](https://adamhsparks.github.io/read.abares/CODE_OF_CONDUCT.html). By contributing to this project, you agree to abide by its terms.
