### 2024-11-02 eea7adb
remove redundant text
