### 2024-11-02 d1bafaf
Use curl to download rather than read directly to use custom handle for dl files
