+++ R/get_agfd.R
@@ -64,7 +64,8 @@ get_agfd <- function(fixed_prices = TRUE, cache = TRUE) {
 
     curl::curl_download(url = url,
                         destfile = download_file,
-                        quiet = FALSE)
+                        quiet = FALSE,
+                        handle =  create_handle())
 
     withr::with_dir(download_dir,
                     utils::unzip(zipfile = download_file,
