### 2024-11-02 f9ecce6
Move code inside tests
