+++ tests/testthat/test-inspect_cache.R
@@ -1,9 +1,9 @@
 
 withr::local_envvar(R_USER_CACHE_DIR = file.path(tempdir()))
-dir.create(file.path(.find_user_cache()), recursive = TRUE)
 # check message with no files present ----
 
 test_that("inspect_cache() works w/ no files", {
+  dir.create(file.path(.find_user_cache()), recursive = TRUE)
   cli_out <- function() {
     return(cli::cli_inform(c(
       "There do not appear to be any files cached for {.pkg {{read.abares}}}."
@@ -14,8 +14,9 @@ test_that("inspect_cache() works w/ no files", {
 })
 
 # Now create a file to check when files are present ----
-test_file <- file.path(.find_user_cache(), "test.R")
+
 test_that("inspect_cache() works, recursive = FALSE", {
+  test_file <- file.path(.find_user_cache(), "test.R")
   file.create(test_file)
 
   f <- .find_user_cache()
@@ -25,6 +26,7 @@ test_that("inspect_cache() works, recursive = FALSE", {
 })
 
 test_that("inspect_cache() works, recursive = TRUE", {
+  test_file <- file.path(.find_user_cache(), "test.R")
   file.create(test_file)
 
   f <- .find_user_cache()
