+++ R/internal_functions.R
@@ -42,3 +42,30 @@ create_handle <- function() {
     http_version = 2
   )
 }
+
+#' Use httr2 to Fetch a File With Retries
+#'
+#' Retries to download the requested resource five times before stopping.  Then
+#'   saves the resource in the `tempdir()` for importing.
+#'
+#' @param url The URL being requested
+#' @param .f a filepath to be written to local storage
+#'
+#' @examples
+#' f <- file.path(tempdir(), "fdp-beta-national-historical.csv")
+#' retry_download(url = "https://www.agriculture.gov.au/sites/default/files/documents/fdp-beta-national-historical.csv",
+#' .f = f)
+#' @return Called for its side-effects, writes an object to the `tempdir()` for
+#'   reading into the active \R session later.
+#' @keywords Internal
+#' @noRd
+
+.retry_download <- function(url, .f) {
+
+  response <- httr2::request(
+    base_url = url) |>
+  httr2::req_retry(max_tries = 5) |>
+  httr2::req_perform()
+
+  writeBin(response$body, con = f)
+}
