### 2024-11-03 9e22991
Fix bug in .retry_download(), delete unused fn
