+++ R/internal_functions.R
@@ -28,21 +28,6 @@
   }
 }
 
-#' Create curl Handles
-#'
-#' @return a \CRANpkg{curl} handle for use in downloading files
-#' @noRd
-#' @keywords Internal
-create_handle <- function() {
-  h <- curl::new_handle()
-  curl::handle_setopt(
-    handle = h,
-    TCP_KEEPALIVE = 200000,
-    CONNECTTIMEOUT = 90,
-    http_version = 2
-  )
-}
-
 #' Use httr2 to Fetch a File With Retries
 #'
 #' Retries to download the requested resource five times before stopping.  Then
@@ -67,5 +52,5 @@ create_handle <- function() {
   httr2::req_retry(max_tries = 5) |>
   httr2::req_perform()
 
-  writeBin(response$body, con = f)
+  writeBin(response$body, con = .f)
 }
