### 2024-11-03 8b6ac16
Remove LazyData from DESCRIPTION, data isn't used
