### 2024-11-03 debc77e
uses httr2 for all downloads, fixes several bugs
