+++ man/get_abares_trade_regions.Rd
@@ -3,6 +3,9 @@
 \name{get_abares_trade_regions}
 \alias{get_abares_trade_regions}
 \title{Get ABARES Trade Data Regions From the ABARES Trade Dashboard}
+\source{
+\url{https://daff.ent.sirsidynix.net.au/client/en_AU/search/asset/1033841/0}
+}
 \usage{
 get_abares_trade_regions(cache = TRUE)
 }
@@ -28,9 +31,6 @@ trade_regions <- get_abares_trade_regions()
 
 trade_regions
 \dontshow{\}) # examplesIf}
-}
-\references{
-\url{https://daff.ent.sirsidynix.net.au/client/en_AU/search/asset/1033841/0}
 }
 \seealso{
 Other Trade: 
