### 2024-11-03 4ebc04b
Try limiting http version to 1.1
