+++ R/internal_functions.R
@@ -37,9 +37,11 @@
 #' @param .f a filepath to be written to local storage
 #'
 #' @examples
+#'
 #' f <- file.path(tempdir(), "fdp-beta-national-historical.csv")
 #' retry_download(url = "https://www.agriculture.gov.au/sites/default/files/documents/fdp-beta-national-historical.csv",
 #' .f = f)
+#'
 #' @return Called for its side-effects, writes an object to the `tempdir()` for
 #'   reading into the active \R session later.
 #' @keywords Internal
@@ -48,8 +50,8 @@
 .retry_download <- function(url, .f) {
 
   response <- httr2::request(base_url = url) |>
-    httr2::req_retry(max_tries = 5) |>
-    httr2::req_options(http_version = 2) |>
+    httr2::req_options(http_version = 2, timeout = 120) |>
+    httr2::req_retry(max_tries = 10) |>
     httr2::req_perform()
 
   writeBin(response$body, con = .f)
