### 2024-11-04 0e93d98
use tryCatch() to retry failed downloads for CI
