### 2024-11-04 5c0b649
Fix cli alerts
