### 2024-11-04 3847ba2
Update internal_functions.R
