### 2024-11-05 0b01dce
format line lengths and fix documentation of internal functions
