### 2024-11-05 aaed972
Check for abares_trade.zip in tempdir() since no longer unzipping
