### 2024-11-05 d643c8d
Use cli_abort for failed downloads
