### 2024-11-05 f28bed5
correct typo in cli  messages
