### 2024-11-05 304eb0b
fix warnings, decrease tries and dleays for download
