### 2024-11-05 3ac39f9
skip tests on CI that require download
