+++ tests/testthat/test-get_abares_trade_regions.R
@@ -6,6 +6,7 @@ withr::local_envvar(R_USER_CACHE_DIR = file.path(tempdir(), "abares.cache.1"))
 
 test_that("get_abares_trade_regions doesn't cache", {
   skip_if_offline()
+  skip_on_ci()
   x <- get_abares_trade_regions(cache = FALSE)
   expect_s3_class(x, c("data.table", "data.frame"))
   expect_named(x,
@@ -35,6 +36,7 @@ test_that("get_abares_trade_regions doesn't cache", {
 
 test_that("get_abares_trade_regions skips downloading if still in tempdir()", {
   skip_if_offline()
+  skip_on_ci()
   x <- .check_existing_trade_regions(cache = FALSE)
   expect_s3_class(x, c("data.table", "data.frame"))
 })
@@ -43,6 +45,7 @@ test_that("get_abares_trade_regions skips downloading if still in tempdir()", {
 
 test_that("get_abares_trade_regions caches", {
   skip_if_offline()
+  skip_on_ci()
   x <- get_abares_trade_regions(cache = TRUE)
   expect_s3_class(x, c("data.table", "data.frame"))
   y <- list.files(file.path(.find_user_cache(), "abares_trade_dir"))
@@ -56,6 +59,7 @@ test_that("get_abares_trade_regions caches", {
 
 test_that("get_abares_trade_regions skips downloading if cache is available", {
   skip_if_offline()
+  skip_on_ci()
   x <- .check_existing_trade_regions(cache = TRUE)
   expect_s3_class(x, c("data.table", "data.frame"))
 })
