+++ tests/testthat/test-get_historical_regional_estimates.R
@@ -1,6 +1,7 @@
 
 test_that("get_historical_regional_estimates works", {
   skip_if_offline()
+  skip_on_ci()
   x <- get_historical_regional_estimates()
   expect_named(x, c("Variable", "Year", "ABARES_region", "Value", "RSE"))
   expect_s3_class(x, c("data.table", "data.frame"))
