+++ tests/testthat/test-get_abares_trade.R
@@ -6,6 +6,7 @@ withr::local_envvar(R_USER_CACHE_DIR = file.path(tempdir(), "abares.cache.1"))
 
 test_that("get_abares_trade doesn't cache", {
   skip_if_offline()
+  skip_on_ci()
   x <- get_abares_trade(cache = FALSE)
   expect_s3_class(x, c("data.table", "data.frame"))
   expect_named(
@@ -56,6 +57,7 @@ test_that("get_abares_trade doesn't cache", {
 
 test_that("get_abares_trade skips downloading if still in tempdir()", {
   skip_if_offline()
+  skip_on_ci()
   x <- .check_existing_trade_regions(cache = FALSE)
   expect_s3_class(x, c("data.table", "data.frame"))
 })
@@ -64,6 +66,7 @@ test_that("get_abares_trade skips downloading if still in tempdir()", {
 
 test_that("get_abares_trade caches", {
   skip_if_offline()
+  skip_on_ci()
   x <- get_abares_trade(cache = TRUE)
   expect_s3_class(x, c("data.table", "data.frame"))
   y <- list.files(file.path(.find_user_cache(), "abares_trade_dir"))
@@ -77,6 +80,7 @@ test_that("get_abares_trade caches", {
 
 test_that("get_abares_trade skips downloading if cache is available", {
   skip_if_offline()
+  skip_on_ci()
   x <- .check_existing_trade_regions(cache = TRUE)
   expect_s3_class(x, c("data.table", "data.frame"))
 })
