+++ tests/testthat/test-read_agfd_terra.R
@@ -1,6 +1,7 @@
 
 test_that("read_agfd_terra() returns a terra object", {
   skip_if_offline()
+  skip_on_ci()
   x <- get_agfd() |>
     read_agfd_terra()
 
