+++ tests/testthat/test-read_agfd_dt.R
@@ -1,5 +1,6 @@
 test_that("read_agfd_dt returns a data.table object", {
   skip_if_offline()
+  skip_on_ci()
   x <- get_agfd() |>
     read_agfd_dt()
 
