### 2024-11-06 #667 @mpadge
@adamhsparks The tests on our machines just keep running for hours and hours and days even. Can you check on your side to make sure there's not some flag that works on GitHub machines to prevent this, yet might not stop tests on our machines?
