### 2024-11-17 2fcddbf
add reference and source for soil thickness docs
