+++ R/get_soil_thickness.R
@@ -27,7 +27,8 @@
 #'  the file path of the resulting \acronym{ESRI} Grid file and text file of
 #'  metadata
 #'
-#' @source <https://data.agriculture.gov.au/geonetwork/srv/eng/catalog.search#/metadata/faa9f157-8e17-4b23-b6a7-37eb7920ead6>
+#' @references <https://data.agriculture.gov.au/geonetwork/srv/eng/catalog.search#/metadata/faa9f157-8e17-4b23-b6a7-37eb7920ead6>
+#' @source <https://anrdl-integration-web-catalog-saxfirxkxt.s3-ap-southeast-2.amazonaws.com/warehouse/staiar9cl__059/staiar9cl__05911a01eg_geo___.zip>
 #' @autoglobal
 #' @family soil_thickness
 #' @export
