+++ man/get_soil_thickness.Rd
@@ -4,7 +4,7 @@
 \alias{get_soil_thickness}
 \title{Download Soil Thickness for Australian Areas of Intensive Agriculture of Layer 1}
 \source{
-\url{https://data.agriculture.gov.au/geonetwork/srv/eng/catalog.search#/metadata/faa9f157-8e17-4b23-b6a7-37eb7920ead6}
+\url{https://anrdl-integration-web-catalog-saxfirxkxt.s3-ap-southeast-2.amazonaws.com/warehouse/staiar9cl__059/staiar9cl__05911a01eg_geo___.zip}
 }
 \usage{
 get_soil_thickness(cache = TRUE)
@@ -42,6 +42,9 @@ library(pander)
 y <- x$metadata
 pander(y)
 \dontshow{\}) # examplesIf}
+}
+\references{
+\url{https://data.agriculture.gov.au/geonetwork/srv/eng/catalog.search#/metadata/faa9f157-8e17-4b23-b6a7-37eb7920ead6}
 }
 \seealso{
 Other soil_thickness: 
