### 2024-11-19 29ee5f0
fix a few more issues with trade region data download
