### 2024-11-19 399e3ce
fixes trade regions bugs
