### 2024-11-20 #667 @adamhsparks
@emilyriederer, the test coverage is >90%, see attached screenshot of local `covr::report()`.

However, there are issues with connectivity between GitHub CI and ABARES where GitHub just won't download some of the data. From what @mpadge and I have sussed out, it's a local issue (GitHub) and I can't really troubleshoot their setup. So I've opted to skip most of the tests on CI using `skip_on_ci()` for any function that downloads data. But I'm open to suggestions as to how to improve this. It's just one reason I've submitted it for review. 😃
