### 2024-11-23 #667 @emilyriederer
Thanks @adamhsparks !

It looks like `{covr}` also has options to explicitly exclude certain functions from code coverage calculations ([details here](https://covr.r-lib.org/reference/exclusions.html)). Would it make sense to use that logic just so our bot and the local review can have consistent results? (curious @mpadge thoughts also)

I'm fairly comfortable moving forward either way; just one thought to flesh out best practices here.
