### 2024-11-25 #667 @adamhsparks
Thanks, @emilyriederer, I'd prefer not to exclude these functions as using testing helped me squash a few bugs last week and there's no reason to exclude them in the package where I might normally consider excluding functions. This appears to be a technical limitation outside of the package that would hinder the package's quality if implemented.
