### 2024-11-25 #667 @mpadge
@emilyriederer I'm happy to accept Adam's explanation there. We do require coverage reports, but have no policies against locally-generated reports like used here. I think this would represent the first submission to do so, and so would maybe only ask @adamhsparks if you could have a look over the [relevant section of the dev guide](https://devguide.ropensci.org/pkg_building.html#testing) to see whether anything might be clarified or improved there?
