### 2024-11-25 #667 @mpadge
For non-binary files (your csvs and maybe more?), you can also use [httptest2 redactor functions](https://enpiar.com/httptest2/articles/redacting.html#setting-a-package-level-redactor) to find some early pattern in a response and `gsub` the rest away. I use a lot of those [like here](https://github.com/ropensci-review-tools/repometrics/blob/main/inst/httptest2/redact.R) to record only tiny portions of otherwise potentially huge response files.
