### 2024-11-25 c29f251
add detailed quote about AGFD data and file sizes
