+++ R/get_agfd.R
@@ -1,6 +1,32 @@
 
 #' Fetch Australian Gridded Farm Data
 #'
+#' @description
+#' A short description...
+#'
+#' From the [ABARES website](https://www.agriculture.gov.au/abares/research-topics/surveys/farm-survey-data/australian-gridded-farm-data):
+#' \dQuote{The Australian Gridded Farm Data (\acronym{AGFD}) are a set of national
+#'  scale maps containing simulated data on historical broadacre farm business
+#'  outcomes including farm profitability on an 0.05-degree (approximately 5 km)
+#'  grid.\cr
+#'  These data have been produced by \acronym{ABARES} as part of the ongoing
+#'  Australian Agricultural Drought Indicator (\acronym{AADI}) project
+#'  (previously known as the Drought Early Warning System Project) and were
+#'  derived using \acronym{ABARES}
+#'  [*farmpredict*](https://www.agriculture.gov.au/abares/research-topics/climate/drought/farmpredict)
+#'  model, which in turn is based on ABARES Agricultural and Grazing Industries
+#'  Survey (\acronym{AAGIS}) data.\cr
+#'  These maps provide estimates of farm business profit, revenue, costs and
+#'  production by location (grid cell) and year for the period 1990-91 to
+#'  2022-23. The data do not include actual observed outcomes but rather model
+#'  predicted outcomes for representative or \sQuote{typical} broadacre farm
+#'  businesses at each location considering likely farm characteristics and
+#'  prevailing weather conditions and commodity prices.}\cr
+#'  -- \acronym{ABARES}, 2024-11-25
+#'
+#' Both sets of data are large in file size, *i.e.*, >1GB, and will require time
+#'   to download.
+#'
 #' @param fixed_prices `Boolean` Download historical climate and prices or
 #'  historical climate and fixed prices as described in  (Hughes *et al.* 2022).
 #'  Defaults to `TRUE` and downloads the data with historical climate and fixed
