### 2024-11-25 adamhsparks/read.abares#1 @mpadge
Desc has github.org not github.com
@adamhsparks I clicked on the link to your package on [the latest checks in the review thread](https://github.com/ropensci/software-review/issues/667#issuecomment-2486224313) and found out it was to "github.org" and not "github.com". I thought of course it was a :bug: in our system somewhere, and looked, and looked ... and then found out it's actually here:
https://github.com/adamhsparks/read.abares/blob/399e3ce2197d5cd7f74a23221ed640b1875f4d71/DESCRIPTION#L14
You might wanna fix that one :wink: Thanks
