### 2024-11-25 5e678cf
add detailed quote about AGFD data and file sizes
