### 2024-11-25 cba15bc
Closes https://github.com/adamhsparks/read.abares/issues/1, thanks @mpadge!
