### 2024-11-27 #667 @emilyriederer
Thanks, @adamhsparks ! That sounds like a great suggestion. I'll hold if you want to investigate including sample data or using @mpadge 's idea. If those prove intractable, I'm also comfortable with moving checks offline / local, so please feel free to ping and we can move the ball forward here.
