### 2024-11-30 89f7042
make read_st() be quiet
