### 2024-11-30 af86f3f
try using mocking and run tests on CI
