+++ tests/testthat/test-get_aagis_regions.R
@@ -5,8 +5,6 @@ withr::local_envvar(R_USER_CACHE_DIR = file.path(tempdir(), "abares.cache.1"))
 # without caching ----
 with_mock_dir("test-get_aagis_regions", {
   test_that("get_aagis_regions doesn't cache", {
-    skip_if_offline()
-    skip_on_ci()
     x <- get_aagis_regions(cache = FALSE)
     expect_s3_class(x, "sf")
     expect_false(file.exists(
@@ -17,8 +15,6 @@ with_mock_dir("test-get_aagis_regions", {
 
 with_mock_dir("test-get_aagis_regions", {
   test_that("get_aagis_regions skips downloading if still in tempdir()", {
-    skip_if_offline()
-    skip_on_ci()
     x <- .check_existing_aagis(cache = FALSE)
     expect_s3_class(x, "sf")
   })
@@ -28,8 +24,6 @@ with_mock_dir("test-get_aagis_regions", {
 
 with_mock_dir("test-get_aagis_regions", {
   test_that("get_aagis_regions caches", {
-    skip_if_offline()
-    skip_on_ci()
     x <- get_aagis_regions(cache = TRUE)
     expect_s3_class(x, "sf")
     expect_true(file.exists(
@@ -46,8 +40,6 @@ with_mock_dir("test-get_aagis_regions", {
 
 with_mock_dir("test-get_aagis_regions", {
   test_that("get_aagis_regions skips downloading if cache is available", {
-    skip_if_offline()
-    skip_on_ci()
     x <- .check_existing_aagis(cache = TRUE)
     expect_s3_class(x, "sf")
   })
@@ -58,8 +50,6 @@ withr::local_envvar(R_USER_CACHE_DIR = file.path(tempdir(), "abares.cache.2"))
 
 with_mock_dir("test-get_aagis_regions", {
   test_that("get_aagis_regions does cache", {
-    skip_if_offline()
-    skip_on_ci()
     x <- get_aagis_regions(cache = TRUE)
     expect_s3_class(x, "sf")
     expect_true(file.exists(
@@ -76,8 +66,6 @@ with_mock_dir("test-get_aagis_regions", {
 
 with_mock_dir("test-get_aagis_regions", {
   test_that("get_aagis_regions skips downloading if still in tempdir()", {
-    skip_if_offline()
-    skip_on_ci()
     x <- .check_existing_aagis(cache = TRUE)
     expect_s3_class(x, "sf")
   })
