### 2024-11-30 a06e557
Revert "try using mocking and run tests on CI"
This reverts commit af86f3f03a07cdb4af20b2f1cba4d9a9438f91bd.
