+++ tests/testthat/test-get_aagis_regions.R
@@ -5,6 +5,8 @@ withr::local_envvar(R_USER_CACHE_DIR = file.path(tempdir(), "abares.cache.1"))
 # without caching ----
 with_mock_dir("test-get_aagis_regions", {
   test_that("get_aagis_regions doesn't cache", {
+    skip_if_offline()
+    skip_on_ci()
     x <- get_aagis_regions(cache = FALSE)
     expect_s3_class(x, "sf")
     expect_false(file.exists(
@@ -15,6 +17,8 @@ with_mock_dir("test-get_aagis_regions", {
 
 with_mock_dir("test-get_aagis_regions", {
   test_that("get_aagis_regions skips downloading if still in tempdir()", {
+    skip_if_offline()
+    skip_on_ci()
     x <- .check_existing_aagis(cache = FALSE)
     expect_s3_class(x, "sf")
   })
@@ -24,6 +28,8 @@ with_mock_dir("test-get_aagis_regions", {
 
 with_mock_dir("test-get_aagis_regions", {
   test_that("get_aagis_regions caches", {
+    skip_if_offline()
+    skip_on_ci()
     x <- get_aagis_regions(cache = TRUE)
     expect_s3_class(x, "sf")
     expect_true(file.exists(
@@ -40,6 +46,8 @@ with_mock_dir("test-get_aagis_regions", {
 
 with_mock_dir("test-get_aagis_regions", {
   test_that("get_aagis_regions skips downloading if cache is available", {
+    skip_if_offline()
+    skip_on_ci()
     x <- .check_existing_aagis(cache = TRUE)
     expect_s3_class(x, "sf")
   })
@@ -50,6 +58,8 @@ withr::local_envvar(R_USER_CACHE_DIR = file.path(tempdir(), "abares.cache.2"))
 
 with_mock_dir("test-get_aagis_regions", {
   test_that("get_aagis_regions does cache", {
+    skip_if_offline()
+    skip_on_ci()
     x <- get_aagis_regions(cache = TRUE)
     expect_s3_class(x, "sf")
     expect_true(file.exists(
@@ -66,6 +76,8 @@ with_mock_dir("test-get_aagis_regions", {
 
 with_mock_dir("test-get_aagis_regions", {
   test_that("get_aagis_regions skips downloading if still in tempdir()", {
+    skip_if_offline()
+    skip_on_ci()
     x <- .check_existing_aagis(cache = TRUE)
     expect_s3_class(x, "sf")
   })
