+++ R/get_aagis_regions.R
@@ -1,9 +1,10 @@
+
 #' Get AAGIS Region Mapping Files
 #'
 #' Download, cache and import the  Australian Agricultural and Grazing
-#'  Industries Survey (\acronym{AAGIS} regions geospatial shapefil. Upon import,
-#'  the geometries  are automatically corrected to fix invalid geometries that
-#'  are present in the original shapefile.
+#'  Industries Survey (\acronym{AAGIS} regions geospatial shapefile. Upon
+#'  import, the geometries  are automatically corrected to fix invalid
+#'  geometries that are present in the original shapefile.
 #'
 #' @param cache `Boolean` Cache the \acronym{AAGIS} regions shape files after
 #'  download using `tools::R_user_dir()` to identify the proper directory for
