+++ man/get_aagis_regions.Rd
@@ -21,9 +21,9 @@ An \CRANpkg{sf} object of the \acronym{AAGIS} regions
 }
 \description{
 Download, cache and import the  Australian Agricultural and Grazing
-Industries Survey (\acronym{AAGIS} regions geospatial shapefil. Upon import,
-the geometries  are automatically corrected to fix invalid geometries that
-are present in the original shapefile.
+Industries Survey (\acronym{AAGIS} regions geospatial shapefile. Upon
+import, the geometries  are automatically corrected to fix invalid
+geometries that are present in the original shapefile.
 }
 \examples{
 \dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
