+++ R/get_aagis_regions.R
@@ -1,7 +1,7 @@
 
 #' Get AAGIS Region Mapping Files
 #'
-#' Download, cache and import the  Australian Agricultural and Grazing
+#' Download, cache and import the Australian Agricultural and Grazing
 #'  Industries Survey (\acronym{AAGIS} regions geospatial shapefile. Upon
 #'  import, the geometries  are automatically corrected to fix invalid
 #'  geometries that are present in the original shapefile.
