+++ man/get_aagis_regions.Rd
@@ -20,7 +20,7 @@ the files locally as a native \R object. If \code{FALSE}, this function uses
 An \CRANpkg{sf} object of the \acronym{AAGIS} regions
 }
 \description{
-Download, cache and import the  Australian Agricultural and Grazing
+Download, cache and import the Australian Agricultural and Grazing
 Industries Survey (\acronym{AAGIS} regions geospatial shapefile. Upon
 import, the geometries  are automatically corrected to fix invalid
 geometries that are present in the original shapefile.
