### 2024-11-30 e73331b
write sf quietly and ensure valid geometries for all imports
