### 2024-11-30 505d808
document that AAGIS regions will be saved as Geopackage file locally
