+++ man/get_aagis_regions.Rd
@@ -10,11 +10,12 @@
 get_aagis_regions(cache = TRUE)
 }
 \arguments{
-\item{cache}{\code{Boolean} Cache the \acronym{AAGIS} regions shape files after
-download using \code{tools::R_user_dir()} to identify the proper directory for
-storing user data in a cache for this package. Defaults to \code{TRUE}, caching
-the files locally as a native \R object. If \code{FALSE}, this function uses
-\code{tempdir()} and the files are deleted upon closing of the \R session.}
+\item{cache}{\code{Boolean} Cache the \acronym{AAGIS} regions' geospatial file
+after downloading using \code{tools::R_user_dir("read.abares", "cache")} to
+identify the proper directory for storing user data in a cache for this
+package. Defaults to \code{TRUE}, caching the files locally as a Geopackage. If
+\code{FALSE}, this function uses \code{tempdir()} and the files are deleted upon
+closing of the \R session.}
 }
 \value{
 An \CRANpkg{sf} object of the \acronym{AAGIS} regions
