+++ man/get_historical_national_estimates.Rd
@@ -17,7 +17,8 @@ A \link[data.table:data.table]{data.table::data.table} object with the \code{Var
 Get Historical National Estimates from ABARES
 }
 \note{
-Columns are renamed and reordered for consistency.
+Columns are renamed for consistency with other \acronym{ABARES} products
+serviced in this package using a snake_case format and ordered consistently.
 }
 \examples{
 \dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
