### 2024-12-01 9ca6e7f
Simplify .retry_download() to actually work as intended with less lines
