### 2024-12-01 fc5dd3c
use httptest2 and test .retry_download()
