### 2024-12-01 0a9ba82
completely rewrite download function to use tryCatch and {brio}
