### 2024-12-01 f9691b1
try fixing bugs in get_agfd()
