### 2024-12-01 babcc35
add references and sources to trade regions
