+++ R/get_abares_trade_regions.R
@@ -17,9 +17,11 @@
 #'
 #' trade_regions
 #'
-#' @return A \CRANpkg{data.table} object of the \acronym{ABARES} trade data regions.
+#' @return A \CRANpkg{data.table} object of the \acronym{ABARES} trade data
+#'  regions.
 #' @family Trade
-#' @source <https://daff.ent.sirsidynix.net.au/client/en_AU/search/asset/1033841/0>
+#' @referencs <https://daff.ent.sirsidynix.net.au/client/en_AU/search/asset/1033841/0>
+#' @source <https://daff.ent.sirsidynix.net.au/client/en_AU/search/asset/1033841/2>
 #' @autoglobal
 #' @export
 
