### 2024-12-01 2708053
Fix references field in get_abares_trade_regions()
