+++ R/read_soil_thickness_stars.R
@@ -2,10 +2,10 @@
 #' Read Soil Thickness File With stars
 #'
 #' Read Soil Thickness for Australian Areas of Intensive Agriculture of Layer 1
-#'  data as a [stars] object.
+#'  data as a \CRANpkg{stars} object.
 #'
-#' @param files An \pkg{read.abares} `read.abares.soil.thickness` object, a `list` that
-#'  contains the \acronym{ESRI} grid file to import
+#' @param files An \pkg{read.abares} `read.abares.soil.thickness` object, a
+#'  `list` that contains the \acronym{ESRI} grid file to import
 #' @return a [stars] object of the Soil Thickness for Australian Areas of
 #'  Intensive Agriculture of Layer 1
 #'
