### 2024-12-02 e39fb4b
increase timeout, remove tryCatch rely on httr2 error
