+++ R/internal_functions.R
@@ -50,21 +50,12 @@
 #' @noRd
 
 .retry_download <- function(url, .f, .max_tries = 3L) {
-  tryCatch({
-    resp <- httr2::request(base_url = url) |>
-      httr2::req_options(http_version = 2, timeout = 500L) |>
-      httr2::req_retry(max_tries = .max_tries) |>
-      httr2::req_cache(path = tempdir()) |>
-      httr2::req_progress() |>
-      httr2::req_perform()
-
-  }, error = function(e) {
-    cli::cli_abort("There was an error with this download, please retry.",
-                   call = rlang::caller_env())
-  })
-
-  # save the file to be imported into the R session.
-    resp |>
-      httr2::resp_body_raw() |>
-      brio::write_file_raw(path = .f)
+  httr2::request(base_url = url) |>
+    httr2::req_options(http_version = 2, timeout = 2000L) |>
+    httr2::req_retry(max_tries = .max_tries) |>
+    httr2::req_cache(path = tempdir()) |>
+    httr2::req_progress() |>
+    httr2::req_perform() |>
+    httr2::resp_body_raw() |>
+    brio::write_file_raw(path = .f)
 }
