### 2024-12-03 6aac6f6
Update links and add source to documentation for estimates
