### 2024-12-03 a4fbc50
Update links and add source to documentation for estimates
