### 2024-12-03 5403abc
Simplify downloading of AGFD files
