### 2024-12-03 1c7170f
Add source and references to documentation and update link to file
