+++ man/read_agfd_stars.Rd
@@ -7,7 +7,7 @@
 read_agfd_stars(files)
 }
 \arguments{
-\item{files}{A list of NetCDF files to import}
+\item{files}{A list of \acronym{AGFD} NetCDF files to import}
 }
 \value{
 a \code{list} object of \CRANpkg{stars} objects of the Australian Gridded
