+++ man/read_agfd_dt.Rd
@@ -7,7 +7,7 @@
 read_agfd_dt(files)
 }
 \arguments{
-\item{files}{A list of NetCDF files to import}
+\item{files}{A list of \acronym{AGFD} NetCDF files to import}
 }
 \value{
 a \link[data.table:data.table]{data.table::data.table} object of the Australian Gridded Farm Data
