+++ man/get_estimates_by_size.Rd
@@ -5,7 +5,7 @@
 \alias{get_est_by_size}
 \title{Get Estimates by Size From ABARES}
 \source{
-\url{https://www.agriculture.gov.au/abares/data/farm-data-portal#data-download}
+\url{https://www.agriculture.gov.au/sites/default/files/documents/fdp-national-historical.csv}
 }
 \usage{
 get_estimates_by_size()
@@ -30,6 +30,9 @@ get_estimates_by_size()
 # or shorter
 get_est_by_size()
 \dontshow{\}) # examplesIf}
+}
+\references{
+\url{https://www.agriculture.gov.au/abares/data/farm-data-portal#data-download}
 }
 \seealso{
 Other Estimates: 
