### 2024-12-03 604fc2d
Update tests to align with new functionality
