### 2024-12-03 0c4d633
Update tests to align with new functionality
