### 2024-12-03 cb32995
Fix source and references for estimates by size/performance
