### 2024-12-03 6e3833b
Add source/rerferences to forecast database
