+++ man/get_historical_forecast_database.Rd
@@ -5,7 +5,7 @@
 \alias{get_historical_forecast}
 \title{Get Historical Forecast Database From ABARES}
 \source{
-\url{https://www.agriculture.gov.au/abares/research-topics/agricultural-outlook/historical-forecasts#:~:text=About\%20the\%20historical\%20agricultural\%20forecast,relevant\%20to\%20Australian\%20agricultural\%20markets}
+\url{https://daff.ent.sirsidynix.net.au/client/en_AU/search/asset/1031941/0}
 }
 \usage{
 get_historical_forecast_database()
@@ -53,4 +53,7 @@ get_historical_forecast_database()
 # or shorter
 get_historical_forecast()
 \dontshow{\}) # examplesIf}
+}
+\references{
+\url{https://www.agriculture.gov.au/abares/research-topics/agricultural-outlook/historical-forecasts#:~:text=About\%20the\%20historical\%20agricultural\%20forecast,relevant\%20to\%20Australian\%20agricultural\%20markets}
 }
