### 2024-12-03 b6f6c9a
add sources for AGFD documentation
