+++ tests/testthat/test-read_agfd_stars.R
@@ -7,8 +7,8 @@ test_that("read_agfd_stars() returns a stars object", {
 
   expect_type(x, "list")
   expect_s3_class(x[[1]], "stars")
-  expect_identical(
-    names(x),
+  expect_named(
+    x,
     c(
       "f2022.c1991.p2022.t2022.nc",
       "f2022.c1992.p2022.t2022.nc",
@@ -45,7 +45,7 @@ test_that("read_agfd_stars() returns a stars object", {
       "f2022.c2023.p2022.t2022.nc"
     )
   )
-  expect_identical(
+  expect_named(
     names(x[[1]]),
     c(
       "farmno",
