+++ tests/testthat/test-get_abares_trade.R
@@ -66,7 +66,7 @@ test_that("get_abares_trade caches", {
   expect_true(file.exists(
     file.path(.find_user_cache(), "abares_trade_dir/abares_trade.rds")
   ))
-  expect_true(!file.exists(
+  expect_false(file.exists(
     file.path(.find_user_cache(), "abares_trade_dir/abares_trade.zip")
   ))
 })
