### 2024-12-04 6d4afc4
use cli::cat_line() and don't force newlines, let {cli} do that
