### 2024-12-04 c0a16d7
Fix source and references for estimates by size/performance
