### 2024-12-04 39b6a71
add sources for AGFD documentation
