### 2024-12-04 e0663e8
Update copyright details
