+++ DESCRIPTION
@@ -4,9 +4,13 @@ Title: Provides simple downloading, parsing and importing of Australian
     Bureau of Agricultural and Resource Economics and Sciences (ABARES)
     data sources
 Version: 0.1.0
-Authors@R: 
+Authors@R: c(
     person("Adam H.", "Sparks", , "adamhsparks@gmail.com", role = c("cre", "aut"),
-           comment = c(ORCID = "0000-0002-0061-8359"))
+           comment = c(ORCID = "0000-0002-0061-8359")),
+    person("Curtin University", role = "cph"),
+    person("Grains Research and Development Corporation", role = c("fnd", "cph"),
+           comment = "GRDC Project CUR2210-005OPX")
+  )
 Description: Download and import data from the Australian Bureau of
     Agricultural and Resource Economics and Sciences (ABARES)
     <https://www.agriculture.gov.au/abares>.
