### 2024-12-04 b3fc009
Update copyright details
