+++ man/read.abares-package.Rd
@@ -22,5 +22,11 @@ Useful links:
 \author{
 \strong{Maintainer}: Adam H. Sparks \email{adamhsparks@gmail.com} (\href{https://orcid.org/0000-0002-0061-8359}{ORCID})
 
+Other contributors:
+\itemize{
+  \item Curtin University [copyright holder]
+  \item Grains Research and Development Corporation (GRDC Project CUR2210-005OPX) [funder, copyright holder]
+}
+
 }
 \keyword{internal}
