### 2024-12-06 a1ac2f3
replace another "\n" with `cat_line()`
