+++ R/get_agfd.R
@@ -133,6 +133,6 @@ get_agfd <- function(fixed_prices = TRUE,
 print.read.abares.agfd.nc.files <- function(x, ...) {
   cli::cli_h1("Locally Available ABARES AGFD NetCDF Files")
   cli::cli_ul(basename(x))
-  cat("\n")
+  cli::cat_line()
   invisible(x)
 }
