### 2024-12-06 a02a562
Rebuild vignette
