+++ vignettes/read.abares.Rmd
@@ -1,7 +1,7 @@
 ---
 title: "read.abares"
 author: "Adam H. Sparks"
-date: "2024-10-26"
+date: "2024-12-06"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{read.abares}
@@ -12,47 +12,9 @@ vignette: >
 
 
 
-## Working With Estimates
-
-You can download the CSV files directly in your R session as illustrated below.
-
-
-``` r
-library(read.abares)
-
-get_hist_nat_est() |>
-  head()
-#>                             Variable  Year Value   RSE      Industry
-#>                               <char> <int> <num> <int>        <char>
-#> 1: AI stud fees and herd testing ($)  1990   790    39 All Broadacre
-#> 2:           Accounting services ($)  1990  2850     4 All Broadacre
-#> 3:             Advisory services ($)  1990   280    37 All Broadacre
-#> 4:        Age of owner manager (yrs)  1990    53     1 All Broadacre
-#> 5:               Age of spouse (yrs)  1990    49     1 All Broadacre
-#> 6:                     Agistment ($)  1990  1950    36 All Broadacre
-
-get_hist_sta_est() |>
-  head()
-#>                             Variable  Year Value   RSE           State      Industry
-#>                               <char> <int> <num> <int>          <char>        <char>
-#> 1: AI stud fees and herd testing ($)  1990   210   118 New South Wales All Broadacre
-#> 2:           Accounting services ($)  1990  2740     7 New South Wales All Broadacre
-#> 3:             Advisory services ($)  1990   190   105 New South Wales All Broadacre
-#> 4:        Age of owner manager (yrs)  1990    55     2 New South Wales All Broadacre
-#> 5:               Age of spouse (yrs)  1990    52     2 New South Wales All Broadacre
-#> 6:                     Agistment ($)  1990  2790    71 New South Wales All Broadacre
-
-get_hist_reg_est() |>
-  head()
-#>                             Variable  Year    ABARES region Value   RSE
-#>                               <char> <int>           <char> <num> <int>
-#> 1: AI stud fees and herd testing ($)  1990 NSW Central West    90    68
-#> 2:           Accounting services ($)  1990 NSW Central West  2420    19
-#> 3:             Advisory services ($)  1990 NSW Central West   120    80
-#> 4:        Age of owner manager (yrs)  1990 NSW Central West    57     5
-#> 5:               Age of spouse (yrs)  1990 NSW Central West    51     4
-#> 6:                     Agistment ($)  1990 NSW Central West   310    52
-```
+This vignette demonstrates some of the functionality of {read.abares}.
+Please note that not all functions are demonstrated here, please refer to the [documentation reference](https://adamhsparks.codeberg.page/read.abares/reference/) for a full list of functionality.
+The worked examples here show some of the more advanced features that {read.abares} offers beyond just fetching and importing data, _e.g._, the Australian Gridded Farm Data, which can be downloaded, cached and then imported using one of four types of object or the soil thickness data, which includes rich metadata.
 
 ## Working With AGFD Data
 
@@ -88,20 +50,20 @@ star <- get_agfd(cache = TRUE) |>
 head(star[[1]])
 #> stars object with 2 dimensions and 6 attributes
 #> attribute(s):
-#>                         Min.         1st Qu.         Median           Mean        3rd Qu.
-#> farmno          15612.000000 233091.50000000 329567.0000000 324737.7187618 418508.5000000
-#> R_total_hat_ha      2.954396      7.88312157     21.7520529    169.5139301    174.8553843
[…]
