### 2024-12-06 414dab4
shaves nanoseconds off the evaluation
