### 2024-12-07 139a2b7
Add description table of the layers in the AGFD data
