+++ R/get_agfd.R
@@ -49,6 +49,60 @@
 #'  the files locally. If `FALSE`, this function uses `tempdir()` and the files
 #'  are deleted upon closing of the \R session.
 #'
+#' @section Data Layers:
+#'
+#' The data layers from the downloaded NetCDF files are described in Table 2
+#'  from Australian Bureau of Agricultural and Resource Economics and Sciences
+#'  (2024).
+#'
+#' Following is a copy of Table 2 for your convenience, please refer to the full
+#' document for all methods and metadata.
+#'
+#' \tabular{lll}{
+#'   \strong{Layer} \tab \strong{Unit} \tab \strong{Description} \cr
+#'   farmno                \tab -         \tab Row index and column index of the grid cell in the form of YYYXXX                                            \cr
+#'   A_barley_hat_ha       \tab -         \tab Proportion of total farm area planted to barley                                                              \cr
+#'   A_oilseeds_hat_ha     \tab -         \tab Proportion of total farm area planted to canola                                                              \cr
+#'   A_sorghum_hat_ha      \tab -         \tab Proportion of total farm area planted to sorghum                                                             \cr
+#'   A_total_cropped_ha    \tab -         \tab Proportion of total farm area planted to crops                                                               \cr
+#'   A_wheat_hat_ha        \tab -         \tab Proportion of total farm area planted to wheat                                                               \cr
+#'   C_chem_hat_ha         \tab $/ha      \tab Expenditure on crop and pasture chemicals per hectare                                                        \cr
+#'   C_fert_hat_ha         \tab $/ha      \tab Expenditure on fertiliser per hectare                                                                        \cr
+#'   C_fodder_hat_ha       \tab $/ha      \tab Expenditure on fodder per hectare                                                                            \cr
+#'   C_fuel_hat_ha         \tab $/ha      \tab Expenditure on fuel, oil and grease per hectare                                                              \cr
+#'   C_total_hat_ha        \tab $/ha      \tab Total cash costs per hectare                                                                                 \cr
+#'   FBP_fci_hat_ha        \tab $/ha      \tab Farm cash income per hectare                                                                                 \cr
+#'   FBP_fbp_hat_ha        \tab $/ha      \tab Farm business profit per hectare, cash income adjusted for family labour, depreciation, and changes in stocks\cr
+#'   FBP_pfe_hat_ha        \tab $/ha      \tab Profit at full equity per hectare                                                                            \cr
+#'   H_barley_dot_hat      \tab t/ha      \tab Barley yield (production per hectare planted)                                                                \cr
+#'   H_oilseeds_dot_hat    \tab t/ha      \tab Oilseeds yield (production per hectare planted)                                                              \cr
+#'   H_sorghum_dot_hat     \tab t/ha      \tab Sorghum yield (production per hectare planted)                                                               \cr
+#'   H_wheat_dot_hat       \tab t/ha      \tab Wheat yield (production per hectare planted)                                                                 \cr
+#'   Q_barley_hat_ha       \tab t/ha      \tab Barley sold per hectare (total farm area)                                                                    \cr
+#'   Q_beef_hat_ha         \tab Number/ha \tab Beef number sold per hectare                                                                                 \cr
+#'   Q_lamb_hat_ha         \tab Number/ha \tab Prime lamb number sold per hectare                                                                           \cr
+#'   Q_oilseeds_hat_ha     \tab t/ha      \tab Canola sold per hectare (total farm area)                                                                    \cr
+#'   Q_sheep_hat_ha        \tab Number/ha \tab Sheep number sold per hectare                                                                                \cr
[…]
