### 2024-12-07 5d75ab5
add comment on the use of piping and caching
