+++ R/read_agfd_terra.R
@@ -13,6 +13,9 @@
 #'  Farm Data with the file names as the list's objects' names
 #'
 #' @examplesIf interactive()
+#'
+#' # using piping, which can use the {read.abares} cache after the first DL
+#'
 #' get_agfd(cache = TRUE) |>
 #'   read_agfd_terra()
 #'
