### 2024-12-07 215a706
add Model scenarios to documentation for AGFD
