+++ tests/testthat/test-get_estimates_by_size.R
@@ -1,7 +1,6 @@
 
 test_that("get_estimates_by_size works", {
   skip_if_offline()
-  skip_on_ci()
   x <- get_estimates_by_size()
   expect_named(x, c("Variable", "Year", "Size", "Industry", "Value", "RSE"))
   expect_s3_class(x, c("data.table", "data.frame"))
