### 2024-12-07 1240578
Revert "Try setting user-agent for CI tests"
This reverts commit 9599142632e62d16cfa4d155d4f5c612556b5121.
