+++ R/get_agfd.R
@@ -87,13 +87,13 @@
 #' Simulation output data are saved as multilayer NetCDF files, which are named
 #'  using following convention:
 #'
-#'    *f< farm year >.c< climate year >.p< price year >.t< technology year >.nc*
+#' \code{\emph{f<farm year>.c<climate year>.p<price year>.t<technology year>.nc}}
 #'
 #'  where:
-#'  * *< farm year >* = Financial year of farm business data is used in simulations.
-#'  * *< climate year >* = Financial year of climate data is used in simulations.
-#'  * *< price year >* = Financial year of output and input prices used in simulations.
-#'  * *< technology year >* = Financial year of farm \sQuote{technology} (equal to farm year in all simulations)
+#'  * \code{\emph{<farm year>}} = Financial year of farm business data is used in simulations.
+#'  * \code{\emph{<climate year>}} = Financial year of climate data is used in simulations.
+#'  * \code{\emph{<price year>}} = Financial year of output and input prices used in simulations.
+#'  * \code{\emph{<technology year>}} = Financial year of farm \sQuote{technology} (equal to farm year in all simulations)
 #'  Here financial years are referred to by the closing calendar year
 #'   (*e.g.*, 2022 = 1 July 2021 to 30 June 2022)."
 #'
