+++ man/read.abares-package.Rd
@@ -24,6 +24,7 @@ Useful links:
 
 Other contributors:
 \itemize{
+  \item Jacob \email{jacob@wujciak.de} (\href{https://orcid.org/0000-0002-7281-3989}{ORCID}) (Assisted with troubleshooting formatting in documentation to display '<' and '>' properly) [contributor]
   \item Curtin University of Technology (Provided support through Adam Sparks's time.) [copyright holder]
   \item Grains Research and Development Corporation (GRDC Project CUR2210-005OPX (AAGI-CU)) [funder, copyright holder]
 }
