+++ DESCRIPTION
@@ -7,6 +7,8 @@ Version: 0.1.0
 Authors@R: c(
     person("Adam H.", "Sparks", , "adamhsparks@gmail.com", role = c("cre", "aut"),
            comment = c(ORCID = "0000-0002-0061-8359")),
+    person("Jacob", , , "jacob@wujciak.de", role = "ctb",
+           comment = c("Assisted with troubleshooting formatting in documentation to display '<' and '>' properly", ORCID = "0000-0002-7281-3989")),
     person("Curtin University of Technology", role = "cph",
            comment = "Provided support through Adam Sparks's time."),
     person("Grains Research and Development Corporation", role = c("fnd", "cph"),
