+++ man/read_agfd_terra.Rd
@@ -89,16 +89,16 @@ Sciences (2024)
 Simulation output data are saved as multilayer NetCDF files, which are named
 using following convention:
 
-\code{\emph{f<farm year>.c<climate year>.p<price year>.t<technology year>.nc}}
+\code{f<farm year>.c<climate year>.p<price year>.t<technology year>.nc}
 
 where:
 \itemize{
-\item \code{\emph{<farm year>}} = Financial year of farm business data is used in simulations.
-\item \code{\emph{<climate year>}} = Financial year of climate data is used in simulations.
-\item \code{\emph{<price year>}} = Financial year of output and input prices used in simulations.
-\item \code{\emph{<technology year>}} = Financial year of farm \sQuote{technology} (equal to farm year in all simulations)
+\item \code{<farm year>} = Financial year of farm business data is used in simulations.
+\item \code{<climate year>} = Financial year of climate data is used in simulations.
+\item \code{<price year>} = Financial year of output and input prices used in simulations.
+\item \code{<technology year>} = Financial year of farm \sQuote{technology} (equal to farm year in all simulations)
 Here financial years are referred to by the closing calendar year
-(\emph{e.g.}, 2022 = 1 July 2021 to 30 June 2022)."
+(\emph{e.g.}, 2022 = 1 July 2021 to 30 June 2022).
 }
 
 -- Taken from Australian Bureau of Agricultural and Resource Economics and
