### 2024-12-07 f69e56d
Use \var and \cite
