+++ man/get_agfd.Rd
@@ -96,8 +96,8 @@ resulting simulation results and percentile indicators are intended to
 reflect the combined impacts of annual climate and commodity price
 variability."
 
--- Taken from Australian Bureau of Agricultural and Resource Economics and
-Sciences (2024)
+-- Taken from  \cite{Australian Bureau of Agricultural and Resource
+Economics and Sciences (2024)}
 }
 }
 
@@ -107,28 +107,28 @@ Sciences (2024)
 Simulation output data are saved as multilayer NetCDF files, which are named
 using following convention:
 
-\code{f<farm year>.c<climate year>.p<price year>.t<technology year>.nc}
+\var{f<farm year>.c<climate year>.p<price year>.t<technology year>.nc}
 
 where:
 \itemize{
-\item \code{<farm year>} = Financial year of farm business data is used in simulations.
-\item \code{<climate year>} = Financial year of climate data is used in simulations.
-\item \code{<price year>} = Financial year of output and input prices used in simulations.
-\item \code{<technology year>} = Financial year of farm \sQuote{technology} (equal to farm year in all simulations)
+\item \var{<farm year>} = Financial year of farm business data is used in simulations.
+\item \var{<climate year>} = Financial year of climate data is used in simulations.
+\item \var{<price year>} = Financial year of output and input prices used in simulations.
+\item \var{<technology year>} = Financial year of farm \sQuote{technology} (equal to farm year in all simulations)
 Here financial years are referred to by the closing calendar year
 (\emph{e.g.}, 2022 = 1 July 2021 to 30 June 2022).
 }
 
--- Taken from Australian Bureau of Agricultural and Resource Economics and
-Sciences (2024)
+-- Taken from  \cite{Australian Bureau of Agricultural and Resource
+Economics and Sciences (2024)}
 }
 
 \section{Data layers}{
 
 
 The data layers from the downloaded NetCDF files are described in Table 2
-as seen in Australian Bureau of Agricultural and Resource Economics and
-Sciences (2024).
+as seen in \cite{Australian Bureau of Agricultural and Resource Economics
+and Sciences (2024)}.
 
 Following is a copy of Table 2 for your convenience, please refer to the full
 document for all methods and metadata.

[…]
