### 2024-12-07 c9246c2
Improve vignette's readability by splitting code into more chunks
