+++ vignettes/read.abares.Rmd
@@ -1,7 +1,7 @@
 ---
 title: "read.abares"
 author: "Adam H. Sparks"
-date: "2024-12-06"
+date: "2024-12-07"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{read.abares}
@@ -35,10 +35,17 @@ Directly from the DAFF website:
 -- Australian Department of Agriculture, Fisheries and Forestry.
 
 
+Load the {read.abares} library.
+
 
 ``` r
 library(read.abares)
+```
 
+Download or load from the local cache and read the AGFD files as a list of {stars} objects.
+
+
+``` r
 ## A list of {stars} objects
 star <- get_agfd(cache = TRUE) |>
   read_agfd_stars()
@@ -68,7 +75,12 @@ head(star[[1]])
 #>     from  to refsys              values x/y
 #> lon    1 886 WGS 84 [886] 112,...,156.2 [x]
 #> lat    1 691 WGS 84 [691] -44.5,...,-10 [y]
+```
 
+Download or load from the local cache and read the AGFD files as a `terra::rast` object.
+
+
+``` r
 ## A {terra} `rast` object
 terr <- get_agfd(cache = TRUE) |>
   read_agfd_terra()
@@ -83,7 +95,12 @@ head(terr[[1]])
 #> names       : farmno, R_tot~at_ha, C_tot~at_ha, FBP_f~at_ha, FBP_f~at_ha, A_whe~at_ha, ... 
 #> min values  :    NaN,         NaN,         NaN,         NaN,         NaN,         NaN, ... 
 #> max values  :    NaN,         NaN,         NaN,         NaN,         NaN,         NaN, ...
+```
 
+Download or load from the local cache and read the AGFD files as a list of {tidync} objects.
+
+
+``` r
 ## A list of {tidync} objects
 tdnc <- get_agfd(cache = TRUE) |>
   read_agfd_tidync()
@@ -91,9 +108,9 @@ tdnc <- get_agfd(cache = TRUE) |>
 head(tdnc[[1]])
 #> $source
 #> # A tibble: 1 × 2
-#>   access              source                                                           
-#>   <dttm>              <chr>                                                            
-#> 1 2024-12-06 10:17:01 /Users/adamsparks/Library/Caches/org.R-project.R/R/read.abares/h…
+#>   access              source                                                            
+#>   <dttm>              <chr>                                                             
+#> 1 2024-12-07 21:23:36 /Users/adamsparks/Library/Caches/org.R-project.R/R/read.abares/hi…
 #> 
 #> $axis
 #> # A tibble: 84 × 3
@@ -157,7 +174,12 @@ head(tdnc[[1]])
 #>  9     0 _FillValue FBP_fbp_hat_ha <dbl [1]>   
 #> 10     0 _FillValue A_wheat_hat_ha <dbl [1]>   
 #> # ℹ 39 more rows
+```
 
+Download or load from the local cache and read the AGFD files as a {data.table} object.
+
+
+``` r
 ## A {data.table} object
 get_agfd(cache = TRUE) |>
   read_agfd_dt() |>
@@ -210,38 +232,38 @@ get_agfd(cache = TRUE) |>
 #> 4:    0.000070963917        0.000062521929        0.000009726773       0.03057606
 #> 5:    0.000007780997        0.000006834211        0.000000000000       0.04944272
 #> 6:    0.000059600116        0.000053976389        0.000008478467       0.03322463
-#>    S_beef_births_hat_ha S_beef_deaths_hat_ha Q_beef_hat_ha Q_sheep_hat_ha
-#>                   <num>                <num>         <num>          <num>
[…]
