### 2024-12-07 3a29c5e
add new function to print netcdf file format info in console for AGFD
