### 2024-12-07 a2fbfcc
add function to display AGFD file data format info and document it
