+++ man/print_agfd_nc_file_format.Rd
@@ -1,0 +1,24 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/get_agfd.R
+\name{print_agfd_nc_file_format}
+\alias{print_agfd_nc_file_format}
+\title{Prints File Format Information for the AGFD NetCDF Files}
+\source{
+\url{https://daff.ent.sirsidynix.net.au/client/en_AU/search/asset/1036161/0}
+}
+\usage{
+print_agfd_nc_file_format()
+}
+\description{
+Print the file format from section 3.2 in \cite{Australian Bureau of
+Agricultural and Resource Economics and Sciences}.
+}
+\examples{
+print_agfd_nc_file_format()
+
+}
+\references{
+Australian gridded farm data, Australian Bureau of Agricultural
+and Resource Economics and Sciences, Canberra, July, DOI:
+<https://doi.org/10.25814/7n6z-ev41. CC BY 4.0.>
+}
