+++ R/read_soil_thickness_terra.R
@@ -4,8 +4,12 @@
 #' Read Soil Thickness for Australian Areas of Intensive Agriculture of Layer 1
 #'  data as a [terra::rast] object.
 #'
-#' @param files An \pkg{read.abares} `read.abares.soil.thickness` object, a `list` that
-#'  contains the \acronym{ESRI} grid file to import
+#' @inheritParams read_soil_thickness_stars
+#'
+#' @references <https://data.agriculture.gov.au/geonetwork/srv/eng/catalog.search#/metadata/faa9f157-8e17-4b23-b6a7-37eb7920ead6>
+#' @source <https://anrdl-integration-web-catalog-saxfirxkxt.s3-ap-southeast-2.amazonaws.com/warehouse/staiar9cl__059/staiar9cl__05911a01eg_geo___.zip>
+
+#'
 #' @return a [terra::rast] object of the Soil Thickness for Australian Areas of
 #'  Intensive Agriculture of Layer 1
 #'
