+++ tests/testthat/test-read_aagis_regions.R
@@ -3,10 +3,10 @@
 withr::local_envvar(R_USER_CACHE_DIR = file.path(tempdir(), "abares.cache.1"))
 
 # without caching ----
-test_that("get_aagis_regions doesn't cache", {
+test_that("read_aagis_regions doesn't cache", {
   skip_if_offline()
   skip_on_ci()
-  x <- get_aagis_regions(cache = FALSE)
+  x <- read_aagis_regions(cache = FALSE)
   expect_s3_class(x, "sf")
   expect_false(file.exists(
     file.path(.find_user_cache(), "aagis_regions_dir/aagis.gpkg")
@@ -14,7 +14,7 @@ test_that("get_aagis_regions doesn't cache", {
 })
 
 
-test_that("get_aagis_regions skips downloading if still in tempdir()", {
+test_that("read_aagis_regions skips downloading if still in tempdir()", {
   skip_if_offline()
   skip_on_ci()
   x <- .check_existing_aagis(cache = FALSE)
@@ -23,10 +23,10 @@ test_that("get_aagis_regions skips downloading if still in tempdir()", {
 
 # with caching ----
 
-test_that("get_aagis_regions caches", {
+test_that("read_aagis_regions caches", {
   skip_if_offline()
   skip_on_ci()
-  x <- get_aagis_regions(cache = TRUE)
+  x <- read_aagis_regions(cache = TRUE)
   expect_s3_class(x, "sf")
   expect_true(file.exists(
     file.path(.find_user_cache(), "aagis_regions_dir/aagis.gpkg")
@@ -39,7 +39,7 @@ test_that("get_aagis_regions caches", {
   )))
 })
 
-test_that("get_aagis_regions skips downloading if cache is available", {
+test_that("read_aagis_regions skips downloading if cache is available", {
   skip_if_offline()
   skip_on_ci()
   x <- .check_existing_aagis(cache = TRUE)
@@ -50,10 +50,10 @@ test_that("get_aagis_regions skips downloading if cache is available", {
 withr::local_envvar(R_USER_CACHE_DIR = file.path(tempdir(), "abares.cache.2"))
 
 
-test_that("get_aagis_regions does cache", {
+test_that("read_aagis_regions does cache", {
   skip_if_offline()
   skip_on_ci()
-  x <- get_aagis_regions(cache = TRUE)
+  x <- read_aagis_regions(cache = TRUE)
   expect_s3_class(x, "sf")
   expect_true(file.exists(
     file.path(.find_user_cache(), "aagis_regions_dir/aagis.gpkg")
@@ -66,7 +66,7 @@ test_that("get_aagis_regions does cache", {
   ))
 })
 
-test_that("get_aagis_regions skips downloading if still in tempdir()", {
+test_that("read_aagis_regions skips downloading if still in tempdir()", {
   skip_if_offline()
   skip_on_ci()
   x <- .check_existing_aagis(cache = TRUE)
