+++ tests/testthat/test-read_historical_national_estimates.R
@@ -1,16 +1,16 @@
 
-test_that("get_historical_regional_estimates works", {
+test_that("read_historical_national_estimates works", {
   skip_if_offline()
   skip_on_ci()
-  x <- get_historical_regional_estimates()
-  expect_named(x, c("Variable", "Year", "ABARES_region", "Value", "RSE"))
+  x <- read_historical_national_estimates()
+  expect_named(x, c("Variable", "Year", "Industry", "Value", "RSE"))
   expect_s3_class(x, c("data.table", "data.frame"))
   expect_identical(
     lapply(x, typeof),
     list(
       Variable = "character",
       Year = "integer",
-      ABARES_region = "character",
+      Industry = "character",
       Value = "double",
       RSE = "double"
     )
