+++ tests/testthat/test-read_historical_regional_estimates.R
@@ -1,17 +1,18 @@
-test_that("get_estimates_by_performance_category() works", {
+
+test_that("read_historical_regional_estimates works", {
   skip_if_offline()
   skip_on_ci()
-  x <- get_estimates_by_performance_category()
-  expect_named(x, c("Variable", "Year", "Size", "Value", "RSE"))
+  x <- read_historical_regional_estimates()
+  expect_named(x, c("Variable", "Year", "ABARES_region", "Value", "RSE"))
   expect_s3_class(x, c("data.table", "data.frame"))
   expect_identical(
     lapply(x, typeof),
     list(
       Variable = "character",
       Year = "integer",
-      Size = "character",
+      ABARES_region = "character",
       Value = "double",
-      RSE = "integer"
+      RSE = "double"
     )
   )
 })
