+++ tests/testthat/test-read_historical_state_estimates.R
@@ -1,8 +1,8 @@
 
-test_that("get_historical_state_estimates works", {
+test_that("read_historical_state_estimates works", {
   skip_if_offline()
   skip_on_ci()
-  x <- get_historical_state_estimates()
+  x <- read_historical_state_estimates()
   expect_named(x, c("Variable", "Year", "State", "Industry", "Value", "RSE"))
   expect_s3_class(x, c("data.table", "data.frame"))
   expect_identical(
