### 2024-12-08 38bf7a6
Update NEWS.md for v1.0.0
