### 2024-12-08 dcb262b
Increment version to v1.0.0
