+++ DESCRIPTION
@@ -3,7 +3,7 @@ Package: read.abares
 Title: Provides simple downloading, parsing and importing of Australian
     Bureau of Agricultural and Resource Economics and Sciences (ABARES)
     data sources
-Version: 0.1.0
+Version: 1.0.0
 Authors@R: c(
     person("Adam H.", "Sparks", , "adamhsparks@gmail.com", role = c("cre", "aut"),
            comment = c(ORCID = "0000-0002-0061-8359")),
