### 2024-12-08 8e510ff
Fix nonascii char
