### 2024-12-08 12fef81
Update NEWS.md
