### 2024-12-08 3f927bf
tidy up the README
