+++ README.Rmd
@@ -29,19 +29,14 @@ However, if there is a data set that you feel would be useful to be serviced by 
 
 Data serviced include:
 
-- [ABARES Estimates](https://www.agriculture.gov.au/abares/data/farm-data-portal#data-download);
-  - Historical National Estimates, `read_historical_national_estimate()`;
-  - Historical State Estimates, `read_historical_state_estimates()`;
-  - Historical Regional Estimates, `read_historical_regional_estimates()`;
-  - Estimates by Size, `read_estimates_by_size()`;
-  - Estimates by Performance Category, `read_estimates_by_performance_category()`;
-- the [Australian Gridded Farm Data (AGFD) set](https://www.agriculture.gov.au/abares/research-topics/surveys/farm-survey-data/australian-gridded-farm-data), `get_agfd()`;
-- the [Australian Agricultural and Grazing Industries Survey (AAGIS)](https://www.agriculture.gov.au/abares/research-topics/surveys/farm-survey-data) region mapping files, `get_aagis_regions()`;
-- the [Historical Agricultural Forecast Database](https://www.agriculture.gov.au/abares/research-topics/agricultural-outlook/historical-forecasts#:~:text=About%20the%20historical%20agricultural%20forecast,relevant%20to%20Australian%20agricultural%20markets), `read_historical_forecast_database()`;
-- a [Soil Thickness for Australian areas of intensive agriculture of Layer 1 (A Horizon - top-soil) (derived from soil mapping)](https://data.agriculture.gov.au/geonetwork/srv/eng/catalog.search#/metadata/faa9f157-8e17-4b23-b6a7-37eb7920ead6) map, `get_soil_thickness()` and;
+- the [ABARES Estimates](https://www.agriculture.gov.au/abares/data/farm-data-portal#data-download);
+- the [Australian Gridded Farm Data (AGFD) set](https://www.agriculture.gov.au/abares/research-topics/surveys/farm-survey-data/australian-gridded-farm-data);
+- the [Australian Agricultural and Grazing Industries Survey (AAGIS)](https://www.agriculture.gov.au/abares/research-topics/surveys/farm-survey-data) region mapping files;
+- the [Historical Agricultural Forecast Database](https://www.agriculture.gov.au/abares/research-topics/agricultural-outlook/historical-forecasts#:~:text=About%20the%20historical%20agricultural%20forecast,relevant%20to%20Australian%20agricultural%20markets);
+- the [Soil Thickness for Australian areas of intensive agriculture of Layer 1 (A Horizon - top-soil) (derived from soil mapping)](https://data.agriculture.gov.au/geonetwork/srv/eng/catalog.search#/metadata/faa9f157-8e17-4b23-b6a7-37eb7920ead6) map and;
 - the [ABARES Trade Data](https://www.agriculture.gov.au/abares/research-topics/trade/dashboard) including;
-  - Trade Data, `read_abares_trade()` and;
-  - Trade Region Data, `read_abares_trade_regions()`.
+  - the trade data and;
+  - the trade region data.
 
 The files are freely available as CSV files, zip archives of NetCDF files or a zip archives of geospatial shape files.
 {read.abares} facilitates downloading, caching and importing these files in your R session with your choice of the class of the resulting object(s).
@@ -70,24 +65,15 @@ Caching is not mandatory, you can just work with the downloaded files in `tempdi
 ### Multiple Classes Supported
 
 {read.abares} supports multiple classes of objects to support your workflow.
-Select from spatial classes:
+Select from spatial classes for the AGFD NetCDF files:
 
 - [{stars}](https://CRAN.R-project.org/package=stars),
[…]
