### 2024-12-08 97ff232
Move contributing and COC to .github
