+++ R/read_soil_thickness_stars.R
@@ -1,4 +1,3 @@
-
 #' Read Soil Thickness File With stars
 #'
 #' Read Soil Thickness for Australian Areas of Intensive Agriculture of Layer 1
