+++ tests/testthat/test-read_estimates_by_size.R
@@ -1,4 +1,3 @@
-
 test_that("read_estimates_by_size works", {
   skip_if_offline()
   skip_on_ci()
