+++ R/read_abares_trade_regions.R
@@ -48,16 +48,19 @@ read_abares_trade_regions <- function(cache = TRUE) {
 #' @keywords Internal
 
 .check_existing_trade_regions <- function(cache) {
-  abares_trade_rds <- file.path(.find_user_cache(),
-                                "abares_trade_dir/abares_trade_regions.rds")
+  abares_trade_rds <- file.path(
+    .find_user_cache(),
+    "abares_trade_dir/abares_trade_regions.rds"
+  )
   tmp_csv <- file.path(tempdir(), "abares_trade_regions.csv")
 
   if (file.exists(abares_trade_rds)) {
     return(readRDS(abares_trade_rds))
   } else if (file.exists(tmp_csv)) {
     abares_trade <- data.table::fread(tmp_csv,
-                                      na.strings = c(""),
-                                      fill = TRUE)
+      na.strings = c(""),
+      fill = TRUE
+    )
     if (cache) {
       dir.create(dirname(abares_trade_rds), recursive = TRUE)
       saveRDS(abares_trade, file = abares_trade_rds)
@@ -85,13 +88,17 @@ read_abares_trade_regions <- function(cache = TRUE) {
 .download_abares_trade_regions <- function(cache) {
   # if you make it this far, the cached file doesn't exist, so we need to
   # download it either to `tempdir()` and dispose or cache it
-  cached_csv <- file.path(.find_user_cache(),
-                          "abares_trade_dir/abares_trade_regions.csv")
+  cached_csv <- file.path(
+    .find_user_cache(),
+    "abares_trade_dir/abares_trade_regions.csv"
+  )
   tmp_csv <- file.path(file.path(tempdir(), "abares_trade_regions.csv"))
   abares_trade_csv <- data.table::fifelse(cache, cached_csv, tmp_csv)
   abares_trade_dir <- dirname(abares_trade_csv)
-  abares_trade_regions_rds <- file.path(abares_trade_dir,
-                                        "abares_trade_regions.rds")
+  abares_trade_regions_rds <- file.path(
+    abares_trade_dir,
+    "abares_trade_regions.rds"
+  )
 
   # the user-cache may not exist if caching is enabled for the 1st time
   if (cache && !dir.exists(abares_trade_dir)) {
@@ -100,11 +107,13 @@ read_abares_trade_regions <- function(cache = TRUE) {
 
   .retry_download(
     "https://daff.ent.sirsidynix.net.au/client/en_AU/search/asset/1033841/2",
-    .f = abares_trade_csv)
+    .f = abares_trade_csv
+  )
 
   abares_trade_regions <- data.table::fread(file.path(tmp_csv),
-                                          na.strings = c(""),
-                                          fill = TRUE)
+    na.strings = c(""),
+    fill = TRUE
+  )
 
   if (cache) {
     saveRDS(abares_trade_regions, file = abares_trade_regions_rds)
