+++ man/clear_cache.Rd
@@ -1,25 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/clear_cache.R
-\name{clear_cache}
-\alias{clear_cache}
-\title{Remove Files in Users' Cache Directory}
-\usage{
-clear_cache()
-}
-\value{
-Nothing, called for its side-effects, clearing the cached files
-}
-\description{
-Removes all files in the \pkg{read.abares} cache if any exist.
-}
-\examples{
-# not run because cached files shouldn't exist on CRAN or testing envs
-\dontrun{
-clear_cache()
-}
-}
-\seealso{
-Other cache: 
-\code{\link{inspect_cache}()}
-}
-\concept{cache}
