+++ pkgdown/_pkgdown.yml
@@ -43,8 +43,7 @@ reference:
   desc: |
     Work with your local file cache
   contents:
-  - clear_cache
-  - inspect_cache
+  - manage_cache
 - title: Soil Thickness
   desc: |
     Soil Thickness for Australian areas of intensive agriculture of Layer 1 (A Horizon - top-soil) (derived from soil mapping)
