+++ DESCRIPTION
@@ -27,6 +27,7 @@ Imports:
     brio,
     cli,
     data.table,
+    hoardr,
     httr2,
     lubridate,
     openxlsx2,
