+++ R/get_agfd.R
@@ -235,7 +235,7 @@ get_agfd <- function(fixed_prices = TRUE,
 
 #' Prints read.abares.agfd.nc.files Objects
 #'
-#' Custom [print()] method for `read.abares.agfd.nc.files` objects.
+#' Custom [print] method for `read.abares.agfd.nc.files` objects.
 #'
 #' @param x a `read.abares.agfd.nc.files` object
 #' @param ... ignored
