### 2024-12-18 c6d383f
lint Roxygen and code
