+++ R/read_historical_forecast_database.R
@@ -28,7 +28,7 @@
 #' @references <https://www.agriculture.gov.au/abares/research-topics/agricultural-outlook/historical-forecasts#:~:text=About%20the%20historical%20agricultural%20forecast,relevant%20to%20Australian%20agricultural%20markets>
 #' @source <https://daff.ent.sirsidynix.net.au/client/en_AU/search/asset/1031941/0>
 #'
-#' @return A [data.table::data.table] object
+#' @return A [data.table::data.table] object.
 #'
 #' @autoglobal
 #' @export
