+++ R/read_soil_thickness_terra.R
@@ -1,4 +1,4 @@
-#' Read Soil Thickness File With terra
+#' Read a Soil Thickness for Australian Areas of Intensive Agriculture of Layer 1 file with terra
 #'
 #' Read Soil Thickness for Australian Areas of Intensive Agriculture of Layer 1
 #'  data as a [terra::rast] object.
@@ -9,8 +9,8 @@
 #' @source <https://anrdl-integration-web-catalog-saxfirxkxt.s3-ap-southeast-2.amazonaws.com/warehouse/staiar9cl__059/staiar9cl__05911a01eg_geo___.zip>
 
 #'
-#' @return a [terra::rast] object of the Soil Thickness for Australian Areas of
-#'  Intensive Agriculture of Layer 1
+#' @return a [terra::rast] object of the 'Soil Thickness for Australian Areas of
+#'  Intensive Agriculture of Layer 1'.
 #'
 #' @examplesIf interactive()
 #' x <- get_soil_thickness(cache = TRUE) |>
