+++ man/get_agfd.Rd
@@ -25,7 +25,7 @@ from the most recently completed financial year.}
 using \link[tools:userdir]{tools::R_user_dir} to identify the proper directory for storing user
 data in a cache for this package. Defaults to \code{TRUE}, caching the files
 locally. If \code{FALSE}, this function uses \code{tempdir()} and the files are deleted
-upon closing of the \R session.}
+upon closing of the active \R session.}
 }
 \value{
 A \code{read.abares.agfd.nc.files} object, a \code{list} of NetCDF files
