+++ man/read_abares_trade.Rd
@@ -14,7 +14,7 @@ read_abares_trade(cache = TRUE)
 using \code{tools::R_user_dir()} to identify the proper directory for storing
 user data in a cache for this package. Defaults to \code{TRUE}, caching the files
 locally as a native \R object. If \code{FALSE}, this function uses \code{tempdir()}
-and the files are deleted upon closing of the \R session.}
+and the files are deleted upon closing of the active \R session.}
 }
 \value{
 A \CRANpkg{data.table} object of the \acronym{ABARES} trade data.
