### 2025-01-24 16cd1cb
Simplify downloading and handling of data. Store as gzip for space
