### 2025-01-24 3b36406
update tests, remove redundant tests
