+++ tests/testthat/test-read_abares_trade.R
@@ -2,10 +2,7 @@
 withr::local_envvar(R_USER_CACHE_DIR = file.path(tempdir(), "abares.cache.1"))
 
 # without caching ----
-
 test_that("read_abares_trade doesn't cache", {
-  skip_if_offline()
-  skip_on_ci()
   x <- read_abares_trade(cache = FALSE)
   expect_s3_class(x, c("data.table", "data.frame"))
   expect_named(
@@ -47,27 +44,20 @@ test_that("read_abares_trade doesn't cache", {
     )
   )
   expect_false(file.exists(
-    file.path(.find_user_cache(), "abares_trade_dir/abares_trade.rds")
-  ))
-  expect_false(file.exists(
-    file.path(.find_user_cache(), "abares_trade_dir/abares_trade.csv")
+    file.path(.find_user_cache(), "abares_trade_dir/abares_trade.gz")
   ))
 })
 
 # with caching ----
-
 test_that("read_abares_trade caches", {
-  skip_if_offline()
-  skip_on_ci()
-  x <- read_abares_trade(cache = TRUE)
-  expect_s3_class(x, c("data.table", "data.frame"))
-  y <- list.files(file.path(.find_user_cache(), "abares_trade_dir"))
+  read_abares_trade(cache = TRUE)
   expect_true(file.exists(
-    file.path(.find_user_cache(), "abares_trade_dir/abares_trade.rds")
-  ))
-  expect_false(file.exists(
-    file.path(.find_user_cache(), "abares_trade_dir/abares_trade.zip")
+    file.path(.find_user_cache(), "abares_trade_dir/abares_trade.gz")
   ))
 })
+
+# cleanup cache if rerunning tests in same R session so first test passes ----
+
+unlink(file.path(.find_user_cache(), "abares_trade_dir/abares_trade.gz"))
 
 withr::deferred_run()
