### 2025-01-24 d750df5
simplifies read_abares_trade_regions()
