### 2025-01-24 5258ed8
more simplification of read_abares_trade_regions()
