### 2025-01-24 fc68036
Ensure tests are skipped offline and on CI, mocking won't work here
