+++ tests/testthat/test-read_abares_trade.R
@@ -3,6 +3,8 @@ withr::local_envvar(R_USER_CACHE_DIR = file.path(tempdir(), "abares.cache.1"))
 
 # without caching ----
 test_that("read_abares_trade doesn't cache", {
+  skip_if_offline()
+  skip_on_ci()
   x <- read_abares_trade(cache = FALSE)
   expect_s3_class(x, c("data.table", "data.frame"))
   expect_named(
@@ -50,6 +52,8 @@ test_that("read_abares_trade doesn't cache", {
 
 # with caching ----
 test_that("read_abares_trade caches", {
+  skip_if_offline()
+  skip_on_ci()
   read_abares_trade(cache = TRUE)
   expect_true(file.exists(
     file.path(.find_user_cache(), "abares_trade_dir/abares_trade.gz")
