### 2025-01-24 5a6b6fb
polish and use mocking in tests, WIP
