### 2025-01-24 2f952a5
Tidy up documentation as this is the only `read()` to offer caching
