+++ man/read_abares_trade.Rd
@@ -21,7 +21,9 @@ closing of the active \R session.}
 A \CRANpkg{data.table} object of the \acronym{ABARES} trade data.
 }
 \description{
-Fetches and imports \acronym{ABARES} trade data.
+Fetches and imports \acronym{ABARES} trade data. As the data file is large,
+~1.4GB uncompressed CSV file, caching is offered to save repeated
+downloading.
 }
 \note{
 Columns are renamed for consistency with other \acronym{ABARES} products
