### 2025-01-24 c548961
fixes reading and removes caching, this is a small file that can be redownloaded
