+++ R/read_abares_trade_regions.R
@@ -6,12 +6,6 @@
 #' Columns are renamed for consistency with other \acronym{ABARES} products
 #'  serviced in this package using a snake_case format and ordered consistently.
 #'
-#' @param cache `Boolean` Cache the \acronym{ABARES} trade regions data after
-#'  download using `tools::R_user_dir()` to identify the proper directory for
-#'  storing user data in a cache for this package. Defaults to `TRUE`, caching
-#'  the files locally as a native \R object. If `FALSE`, this function uses
-#'  `tempdir()` and the files are deleted upon closing of the active \R session.
-#'
 #' @examplesIf interactive()
 #' trade_regions <- read_abares_trade_regions()
 #'
@@ -25,55 +19,14 @@
 #' @autoglobal
 #' @export
 
-read_abares_trade_regions <- function(cache = TRUE) {
-  abares_trade_gz <- file.path(
-    .find_user_cache(),
-    "abares_trade_dir/abares_trade_regions.gz"
-  )
-
-  if (file.exists(abares_trade_gz)) {
-    return(data.table::fread(abares_trade_gz))
-  } else {
-    return(.download_abares_trade_regions(cache))
-  }
-}
-
-#' Download the ABARES Trade CSV file
-#'
-#' Handles downloading and caching (if requested) of ABARES Trade data files.
-#'
-#' @param cache `Boolean` Cache the \acronym{ABARES} trade CSV file after
-#'  download using `tools::R_user_dir()` to identify the proper directory for
-#'  storing user data in a cache for this package. Defaults to `TRUE`, caching
-#'  the files locally as a native \R object. If `FALSE`, this function uses
-#'  `tempdir()` and the files are deleted upon closing of the active \R session.
-#'
-#' @return A \CRANpkg{data.table} object of the \acronym{ABARES} trade data.
-#' @noRd
-#' @autoglobal
-#' @keywords Internal
-.download_abares_trade_regions <- function(cache) {
-  abares_trade_regions_dir <- file.path(
-    .find_user_cache(),
-    "abares_trade_dir/"
-  )
-  if (cache && !dir.exists(abares_trade_regions_dir)) {
-    dir.create(abares_trade_regions_dir, recursive = TRUE)
-  }
-  trade_regions_zip <- file.path(tempdir(), "abares_trade_regions_data.zip")
-
+read_abares_trade_regions <- function() {
+  trade_regions <- file.path(tempdir(), "trade_regions")
   .retry_download(
     url = "https://daff.ent.sirsidynix.net.au/client/en_AU/search/asset/1033841/2",
-    .f = trade_regions_zip
+    .f = trade_regions
   )
 
-  abares_trade_regions <- data.table::fread(trade_regions_zip)
-
-  if (cache) {
-    data.table::fwrite(abares_trade_regions,
-      file = file.path(abares_trade_regions_dir, "abares_trade_regions.gz")
-    )
-  }
+  abares_trade_regions <- data.table::fread(trade_regions, fill = TRUE)
 
   return(abares_trade_regions[])
 }
