### 2025-02-02 #667 @maurolepore
@adamhsparks FYI my EiC rotation just started so I'm checking the status of all issues. Seems like a good time to touch base. Should we keep holding?
