### 2025-02-03 aaabb46
fix tests, no caching in regional performance
