### 2025-02-04 #667 @adamhsparks
Hi @maurolepore, in the spirit of the recent discussion around the "holding" tag, let's go ahead with this one.

I've spent several hours trying to make mocking work with the tests and cannot make it reasonably work. In the end, the tests work and pass quite quickly, locally, after the first pass. But I'm open to improvements and hope that reviewers might be able to help me out here. Though I have tidied up a few things with my efforts, so there's that.

I'm happy to share what I've tried and learned along the way for the mocking with the reviewers.
