+++ tests/testthat/test-read_historical_regional_estimates.R
@@ -1,19 +1,17 @@
-with_mock_dir("abares/read_historical_regional_estimates", {
-  test_that("read_historical_regional_estimates works", {
-    skip_if_offline()
-    skip_on_ci()
-    x <- read_historical_regional_estimates()
-    expect_named(x, c("Variable", "Year", "ABARES_region", "Value", "RSE"))
-    expect_s3_class(x, c("data.table", "data.frame"))
-    expect_identical(
-      lapply(x, typeof),
-      list(
-        Variable = "character",
-        Year = "integer",
-        ABARES_region = "character",
-        Value = "double",
-        RSE = "double"
-      )
+test_that("read_historical_regional_estimates works", {
+  skip_if_offline()
+  skip_on_ci()
+  x <- read_historical_regional_estimates()
+  expect_named(x, c("Variable", "Year", "ABARES_region", "Value", "RSE"))
+  expect_s3_class(x, c("data.table", "data.frame"))
+  expect_identical(
+    lapply(x, typeof),
+    list(
+      Variable = "character",
+      Year = "integer",
+      ABARES_region = "character",
+      Value = "double",
+      RSE = "double"
     )
-  })
+  )
 })
