+++ DESCRIPTION
@@ -51,11 +51,11 @@ VignetteBuilder:
     knitr
 Config/roxyglobals/filename: globals.R
 Config/roxyglobals/unique: FALSE
-Config/roxylint:list(linters = roxylint::tidy)
+Config/roxylint: list(linters = roxylint::tidy)
 Config/testthat/edition: 3
 Config/testthat/parallel: true
 Encoding: UTF-8
 Language: en-US
-Roxygen: list(markdown = TRUE, roclets = c("collate", "namespace", "rd",
-    "roxyglobals::global_roclet", "roxylint::roxylint"))
+Roxygen: list(markdown = TRUE, roclets = c("collate", "namespace", "rd", "roxyglobals::global_roclet", "roxylint::roxylint", "devtag::dev_roclet"))
 RoxygenNote: 7.3.2
+Config/Needs/build: moodymudskipper/devtag
