+++ man/dot-retry_download.Rd
@@ -1,0 +1,37 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/internal_functions.R
+\name{.retry_download}
+\alias{.retry_download}
+\title{Use httr2 to fetch a file with retries}
+\usage{
+.retry_download(url, .f, .max_tries = 3L)
+}
+\arguments{
+\item{url}{\code{Character} The URL being requested.}
+
+\item{.f}{\code{Character} A filepath to be written to local storage.}
+
+\item{.max_tries}{\code{Integer} The number of times to retry a failed download
+before emitting an error message.}
+
+\item{.initial_delay}{\code{Integer} The number of seconds to delay before
+retrying the download.  This increases as the tries increment.}
+}
+\value{
+Called for its side-effects, writes an object to the \code{tempdir()} for
+reading into the active \R session later.
+}
+\description{
+Retries to download the requested resource before stopping. Uses
+\CRANpkg{httr2} to cache in-session results in the \code{tempdir()}.
+}
+\examples{
+
+f <- file.path(tempdir(), "fdp-beta-national-historical.csv")
+retry_download(
+  url = "https://www.agriculture.gov.au/sites/default/files/documents/fdp-beta-national-historical.csv",
+  .f = f
+)
+
+}
+\keyword{internal}
