+++ R/internal_functions.R
@@ -2,8 +2,7 @@
 #'
 #' @return A `character` string value of a file path indicating the proper
 #'  directory to use for cached files.
-#' @noRd
-#' @keywords Internal
+#' @dev
 #' @autoglobal
 .find_user_cache <- function() {
   tools::R_user_dir(package = "read.abares", which = "cache")
