+++ man/dot-download_soil_thickness.Rd
@@ -1,0 +1,23 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/get_soil_thickness.R
+\name{.download_soil_thickness}
+\alias{.download_soil_thickness}
+\title{Downloads soil thickness data if not already found locally}
+\usage{
+.download_soil_thickness(cache)
+}
+\arguments{
+\item{cache}{\code{Boolean} Cache the soil thickness data files after download
+using \code{tools::R_user_dir()} to identify the proper directory for storing
+user data in a cache for this package. Defaults to \code{TRUE}, caching the files
+locally. If \code{FALSE}, this function uses \code{tempdir()} and the files are deleted
+upon closing of the active \R session.}
+}
+\value{
+Nothing, called for its side-effects of downloading and unzipping
+files.
+}
+\description{
+Downloads soil thickness data if not already found locally
+}
+\keyword{internal}
