### 2025-02-04 d263385
Fix macro in rd-dev docs
