### 2025-02-04 af98fbf
adds caching for trade region data
