+++ tests/testthat/test-read_abares_trade_regions.R
@@ -1,8 +1,3 @@
-# sets up a custom cache environment in `tempdir()` just for testing
-withr::local_envvar(R_USER_CACHE_DIR = file.path(tempdir(), "abares.cache.1"))
-
-# without caching ----
-
 test_that("read_abares_trade_regions works", {
   skip_if_offline()
   skip_on_ci()
