### 2025-02-05 a5dbae6
fix @returns tag in ROxygen
