+++ R/read_soil_thickness_stars.R
@@ -9,7 +9,7 @@
 #' @references <https://data.agriculture.gov.au/geonetwork/srv/eng/catalog.search#/metadata/faa9f157-8e17-4b23-b6a7-37eb7920ead6>
 #' @source <https://anrdl-integration-web-catalog-saxfirxkxt.s3-ap-southeast-2.amazonaws.com/warehouse/staiar9cl__059/staiar9cl__05911a01eg_geo___.zip>
 #'
-#' @return a [stars] object of the 'Soil Thickness for Australian Areas of
+#' @returns a [stars] object of the 'Soil Thickness for Australian Areas of
 #'  Intensive Agriculture of Layer 1'.
 #'
 #' @examplesIf interactive()
