+++ R/read_agfd_dt.R
@@ -9,7 +9,7 @@
 #' @inherit get_agfd references
 #' @param files A list of \acronym{AGFD} NetCDF files to import.
 #'
-#' @return a [data.table::data.table] object of the 'Australian Gridded Farm
+#' @returns a [data.table::data.table] object of the 'Australian Gridded Farm
 #'  Data'.
 #'
 #' @examplesIf interactive()
