### 2025-02-05 #667 @maurolepore
@adamhsparks thanks! Seems like a great mindset  💯

Great then let's move on to finding a handling editor.

Would you mind summarizing this threat to highlight the main challenge we need to tackle and the expertise you think would help the most? That may help us find a suitable handling editor, and later suitable reviewers.
