### 2025-02-06 #667 @adamhsparks
The biggest issue that this package has that I perceive and from this Issue, is that tests can't reliably run on CI and must use Internet access to download the files necessary for testing. So, the first tests take a while as they download oodles of data, but if they cache the next runs are faster. There are connectivity issues with GitHub and this server, I'm able to reliably access the data from my computers that I can physically interact with but GitHub's CI seems to have issues with downloads from the servers, so I just skipped nearly all tests on CI.

Therefore, I'd love to use the mocking capabilities in {httptest2} in my tests to get around this limitation. And, I'm already using that package to test no connectivity, so it's not an additional package to be added. However, when I've tried to set this up, I've i) managed to get one or two tests to successfully mock as the server properly returns a text file that I can then edit to reduce the size of the file since they are LARGE but, the file-path remains an issue for CRAN as it becomes several layers deep, ii) the server seems to be serving binary files when a CSV file is requested; @jonthegeek had provided me with a suggestion the rOpenSci Slack that I attempted to use and it sorta worked to create an edited file and re-gzip it and insert it into the mocking file but I had more issues with testing then that {httptest2} wasn't happy with the file itself, and iii) some of the files are binary, _i.e._, Excel, and so point ii comes up again. I can edit and try to reinsert but then I sometimes get it to work and sometimes it fails to work.

I'll create a new branch and upload the script I was using to try to do all of this and share that later.
