### 2025-02-06 #667 @adamhsparks
@maurolepore, I've created a new branch, `mocking`, https://github.com/adamhsparks/read.abares/tree/mocking.

In it, the `inst` directory has [edit_mocking_files.R](https://github.com/adamhsparks/read.abares/blob/mocking/inst/edit_mocking_files.R) that has what I've done to try to fix these issues with mocking for reviewers' reference if they like.
