### 2025-02-06 df56333
Add @returns for print_agfd_nc_files()
