### 2025-02-06 #667 @maurolepore
I did a preliminary editor check just to know what I would be passing on to potential handling editor. What I see superficially looks in great shape and suggests that  past the disclosed test-challenge the review should go smoothly.

Ideally we would try to speed up the setup of at least some tests so as to use reviewer's time/internet-connection most effectively.

### Preliminary editor checks:

      > ml01. I see sufficient info to judge de value of the package. Consider opening README with a short explanation for why this package is useful.

  > ml02. I haven't installed myself, but looks good. My personal style would look more like this:
  ```
  # install.packages("pak")
  pak::pak("adamhsparks/read.abares")
  ```
  > ml03. The review issue mentions some test challenges. README too but README doesn't seem like the place for such developer-facing (and hopefully temporary) information.
  > ml04. I haven't run them but I see what seems to be sufficient number of test-files, clearly named, and a random peek suggests they are meaningful and complete.

  > ml05. Although I haven't tried it myself, what I read suggests the setup for tests might be too slow to ask our reviewers to try it. Maybe we could create a tests/testthat/slow/ directory that runs on demand with `testthat::test_dir("tests/testthat/slow")`?

---

---
