### 2025-02-07 #667 @adamhsparks
@maurolepore and @mpadge, I'm puzzled by the failed test. It's passing GitHub CI and locally and passed initial checks but these recent ones fail and it's not something I've touched in my changes that I'm aware of.

On the test coverage note. I've just rechecked it for sake of being open, {read.abares} has 97.36% coverage.
