### 2025-02-08 27bf97a
follow @maurolepore's suggestion for install instructions
