+++ README.md
@@ -1,87 +1,63 @@
+---
+output: github_document
+---
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
+
+
 
 # {read.abares}: Simple downloading and importing of ABARES Data <img src="man/figures/logo.png" align="right"/>
 
 <!-- badges: start -->
-
-[![Lifecycle:
-stable](https://img.shields.io/badge/lifecycle-stable-green.svg)](https://lifecycle.r-lib.org/articles/stages.html#stable)
-[![Project Status: Active – The project has reached a stable, usable
-state and is being actively
-developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
+[![Lifecycle: stable](https://img.shields.io/badge/lifecycle-stable-green.svg)](https://lifecycle.r-lib.org/articles/stages.html#stable)
+[![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
 [![R-CMD-check](https://github.com/adamhsparks/read.abares/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/adamhsparks/read.abares/actions/workflows/R-CMD-check.yaml)
-[![Codecov test
-coverage](https://codecov.io/gh/adamhsparks/read.abares/graph/badge.svg)](https://app.codecov.io/gh/adamhsparks/read.abares)
+[![Codecov test coverage](https://codecov.io/gh/adamhsparks/read.abares/graph/badge.svg)](https://app.codecov.io/gh/adamhsparks/read.abares)
 <!-- badges: end -->
 
-An R package for automated downloading and ingestion of data from the
-Australian Bureau of Agricultural and Resource Economics and Sciences.
-Not all ABARES data are serviced. The list is hand-picked to be
-reasonably useful and maintainable, *i.e.*, frequently updated values
-are not included in this, *e.g.*, [Australian crop
-reports](https://daff.ent.sirsidynix.net.au/client/en_AU/ABARES/search/results?te=ASSET&st=PD#).
-However, if there is a data set that you feel would be useful to be
-serviced by {read.abares}, please feel free to [open an
-issue](https://github.com/adamhsparks/read.abares/issues/new) with
-details about the data set or better yet, open a pull request!
+An R package for automated downloading and ingestion of data from the Australian Bureau of Agricultural and Resource Economics and Sciences.
+Not all ABARES data are serviced.
+The list is hand-picked to be reasonably useful and maintainable, _i.e._, frequently updated values are not included in this, _e.g._, [Australian crop reports](https://daff.ent.sirsidynix.net.au/client/en_AU/ABARES/search/results?te=ASSET&st=PD#).
+However, if there is a data set that you feel would be useful to be serviced by {read.abares}, please feel free to [open an issue](https://github.com/adamhsparks/read.abares/issues/new) with details about the data set or better yet, open a pull request!
 
 Data serviced include:
 
-- the [ABARES
-  Estimates](https://www.agriculture.gov.au/abares/data/farm-data-portal#data-download);
-- the [Australian Gridded Farm Data (AGFD)
-  set](https://www.agriculture.gov.au/abares/research-topics/surveys/farm-survey-data/australian-gridded-farm-data);
-- the [Australian Agricultural and Grazing Industries Survey
-  (AAGIS)](https://www.agriculture.gov.au/abares/research-topics/surveys/farm-survey-data)
-  region mapping files;
-- the [Historical Agricultural Forecast
[…]
