### 2025-02-09 f0db195
move slow tests to `testthat/slow_tests` as per @maurolepore's excellent idea
