### 2025-02-09 619d272
update note on testing protocols
