### 2025-02-09 #667 @adamhsparks
Thanks for your ideas/suggestions, @maurolepore, I've incorporated two of them. I'm not sure where to put the information about the tests, otherwise, so I've updated the instructions for running the slow tests.

* Use {pak} for install: https://github.com/adamhsparks/read.abares/commit/27bf97a05cb570a2ebf5eae5abc4e7f1e891a452
* Move tests to `testthat/slow_tests`,
	* https://github.com/adamhsparks/read.abares/commit/f0db1954ebb4941fd62ddc09a609d1dcf66d5db4 and  	* https://github.com/adamhsparks/read.abares/commit/619d2727ba0c20dbb6faae4b3fd24fe1ec74cef9
