### 2025-02-10 583337a
fix a bad test, file should not exist
