### 2025-02-10 #667 @adamhsparks
@mpadge, I thought `skip_on_ci()` was now recognised here and so the tests wouldn't be run but it appears that they're running when this option is set in my package.
