### 2025-03-02 #667 @obrl-soil
Hi @adamhsparks - I'm unfortunately having a download-timeout problem similar to the discussion about CI issues above, only on my local machine (Surface laptop running win 11). I can download the target files via the browser just fine, and I can't see anything of note when I run `httr2::req_perform(., verbosity = 3)`. Not sure where to go from here. Maybe the httr2 maintainers might have some tips?
