+++ tests/testthat/test-check_class.R
@@ -1,5 +1,6 @@
 test_that("class checking works properly", {
-  expect_error(.check_class(x = 1, "read.abares"),
+  expect_error(
+    .check_class(x = 1, "read.abares"),
     regexp = "You must provide a*"
   )
   expect_silent(.check_class(x = 1, "numeric"))
