### 2025-03-05 #667 @adamhsparks
Hi @obrl-soil, which file is causing the issue?

Unfortunately, I'm unable to duplicate, it's passing all the tests for me locally on my Mac, which means the files are available and downloading through R. With the exception of GitHub CI which isn't really much help in troubleshooting any issues, itworksforme. 🙈

I can see if I can get one of the team to try some of the functions tomorrow or running tests on Windows laptops at work, maybe?
