### 2025-03-06 #667 @obrl-soil
I'd be surprised if it was, this feels like more of a curl/httr/browser(??) type problem. My default browser is Firefox, but I also tried this on a much more locked down work machine (win10 and Edge only) with the same result.

I tried blowing away my entire R/RTools/RStudio setup on my home machine tonight and started from scratch. No luck, but here's the check logs:

````
> library(devtools)
Loading required package: usethis
> check()
══ Documenting ════════════════════════════════════════════════════════════════════════════
ℹ Updating read.abares documentation
ℹ Loading read.abares
✖ get_soil_thickness.R:175: @description refers to unavailable topic pander::pander.
Caused by error in `find.package()`:
! there is no package called ‘pander’
ℹ [get_agfd.R:1] @title should be 'Sentence case'
ℹ [get_agfd.R:38] @param should be 'Sentence case'
ℹ [get_agfd.R:45] @param should be 'Sentence case'
ℹ [get_agfd.R:239] @param should be 'Sentence case'
ℹ [get_agfd.R:240] @param should be 'Sentence case'
ℹ [get_agfd.R:251] @title should be 'Sentence case'
ℹ [get_soil_thickness.R:1] @title should be 'Sentence case'
ℹ [get_soil_thickness.R:77] @param should be 'Sentence case'
ℹ [get_soil_thickness.R:129] @param should be 'Sentence case'
ℹ [get_soil_thickness.R:130] @param should be 'Sentence case'
ℹ [inspect_cache.R:1] @title should be 'Sentence case'
ℹ [internal_functions.R:11] @title should be 'Sentence case'
ℹ [internal_functions.R:14] @param should be 'Sentence case'
ℹ [internal_functions.R:33] @param should be 'Sentence case'
ℹ [internal_functions.R:34] @param should be 'Sentence case'
ℹ [internal_functions.R:35] @param should be 'Sentence case'
ℹ [internal_functions.R:37] @param should be 'Sentence case'
ℹ [read.abares-package.R:2] @seealso should be 'Sentence case'
ℹ [read.abares-package.R:2] @seealso should terminate with a full stop, `.`
ℹ [read_aagis_regions.R:1] @title should be 'Sentence case'
ℹ [read_aagis_regions.R:8] @param should be 'Sentence case'
ℹ [read_aagis_regions.R:69] @title should be 'Sentence case'
ℹ [read_aagis_regions.R:75] @param should be 'Sentence case'
ℹ [read_abares_trade.R:1] @title should be 'Sentence case'
ℹ [read_abares_trade.R:12] @param should be 'Sentence case'
ℹ [read_abares_trade.R:48] @title should be 'Sentence case'
ℹ [read_abares_trade.R:52] @param should be 'Sentence case'
ℹ [read_abares_trade_regions.R:1] @title should be 'Sentence case'
ℹ [read_agfd_dt.R:1] @title should be 'Sentence case'
ℹ [read_agfd_dt.R:10] @param should be 'Sentence case'
ℹ [read_agfd_stars.R:1] @title should be 'Sentence case'
ℹ [read_agfd_terra.R:1] @title should be 'Sentence case'
ℹ [read_agfd_tidync.R:1] @title should be 'Sentence case'
ℹ [read_estimates_by_performance_category.R:1] @title should be 'Sentence case'
ℹ [read_estimates_by_size.R:1] @title should be 'Sentence case'
ℹ [read_historical_forecast_database.R:1] @title should be 'Sentence case'
ℹ [read_historical_national_estimates.R:1] @title should be 'Sentence case'
ℹ [read_historical_regional_estimates.R:1] @title should be 'Sentence case'
ℹ [read_historical_state_estimates.R:1] @title should be 'Sentence case'
ℹ [read_soil_thickness_stars.R:1] @title should be 'Sentence case'
ℹ [read_soil_thickness_terra.R:1] @title should be 'Sentence case'
Writing dot-create_soil_thickness_list.Rd
Writing dot-download_soil_thickness.Rd
Writing dot-find_user_cache.Rd
Writing dot-check_class.Rd
Writing dot-retry_download.Rd
Writing dot-check_existing_aagis.Rd
Writing dot-download_aagis_shp.Rd

══ Building ═══════════════════════════════════════════════════════════════════════════════
Setting env vars:
• CFLAGS    : -Wall -pedantic -fdiagnostics-color=always
• CXXFLAGS  : -Wall -pedantic -fdiagnostics-color=always
• CXX11FLAGS: -Wall -pedantic -fdiagnostics-color=always
• CXX14FLAGS: -Wall -pedantic -fdiagnostics-color=always
• CXX17FLAGS: -Wall -pedantic -fdiagnostics-color=always
• CXX20FLAGS: -Wall -pedantic -fdiagnostics-color=always
── R CMD build ────────────────────────────────────────────────────────────────────────────
✔  checking for file 'C:\Data\code\github\rOpenSci\editorials\read.abares/DESCRIPTION' (573ms)
─  preparing 'read.abares': (671ms)
─  installing the package to build vignettes (390ms)
✔  creating vignettes (24.4s)
─  checking for LF line-endings in source and make files and shell scripts (1.2s)
─  building 'read.abares_1.0.0.tar.gz'

══ Checking ═══════════════════════════════════════════════════════════════════════════════
Setting env vars:
• _R_CHECK_CRAN_INCOMING_REMOTE_               : FALSE
• _R_CHECK_CRAN_INCOMING_                      : FALSE
• _R_CHECK_FORCE_SUGGESTS_                     : FALSE
• _R_CHECK_PACKAGES_USED_IGNORE_UNUSED_IMPORTS_: FALSE
• NOT_CRAN                                     : true
── R CMD check ────────────────────────────────────────────────────────────────────────────
─  using log directory 'C:/Users/leobr/AppData/Local/Temp/RtmpghvyIL/file14cc2703fd1/read.abares.Rcheck' (429ms)
─  using R version 4.4.3 (2025-02-28 ucrt)
─  using platform: x86_64-w64-mingw32
─  R was compiled by
       gcc.exe (GCC) 13.3.0
       GNU Fortran (GCC) 13.3.0
─  running under: Windows 11 x64 (build 26100)
─  using session charset: UTF-8
─  using options '--no-manual --as-cran'
✔  checking for file 'read.abares/DESCRIPTION' ...
─  checking extension type ... Package
─  this is package 'read.abares' version '1.0.0'
─  package encoding: UTF-8
✔  checking package namespace information ...
N  checking package dependencies (3.5s)
   Packages suggested but not available for checking: 'pander', 'spelling'
✔  checking if this is a source package ...
✔  checking if there is a namespace
✔  checking for executable files (1.6s)
✔  checking for hidden files and directories ...
✔  checking for portable file names ...
─  checking whether package 'read.abares' can be installed ... [25s] OK (25.3s)
✔  checking installed package size ...
✔  checking package directory (431ms)
N  checking for future file timestamps ...
   unable to verify current time
✔  checking 'build' directory
✔  checking DESCRIPTION meta-information (638ms)
N  checking top-level files
   Non-standard file/directory found at top level:
     'air.toml'
✔  checking for left-over files
✔  checking index information (531ms)
✔  checking package subdirectories (2.4s)
✔  checking code files for non-ASCII characters ...
✔  checking R files for syntax errors ...
✔  checking whether the package can be loaded (6.1s)
✔  checking whether the package can be loaded with stated dependencies (5.5s)
✔  checking whether the package can be unloaded cleanly (5.9s)
✔  checking whether the namespace can be loaded with stated dependencies (5.9s)
✔  checking whether the namespace can be unloaded cleanly (6.9s)
✔  checking loading without being on the library search path (7.2s)
✔  checking dependencies in R code (13.9s)
✔  checking S3 generic/method consistency (6s)
✔  checking replacement functions (5.8s)
✔  checking foreign function calls (6.3s)
─  checking R code for possible problems ... [33s] OK (33.4s)
✔  checking Rd files (1s)
✔  checking Rd metadata ...
✔  checking Rd line widths (341ms)
N  checking Rd cross-references (5.5s)
   Package unavailable to check Rd xrefs: 'pander'
✔  checking for missing documentation entries (7.2s)
✔  checking for code/documentation mismatches (18.4s)
✔  checking Rd \usage sections (6.8s)
✔  checking Rd contents (449ms)
✔  checking for unstated dependencies in examples (666ms)
✔  checking installed files from 'inst/doc' (664ms)
✔  checking files in 'vignettes' (661ms)
✔  checking examples (10.7s)
✔  checking for unstated dependencies in 'tests' ...
─  checking tests (406ms)
─  Running 'spelling.R' (457ms)
   Running the tests in 'tests/testthat.R' failed.
   Last 13 lines of output:
       1. └─read.abares::read_historical_state_estimates() at test-read_historical_state_estimates.R:4:3
       2.   └─read.abares:::.retry_download(...)
       3.     ├─brio::write_file_raw(...)
       4.     ├─httr2::resp_body_raw(...)
       5.     │ └─httr2:::check_response(resp)
       6.     │   └─httr2:::is_response(resp)
       7.     └─httr2::req_perform(...)
       8.       └─base::tryCatch(...)
       9.         └─base (local) tryCatchList(expr, classes, parentenv, handlers)
      10.           └─base (local) tryCatchOne(expr, names, parentenv, handlers[[1L]])
      11.             └─value[[3L]](cond)

     [ FAIL 11 | WARN 0 | SKIP 0 | PASS 20 ]
     Error: Test failures
     Execution halted
✔  checking for unstated dependencies in vignettes (4.9s)
✔  checking package vignettes ...
N  checking re-building of vignette outputs (8.1s)
   Note: skipping 'read.abares.Rmd' due to unavailable dependencies: 'pander'
✔  checking for non-standard things in the check directory
✔  checking for detritus in the temp directory ...

   See
     'C:/Users/leobr/AppData/Local/Temp/RtmpghvyIL/file14cc2703fd1/read.abares.Rcheck/00check.log'
   for details.

   ERROR: Unknown command "TMPDIR=C:/Users/leobr/AppData/Local/Temp/Rtmp6n5EqG/file8945a3e2a0c". Did you mean command "create"?
   Warning in system2("quarto", "-V", stdout = TRUE, env = paste0("TMPDIR=",  :
     running command '"quarto" TMPDIR=C:/Users/leobr/AppData/Local/Temp/Rtmp6n5EqG/file8945a3e2a0c -V' had status 1

── R CMD check results ───────────────────────────────────────────── read.abares 1.0.0 ────
Duration: 48m 59.7s

❯ checking tests ...
  See below...

❯ checking package dependencies ... NOTE
  Packages suggested but not available for checking: 'pander', 'spelling'

❯ checking for future file timestamps ... NOTE
  unable to verify current time

❯ checking top-level files ... NOTE
  Non-standard file/directory found at top level:
    'air.toml'

❯ checking Rd cross-references ... NOTE
  Package unavailable to check Rd xrefs: 'pander'

❯ checking re-building of vignette outputs ... NOTE
  Note: skipping 'read.abares.Rmd' due to unavailable dependencies: 'pander'

── Test failures ──────────────────────────────────────────────────────────── testthat ────

> # This file is part of the standard setup for testthat.
> # It is recommended that you do not modify it.
> #
> # Where should you do additional test configuration?
> # Learn more about the roles of various files in:
> # * https://r-pkgs.org/testing-design.html#sec-tests-files-overview
> # * https://testthat.r-lib.org/articles/special-files.html
>
> library(testthat)
> library(read.abares)

Attaching package: 'read.abares'

The following object is masked from 'package:graphics':

    plot

The following object is masked from 'package:base':

    plot

>
> test_check("read.abares")
Starting 2 test processes
[ FAIL 11 | WARN 0 | SKIP 0 | PASS 20 ]

── Error ('test-read_estimates_by_performance_category.R:4:3'): read_estimates_by_performance_category() works ──
<httr2_failure/httr2_error/rlang_error/error/condition>
Error in `httr2::req_perform(httr2::req_progress(httr2::req_cache(httr2::req_retry(httr2::req_options(httr2::req_user_agent(httr2::request(base_url = url),
    "read.abares"), http_version = 2, timeout = 2000L), max_tries = .max_tries),
    path = tempdir())))`: Failed to perform HTTP request.
Caused by error in `curl::curl_fetch_memory()`:
! Timeout was reached [www.agriculture.gov.au]: Operation too slow. Less than 1 bytes/sec transferred the last 600 seconds
Backtrace:
     ▆
  1. └─read.abares::read_estimates_by_performance_category() at test-read_estimates_by_performance_category.R:4:3
  2.   └─read.abares:::.retry_download(...)
  3.     ├─brio::write_file_raw(...)
  4.     ├─httr2::resp_body_raw(...)
  5.     │ └─httr2:::check_response(resp)
  6.     │   └─httr2:::is_response(resp)
  7.     └─httr2::req_perform(...)
  8.       └─base::tryCatch(...)
  9.         └─base (local) tryCatchList(expr, classes, parentenv, handlers)
 10.           └─base (local) tryCatchOne(expr, names, parentenv, handlers[[1L]])
 11.             └─value[[3L]](cond)
── Error ('test-read_estimates_by_size.R:4:3'): read_estimates_by_size works ───
<httr2_failure/httr2_error/rlang_error/error/condition>
Error in `httr2::req_perform(httr2::req_progress(httr2::req_cache(httr2::req_retry(httr2::req_options(httr2::req_user_agent(httr2::request(base_url = url),
    "read.abares"), http_version = 2, timeout = 2000L), max_tries = .max_tries),
    path = tempdir())))`: Failed to perform HTTP request.
Caused by error in `curl::curl_fetch_memory()`:
! Timeout was reached [www.agriculture.gov.au]: Send failure: Connection was aborted
Backtrace:
     ▆
  1. └─read.abares::read_estimates_by_size() at test-read_estimates_by_size.R:4:3
  2.   └─read.abares:::.retry_download(...)
  3.     ├─brio::write_file_raw(...)
  4.     ├─httr2::resp_body_raw(...)
  5.     │ └─httr2:::check_response(resp)
  6.     │   └─httr2:::is_response(resp)
  7.     └─httr2::req_perform(...)
  8.       └─base::tryCatch(...)
  9.         └─base (local) tryCatchList(expr, classes, parentenv, handlers)
 10.           └─base (local) tryCatchOne(expr, names, parentenv, handlers[[1L]])
 11.             └─value[[3L]](cond)
── Error ('test-read_aagis_regions.R:8:3'): read_aagis_regions doesn't cache ───
<httr2_failure/httr2_error/rlang_error/error/condition>
Error in `httr2::req_perform(httr2::req_progress(httr2::req_cache(httr2::req_retry(httr2::req_options(httr2::req_user_agent(httr2::request(base_url = url),
    "read.abares"), http_version = 2, timeout = 2000L), max_tries = .max_tries),
    path = tempdir())))`: Failed to perform HTTP request.
Caused by error in `curl::curl_fetch_memory()`:
! Timeout was reached [www.agriculture.gov.au]: Operation too slow. Less than 1 bytes/sec transferred the last 600 seconds
Backtrace:
     ▆
  1. └─read.abares::read_aagis_regions(cache = FALSE) at test-read_aagis_regions.R:8:3
  2.   └─read.abares:::.download_aagis_shp(cache)
  3.     └─read.abares:::.retry_download(...)
  4.       ├─brio::write_file_raw(...)
  5.       ├─httr2::resp_body_raw(...)
  6.       │ └─httr2:::check_response(resp)
  7.       │   └─httr2:::is_response(resp)
  8.       └─httr2::req_perform(...)
  9.         └─base::tryCatch(...)
 10.           └─base (local) tryCatchList(expr, classes, parentenv, handlers)
 11.             └─base (local) tryCatchOne(expr, names, parentenv, handlers[[1L]])
 12.               └─value[[3L]](cond)
── Failure ('test-read_aagis_regions.R:20:3'): read_aagis_regions skips downloading if still in tempdir() ──
`x` is not an S3 object
── Error ('test-read_aagis_regions.R:28:3'): read_aagis_regions caches ─────────
<httr2_failure/httr2_error/rlang_error/error/condition>
Error in `httr2::req_perform(httr2::req_progress(httr2::req_cache(httr2::req_retry(httr2::req_options(httr2::req_user_agent(httr2::request(base_url = url),
    "read.abares"), http_version = 2, timeout = 2000L), max_tries = .max_tries),
    path = tempdir())))`: Failed to perform HTTP request.
Caused by error in `curl::curl_fetch_memory()`:
! Timeout was reached [www.agriculture.gov.au]: Operation too slow. Less than 1 bytes/sec transferred the last 600 seconds
Backtrace:
     ▆
  1. └─read.abares::read_aagis_regions(cache = TRUE) at test-read_aagis_regions.R:28:3
  2.   └─read.abares:::.download_aagis_shp(cache)
  3.     └─read.abares:::.retry_download(...)
  4.       ├─brio::write_file_raw(...)
  5.       ├─httr2::resp_body_raw(...)
  6.       │ └─httr2:::check_response(resp)
  7.       │   └─httr2:::is_response(resp)
  8.       └─httr2::req_perform(...)
  9.         └─base::tryCatch(...)
 10.           └─base (local) tryCatchList(expr, classes, parentenv, handlers)
 11.             └─base (local) tryCatchOne(expr, names, parentenv, handlers[[1L]])
 12.               └─value[[3L]](cond)
── Failure ('test-read_aagis_regions.R:46:3'): read_aagis_regions skips downloading if cache is available ──
`x` is not an S3 object
── Error ('test-read_aagis_regions.R:56:3'): read_aagis_regions does cache ─────
<httr2_failure/httr2_error/rlang_error/error/condition>
Error in `httr2::req_perform(httr2::req_progress(httr2::req_cache(httr2::req_retry(httr2::req_options(httr2::req_user_agent(httr2::request(base_url = url),
    "read.abares"), http_version = 2, timeout = 2000L), max_tries = .max_tries),
    path = tempdir())))`: Failed to perform HTTP request.
Caused by error in `curl::curl_fetch_memory()`:
! Timeout was reached [www.agriculture.gov.au]: Operation too slow. Less than 1 bytes/sec transferred the last 600 seconds
Backtrace:
     ▆
  1. └─read.abares::read_aagis_regions(cache = TRUE) at test-read_aagis_regions.R:56:3
  2.   └─read.abares:::.download_aagis_shp(cache)
  3.     └─read.abares:::.retry_download(...)
  4.       ├─brio::write_file_raw(...)
  5.       ├─httr2::resp_body_raw(...)
  6.       │ └─httr2:::check_response(resp)
  7.       │   └─httr2:::is_response(resp)
  8.       └─httr2::req_perform(...)
  9.         └─base::tryCatch(...)
 10.           └─base (local) tryCatchList(expr, classes, parentenv, handlers)
 11.             └─base (local) tryCatchOne(expr, names, parentenv, handlers[[1L]])
 12.               └─value[[3L]](cond)
── Failure ('test-read_aagis_regions.R:73:3'): read_aagis_regions skips downloading if still in tempdir() ──
`x` is not an S3 object
── Error ('test-read_historical_national_estimates.R:4:3'): read_historical_national_estimates works ──
<httr2_failure/httr2_error/rlang_error/error/condition>
Error in `httr2::req_perform(httr2::req_progress(httr2::req_cache(httr2::req_retry(httr2::req_options(httr2::req_user_agent(httr2::request(base_url = url),
    "read.abares"), http_version = 2, timeout = 2000L), max_tries = .max_tries),
    path = tempdir())))`: Failed to perform HTTP request.
Caused by error in `curl::curl_fetch_memory()`:
! Timeout was reached [www.agriculture.gov.au]: Operation too slow. Less than 1 bytes/sec transferred the last 600 seconds
Backtrace:
     ▆
  1. └─read.abares::read_historical_national_estimates() at test-read_historical_national_estimates.R:4:3
  2.   └─read.abares:::.retry_download(...)
  3.     ├─brio::write_file_raw(...)
  4.     ├─httr2::resp_body_raw(...)
  5.     │ └─httr2:::check_response(resp)
  6.     │   └─httr2:::is_response(resp)
  7.     └─httr2::req_perform(...)
  8.       └─base::tryCatch(...)
  9.         └─base (local) tryCatchList(expr, classes, parentenv, handlers)
 10.           └─base (local) tryCatchOne(expr, names, parentenv, handlers[[1L]])
 11.             └─value[[3L]](cond)
── Error ('test-read_historical_regional_estimates.R:4:3'): read_historical_regional_estimates works ──
<httr2_failure/httr2_error/rlang_error/error/condition>
Error in `httr2::req_perform(httr2::req_progress(httr2::req_cache(httr2::req_retry(httr2::req_options(httr2::req_user_agent(httr2::request(base_url = url),
    "read.abares"), http_version = 2, timeout = 2000L), max_tries = .max_tries),
    path = tempdir())))`: Failed to perform HTTP request.
Caused by error in `curl::curl_fetch_memory()`:
! Timeout was reached [www.agriculture.gov.au]: Operation too slow. Less than 1 bytes/sec transferred the last 600 seconds
Backtrace:
     ▆
  1. └─read.abares::read_historical_regional_estimates() at test-read_historical_regional_estimates.R:4:3
  2.   └─read.abares:::.retry_download(...)
  3.     ├─brio::write_file_raw(...)
  4.     ├─httr2::resp_body_raw(...)
  5.     │ └─httr2:::check_response(resp)
  6.     │   └─httr2:::is_response(resp)
  7.     └─httr2::req_perform(...)
  8.       └─base::tryCatch(...)
  9.         └─base (local) tryCatchList(expr, classes, parentenv, handlers)
 10.           └─base (local) tryCatchOne(expr, names, parentenv, handlers[[1L]])
 11.             └─value[[3L]](cond)
── Error ('test-read_historical_state_estimates.R:4:3'): read_historical_state_estimates works ──
<httr2_failure/httr2_error/rlang_error/error/condition>
Error in `httr2::req_perform(httr2::req_progress(httr2::req_cache(httr2::req_retry(httr2::req_options(httr2::req_user_agent(httr2::request(base_url = url),
[…]
