### 2025-03-06 #667 @adamhsparks
@obrl-soil, I'm at a loss here. I don't know what to actually do here, I'm open to suggestions since I'm not able to reproduce this except on GitHub CI. I thought that was just the way GitHub had stuff configured. I don't know how to do any troubleshooting with this. I thought I'd dealt with the flakey connection and slow responses, which in all my testing I had and trying on a Windows machine (I use Mac exclusively) allowed me to rule out a platform issue. So I'm unsure where to look if it's not a curl lib issue on Windows, my first thought when you raised the issue here, I'm not sure what it is. I've never seen anything like it.

@maurolepore, do you have any ideas what we can do here?
