### 2025-03-06 #667 @mpadge
@adamhsparks FYI I'd suggest the key seems to be the line:
```
! Timeout was reached [www.agriculture.gov.au]: Operation too slow. Less than 1 bytes/sec transferred the last 600 seconds
```
This is similar to behaviour that can happen with the server used in the {osmdata} package, which has very clever throttling mechanisms dependent on a host of factors. Maybe the API you're calling throttles requests from GitHub IP addresses? In which case, you won't be able to do anything about it except re-design tests to avoid downloading ... again 😱 )
