### 2025-03-06 #667 @adamhsparks
@mpadge, I've given up with trying to DL from GitHub, this is a problem for @obrl-soil who can download the files with her browser but not the package, I'm stumped.
