### 2025-03-06 #667 @mpadge
Browser has a different user agent. Maybe they're even throttling everything that is not a standard browser user agent? You can fake user agents in the package to pretend it's a browser, but that just starts a game of cat-and-mouse.
