### 2025-03-06 #667 @adamhsparks
I thought about that, but it works on every computer I’ve tried with the default user agent for {httr2}.
