### 2025-03-06 6fdc112
remove custom user-agent to see if that fixes DL issues
