### 2025-03-07 #667 @obrl-soil
I had previously tried messing around with the `.retry_download()` function because I wondered whether the use of piped code was causing a problem (no) and as part of that I did try skipping the user agent at one point. Didn't help.

I also tried the SSL mod described [here](https://cran.r-project.org/web/packages/curl/vignettes/windows.html), no change.

Is it possible this is a server-specific issue? @adamhsparks do you have any contacts at ABARES' IT dept?
