### 2025-03-07 #667 @adamhsparks
I don’t know anyone at ABARES, no.

I’m baffled. It works in a browser but suddenly not here and no real error either. I wish I could recreate this. Maybe using curl from the CLI would be insightful. I can try that tomorrow and see what happens if I’m not using R, I guess.
