### 2025-03-07 #667 @obrl-soil
This is definitely domain-specific. I've tested every URL stored in your package. The ones linking to the ABARES publications library at `https://daff.ent.sirsidynix.net.au` are fine, as is the link to their AWS-hosted soil thickness data. Only the data stored in `https://www.agriculture.gov.au/sites/default/files/documents/` is a problem. FWIW I also have no problems downloading from e.g. data.gov.au or various other state-based portals.

I'm of the opinion that the datasets stored on the website should be moved into the ABARES publications library or onto data.gov.au as a matter of best practice (FAIR principles, proper metadata management etc) but you'd have to take that up with the site admins. If they don't action your request, you may be down to one or both of the following options:

- re-host copies of the affected files yourself, and take on the maintenance burden of keeping your version of the data up to date
- Change the workflow for the affected datasets to handle the issue. E.g. if the download times out, have a message print to the console describing the problem and directing the user to download the data manually. `read.abares` could then be used to do the data cleaning steps you've built in to the package. Less maintenance for you but more overhead for the user.

Not ideal, I know :/
