### 2025-03-07 #667 @adamhsparks
Not ideal, but useful info. Let me see what I can find.
