### 2025-03-07 #667 @adamhsparks
@obrl-soil, can you use curl and give me the output of, `curl -vO https://www.agriculture.gov.au/sites/default/files/documents/fdp-national-historical.csv`?

This is what I see on my machine.

```fish
~/tmp via  v3.4.1
♥ ❯ curl -vO https://www.agriculture.gov.au/sites/default/files/documents/fdp-national-historical.csv
                                 Dload  Upload   Total   Spent    Left  Speed
  0     0    0     0    0     0      0      0 --:--:-- --:--:-- --:--:--     0* Host www.agriculture.gov.au:443 was resolved.
* IPv6: 2600:1415:11::1737:f29a, 2600:1415:11::1737:f283
* IPv4: 184.84.150.48, 184.84.150.145
*   Trying 184.84.150.48:443...
* Connected to www.agriculture.gov.au (184.84.150.48) port 443
* ALPN: curl offers h2,http/1.1
* (304) (OUT), TLS handshake, Client hello (1):
} [327 bytes data]
*  CAfile: /etc/ssl/cert.pem
*  CApath: none
* (304) (IN), TLS handshake, Server hello (2):
{ [122 bytes data]
* (304) (IN), TLS handshake, Unknown (8):
{ [29 bytes data]
* (304) (IN), TLS handshake, Certificate (11):
{ [2616 bytes data]
* (304) (IN), TLS handshake, CERT verify (15):
{ [264 bytes data]
* (304) (IN), TLS handshake, Finished (20):
{ [36 bytes data]
* (304) (OUT), TLS handshake, Finished (20):
} [36 bytes data]
* SSL connection using TLSv1.3 / AEAD-CHACHA20-POLY1305-SHA256 / [blank] / UNDEF
* ALPN: server accepted h2
* Server certificate:
*  subject: CN=www.agriculture.gov.au
*  start date: Mar  6 10:20:27 2025 GMT
*  expire date: Jun  4 10:20:26 2025 GMT
*  subjectAltName: host "www.agriculture.gov.au" matched cert's "www.agriculture.gov.au"
*  issuer: C=US; O=Let's Encrypt; CN=R11
*  SSL certificate verify ok.
* using HTTP/2
* [HTTP/2] [1] OPENED stream for https://www.agriculture.gov.au/sites/default/files/documents/fdp-national-historical.csv
* [HTTP/2] [1] [:method: GET]
* [HTTP/2] [1] [:scheme: https]
* [HTTP/2] [1] [:authority: www.agriculture.gov.au]
* [HTTP/2] [1] [:path: /sites/default/files/documents/fdp-national-historical.csv]
* [HTTP/2] [1] [user-agent: curl/8.7.1]
* [HTTP/2] [1] [accept: */*]
> GET /sites/default/files/documents/fdp-national-historical.csv HTTP/2
> Host: www.agriculture.gov.au
> User-Agent: curl/8.7.1
> Accept: */*
>
* Request completely sent off
< HTTP/2 200
< expires: Sat, 08 Mar 2025 00:10:01 GMT
< last-modified: Mon, 02 Dec 2024 23:51:40 GMT
< etag: "674e480c-1742c1"
< cache-control: max-age=1800
< x-frame-options: SAMEORIGIN
< x-robots-tag: all
< x-content-type-options: nosniff
< strict-transport-security: max-age=31536000; includeSubDomains
< x-lagoon: amazeeio-govcms5>sigsci-ingress-nginx>daff-master:nginx>
< content-type: application/octet-stream
< date: Fri, 07 Mar 2025 23:41:14 GMT
<
{ [16375 bytes data]
100 1488k    0 1488k    0     0  2585k      0 --:--:-- --:--:-- --:--:-- 2589k
* Connection #0 to host www.agriculture.gov.au left intact
```
