### 2025-03-08 #667 @obrl-soil
Sure,
```
C:\Users\leobr>curl -vO https://www.agriculture.gov.au/sites/default/files/documents/fdp-national-historical.csv
                                 Dload  Upload   Total   Spent    Left  Speed
  0     0    0     0    0     0      0      0 --:--:-- --:--:-- --:--:--     0* Host www.agriculture.gov.au:443 was resolved.
* IPv6: (none)
* IPv4: 131.203.3.10
  0     0    0     0    0     0      0      0 --:--:-- --:--:-- --:--:--     0*   Trying 131.203.3.10:443...
* schannel: disabled automatic use of client certificate
* ALPN: curl offers http/1.1
* ALPN: server accepted http/1.1
* Connected to www.agriculture.gov.au (131.203.3.10) port 443
* using HTTP/1.x
> GET /sites/default/files/documents/fdp-national-historical.csv HTTP/1.1
> Host: www.agriculture.gov.au
> User-Agent: curl/8.10.1
> Accept: */*
>
* schannel: remote party requests renegotiation
* schannel: renegotiating SSL/TLS connection
* schannel: SSL/TLS connection renegotiated
* schannel: remote party requests renegotiation
* schannel: renegotiating SSL/TLS connection
* schannel: SSL/TLS connection renegotiated
* Request completely sent off
  0     0    0     0    0     0      0      0 --:--:--  0:10:00 --:--:--     0* Recv failure: Connection was reset
* schannel: recv returned CURLE_RECV_ERROR
  0     0    0     0    0     0      0      0 --:--:--  0:10:00 --:--:--     0
* closing connection #0
curl: (56) Recv failure: Connection was reset
```
