+++ tests/testthat/test-read_estimates_by_performance_category.R
@@ -1,6 +1,5 @@
 test_that("read_estimates_by_performance_category() works", {
   skip_if_offline()
-  skip_on_ci()
   x <- read_estimates_by_performance_category()
   expect_named(x, c("Variable", "Year", "Size", "Value", "RSE"))
   expect_s3_class(x, c("data.table", "data.frame"))
