### 2025-03-09 #667 @adamhsparks
Ok, this clearly is not a {read.abares} package issue given that it fails with curl from the CLI for specific computers. So, based on the logs provided, I have a faint memory of having issues with Windows, {curl}/curl and schannel when writing {weatherOz} and we changed the backend and it fixed our issues (locally). @obrl-soil, can we try changing the curl backend on your machine machine, _e.g., https://jeroen.r-universe.dev/articles/curl/windows.html, and see if that fixes the issues we're seeing?
