### 2025-03-10 #667 @obrl-soil
I already tried that in R without success, and setting a system-wide environment variable didn't change anything. I did notice a potential issue with multiple versions of `curl` (see your slack thread). Running the following on the command line to specifically invoke the Rtools instance of curl produced another failure, but at least its a new and different one?

```
c:\Data>C:\rtools44\usr\bin\curl.exe -vO https://www.agriculture.gov.au/sites/default/files/documents/fdp-national-historical.csv
                                 Dload  Upload   Total   Spent    Left  Speed
  0     0    0     0    0     0      0      0 --:--:-- --:--:-- --:--:--     0* Host www.agriculture.gov.au:443 was resolved.
* IPv6: 2404:4408:5:51cc::83cb:30a, 2404:4408:5:51cc::83cb:309
* IPv4: 131.203.3.10, 131.203.3.9
*   Trying [2404:4408:5:51cc::83cb:30a]:443...
* connect to 2404:4408:5:51cc::83cb:30a port 443 from :: port 50129 failed: Network is unreachable
*   Trying [2404:4408:5:51cc::83cb:309]:443...
* connect to 2404:4408:5:51cc::83cb:309 port 443 from :: port 50130 failed: Network is unreachable
*   Trying 131.203.3.10:443...
* ALPN: curl offers h2,http/1.1
} [5 bytes data]
* TLSv1.3 (OUT), TLS handshake, Client hello (1):
} [512 bytes data]
*  CAfile: /usr/ssl/certs/ca-bundle.crt
*  CApath: none
{ [5 bytes data]
* TLSv1.3 (IN), TLS handshake, Server hello (2):
{ [122 bytes data]
* TLSv1.3 (IN), TLS change cipher, Change cipher spec (1):
{ [1 bytes data]
* TLSv1.3 (IN), TLS handshake, Encrypted Extensions (8):
{ [29 bytes data]
* TLSv1.3 (IN), TLS handshake, Certificate (11):
{ [2616 bytes data]
* TLSv1.3 (IN), TLS handshake, CERT verify (15):
{ [264 bytes data]
* TLSv1.3 (IN), TLS handshake, Finished (20):
{ [52 bytes data]
* TLSv1.3 (OUT), TLS change cipher, Change cipher spec (1):
} [1 bytes data]
* TLSv1.3 (OUT), TLS handshake, Finished (20):
} [52 bytes data]
* SSL connection using TLSv1.3 / TLS_AES_256_GCM_SHA384 / x25519 / RSASSA-PSS
* ALPN: server accepted h2
* Server certificate:
*  subject: CN=www.agriculture.gov.au
*  start date: Mar  6 10:20:27 2025 GMT
*  expire date: Jun  4 10:20:26 2025 GMT
*  subjectAltName: host "www.agriculture.gov.au" matched cert's "www.agriculture.gov.au"
*  issuer: C=US; O=Let's Encrypt; CN=R11
*  SSL certificate verify ok.
*   Certificate level 0: Public key type RSA (2048/112 Bits/secBits), signed using sha256WithRSAEncryption
*   Certificate level 1: Public key type RSA (2048/112 Bits/secBits), signed using sha256WithRSAEncryption
*   Certificate level 2: Public key type RSA (4096/152 Bits/secBits), signed using sha256WithRSAEncryption
* Connected to www.agriculture.gov.au (131.203.3.10) port 443
* using HTTP/2
* [HTTP/2] [1] OPENED stream for https://www.agriculture.gov.au/sites/default/files/documents/fdp-national-historical.csv
* [HTTP/2] [1] [:method: GET]
* [HTTP/2] [1] [:scheme: https]
* [HTTP/2] [1] [:authority: www.agriculture.gov.au]
* [HTTP/2] [1] [:path: /sites/default/files/documents/fdp-national-historical.csv]
* [HTTP/2] [1] [user-agent: curl/8.12.1]
* [HTTP/2] [1] [accept: */*]
} [5 bytes data]
> GET /sites/default/files/documents/fdp-national-historical.csv HTTP/2
> Host: www.agriculture.gov.au
> User-Agent: curl/8.12.1
> Accept: */*
>
* Request completely sent off
{ [5 bytes data]
* TLSv1.3 (IN), TLS handshake, Newsession Ticket (4):
{ [265 bytes data]
* TLSv1.3 (IN), TLS handshake, Newsession Ticket (4):
{ [265 bytes data]
* HTTP/2 stream 1 was not closed cleanly: INTERNAL_ERROR (err 2)
} [5 bytes data]
* Connection #0 to host www.agriculture.gov.au left intact
curl: (92) HTTP/2 stream 1 was not closed cleanly: INTERNAL_ERROR (err 2)
```
