### 2025-03-11 9ba2d9b
Revert "remove custom user-agent to see if that fixes DL issues"
This reverts commit 6fdc11267f258496bc705a80a0bd50ab731e81b8.
