### 2025-03-11 f33bffa
remove explicit returns where possible and try to fix a doc link issue
