### 2025-03-11 3c01e09
remove debug-shell.yaml
