### 2025-03-11 b0632df
try fixing DL issue via https://forum.posit.co/t/httr2-request-fails-ok-in-browser/197965/3
