### 2025-03-11 9ab045a
fixes bug in read_estimates_by_performance_category()
