### 2025-03-11 405fb64
delete duplicated req_perform()
