### 2025-03-11 #667 @adamhsparks
Ok, so, I just had a browse over my e-mail notification of discussions on the Posit forum and this gem popped up, <https://forum.posit.co/t/httr2-request-fails-ok-in-browser/197965/2>, look familiar? I thought so.

So, I've tried the suggested fix therein,

[Lines 54:57](https://github.com/adamhsparks/read.abares/blob/3be848125c387941425d77181fa7874c0445b133/R/internal_functions.R#L54C4-L57C61) of internal_functions.R are now:

```r
 httr2::req_user_agent("Mozilla/5.0") |>
    httr2::req_headers("Accept-Encoding" = "identity") |>
    httr2::req_headers("Connection" = "Keep-Alive") |>
    httr2::req_options(http_version = 2, timeout = 2000L) |>
```

Keeping the `http_version = 2, timeout = 2000L)` that I already had in place. This again worksforme. 🙃
