### 2025-03-11 #667 @obrl-soil
It finally worked!!! I did have to remove that openssl variable though
