### 2025-03-12 e1ec3cf
try running test on CI with new httr2 options
