### 2025-03-12 c05eab6
don't skip quick-running tests on CI now that it seems to work
