+++ tests/testthat/test-read_historical_forecast_database.R
@@ -1,6 +1,5 @@
 test_that("read_historical_forecast() works", {
   skip_if_offline()
-  skip_on_ci()
   x <- read_historical_forecast()
   expect_named(
     x,
