+++ tests/testthat/test-read_historical_state_estimates.R
@@ -1,6 +1,5 @@
 test_that("read_historical_state_estimates works", {
   skip_if_offline()
-  skip_on_ci()
   x <- read_historical_state_estimates()
   expect_named(x, c("Variable", "Year", "State", "Industry", "Value", "RSE"))
   expect_s3_class(x, c("data.table", "data.frame"))
