### 2025-03-12 b9e7b17
Revert "don't skip quick-running tests on CI now that it seems to work"
This reverts commit c05eab6d75ba31a7cb406c39ec4cda37f159d897.
