+++ tests/testthat/test-read_aagis_regions.R
@@ -4,6 +4,7 @@ withr::local_envvar(R_USER_CACHE_DIR = file.path(tempdir(), "abares.cache.1"))
 # without caching ----
 test_that("read_aagis_regions doesn't cache", {
   skip_if_offline()
+  skip_on_ci()
   x <- read_aagis_regions(cache = FALSE)
   expect_s3_class(x, "sf")
   expect_false(file.exists(
@@ -23,6 +24,7 @@ test_that("read_aagis_regions skips downloading if still in tempdir()", {
 
 test_that("read_aagis_regions caches", {
   skip_if_offline()
+  skip_on_ci()
   x <- read_aagis_regions(cache = TRUE)
   expect_s3_class(x, "sf")
   expect_true(file.exists(
@@ -50,6 +52,7 @@ withr::local_envvar(R_USER_CACHE_DIR = file.path(tempdir(), "abares.cache.2"))
 
 test_that("read_aagis_regions does cache", {
   skip_if_offline()
+  skip_on_ci()
   x <- read_aagis_regions(cache = TRUE)
   expect_s3_class(x, "sf")
   expect_true(file.exists(
@@ -65,6 +68,7 @@ test_that("read_aagis_regions does cache", {
 
 test_that("read_aagis_regions skips downloading if still in tempdir()", {
   skip_if_offline()
+  skip_on_ci()
   x <- .check_existing_aagis(cache = TRUE)
   expect_s3_class(x, "sf")
 })
