+++ tests/testthat/test-read_historical_national_estimates.R
@@ -1,5 +1,6 @@
 test_that("read_historical_national_estimates works", {
   skip_if_offline()
+  skip_on_ci()
   x <- read_historical_national_estimates()
   expect_named(x, c("Variable", "Year", "Industry", "Value", "RSE"))
   expect_s3_class(x, c("data.table", "data.frame"))
