### 2025-03-12 a22f735
Revert "try running test on CI with new httr2 options"
This reverts commit e1ec3cf7d1ea4e54f8b99db0df3ab4e76280cabc.
