### 2025-03-17 #667 @obrl-soil
Hey - just trying to finalise my editor comments. Package check functions are working fine but I'm running into odd hiccups using the package interactively. For instance, running the example code from `read_soil_thickness_terra()`:

```
x <- get_soil_thickness(cache = TRUE) |>
+   read_soil_thickness_terra()
Error in list_files(file, ignore_missing, TRUE, verbosity) :
  File '' does not exist.
In addition: Warning message:
In file.rename(file.path(download_dir, "staiar9cl__05911a01eg_geo___/"),  :
  cannot rename file 'C:/Users/leobr/AppData/Local/R/cache/R/read.abares/staiar9cl__05911a01eg_geo___' to 'C:/Users/leobr/AppData/Local/R/cache/R/read.abares/soil_thickness_dir', reason 'Access is denied'
```

Not sure why this is happening. The files have downloaded and nothing is tagged as read-only in that folder.

Also, for

```
agfd <- get_agfd(cache = TRUE) |>
+   read_agfd_terra()
Error in `purrr::map()`:
ℹ In index: 1.
Caused by error:
! [rast] file does not exist: read.abares.agfd.nc.files
Run `rlang::last_trace()` to see where the error occurred.
Warning message:
read.abares.agfd.nc.files: No such file or directory (GDAL error 4)
```

The data is downloaded and present in the cache, but I don't think `get_agfd()` is listing the files correctly in its output object.

Can you replicate these issues?
