### 2025-03-18 #667 @adamhsparks
Yes, and I found where the issue is. Now I need to understand why.

Sorry about this, I thought it was all in good order.
