### 2025-03-19 1874728
refactors get_soil_thickness(), fixes bug reported by @obrl-soil
