### 2025-03-19 d7131c9
replace base functions with {fs}
