### 2025-03-19 32aaef3
update test objects and replace base fns with {fs} in tests
