+++ tests/testthat/test-read_abares_trade.R
@@ -1,5 +1,5 @@
 # sets up a custom cache environment in `tempdir()` just for testing
-withr::local_envvar(R_USER_CACHE_DIR = file.path(tempdir(), "abares.cache.1"))
+withr::local_envvar(R_USER_CACHE_DIR = fs::path(tempdir(), "abares.cache.1"))
 
 # without caching ----
 test_that("read_abares_trade doesn't cache", {
@@ -46,7 +46,7 @@ test_that("read_abares_trade doesn't cache", {
     )
   )
   expect_false(file.exists(
-    file.path(.find_user_cache(), "abares_trade_dir/abares_trade.gz")
+    fs::path(.find_user_cache(), "abares_trade_dir/abares_trade.gz")
   ))
 })
 
@@ -56,12 +56,12 @@ test_that("read_abares_trade caches", {
   skip_on_ci()
   read_abares_trade(cache = TRUE)
   expect_true(file.exists(
-    file.path(.find_user_cache(), "abares_trade_dir/abares_trade.gz")
+    fs::path(.find_user_cache(), "abares_trade_dir/abares_trade.gz")
   ))
 })
 
 # cleanup cache if rerunning tests in same R session so first test passes ----
 
-unlink(file.path(.find_user_cache(), "abares_trade_dir/abares_trade.gz"))
+unlink(fs::path(.find_user_cache(), "abares_trade_dir/abares_trade.gz"))
 
 withr::deferred_run()
