+++ R/read_agfd_tidync.R
@@ -25,6 +25,11 @@
 #' @export
 
 read_agfd_tidync <- function(files) {
+  if (!inherits(files, "read.abares.agfd.nc.files")) {
+    cli::cli_abort(
+      "{.var files} must be a list of class 'read.abares.agfd.nc.files'"
+    )
+  }
   tnc <- purrr::map(files, tidync::tidync)
   names(tnc) <- basename(files)
 }
