### 2025-03-19 e5669b1
delete slow_tests, move back to tests
