### 2025-03-19 2be6503
re-add slow tests to main tests
