+++ tests/testthat/test-read_agfd_stars.R
@@ -1,0 +1,93 @@
+test_that("read_agfd_stars() returns a stars object", {
+  skip_if_offline()
+  skip_on_ci()
+  x <- get_agfd() |>
+    read_agfd_stars()
+
+  expect_type(x, "list")
+  expect_s3_class(x[[1]], "stars")
+  expect_named(
+    x,
+    c(
+      "f2022.c1991.p2022.t2022.nc",
+      "f2022.c1992.p2022.t2022.nc",
+      "f2022.c1993.p2022.t2022.nc",
+      "f2022.c1994.p2022.t2022.nc",
+      "f2022.c1995.p2022.t2022.nc",
+      "f2022.c1996.p2022.t2022.nc",
+      "f2022.c1997.p2022.t2022.nc",
+      "f2022.c1998.p2022.t2022.nc",
+      "f2022.c1999.p2022.t2022.nc",
+      "f2022.c2000.p2022.t2022.nc",
+      "f2022.c2001.p2022.t2022.nc",
+      "f2022.c2002.p2022.t2022.nc",
+      "f2022.c2003.p2022.t2022.nc",
+      "f2022.c2004.p2022.t2022.nc",
+      "f2022.c2005.p2022.t2022.nc",
+      "f2022.c2006.p2022.t2022.nc",
+      "f2022.c2007.p2022.t2022.nc",
+      "f2022.c2008.p2022.t2022.nc",
+      "f2022.c2009.p2022.t2022.nc",
+      "f2022.c2010.p2022.t2022.nc",
+      "f2022.c2011.p2022.t2022.nc",
+      "f2022.c2012.p2022.t2022.nc",
+      "f2022.c2013.p2022.t2022.nc",
+      "f2022.c2014.p2022.t2022.nc",
+      "f2022.c2015.p2022.t2022.nc",
+      "f2022.c2016.p2022.t2022.nc",
+      "f2022.c2017.p2022.t2022.nc",
+      "f2022.c2018.p2022.t2022.nc",
+      "f2022.c2019.p2022.t2022.nc",
+      "f2022.c2020.p2022.t2022.nc",
+      "f2022.c2021.p2022.t2022.nc",
+      "f2022.c2022.p2022.t2022.nc",
+      "f2022.c2023.p2022.t2022.nc"
+    )
+  )
+  expect_named(
+    x[[1]],
+    c(
+      "farmno",
+      "R_total_hat_ha",
+      "C_total_hat_ha",
+      "FBP_fci_hat_ha",
+      "FBP_fbp_hat_ha",
+      "A_wheat_hat_ha",
+      "H_wheat_dot_hat",
+      "A_barley_hat_ha",
+      "H_barley_dot_hat",
+      "A_sorghum_hat_ha",
+      "H_sorghum_dot_hat",
+      "A_oilseeds_hat_ha",
+      "H_oilseeds_dot_hat",
+      "R_wheat_hat_ha",
+      "R_sorghum_hat_ha",
+      "R_oilseeds_hat_ha",
+      "R_barley_hat_ha",
+      "Q_wheat_hat_ha",
+      "Q_barley_hat_ha",
+      "Q_sorghum_hat_ha",
+      "Q_oilseeds_hat_ha",
+      "S_wheat_cl_hat_ha",
+      "S_sheep_cl_hat_ha",
+      "S_sheep_births_hat_ha",
+      "S_sheep_deaths_hat_ha",
+      "S_beef_cl_hat_ha",
+      "S_beef_births_hat_ha",
+      "S_beef_deaths_hat_ha",
+      "Q_beef_hat_ha",
+      "Q_sheep_hat_ha",
+      "Q_lamb_hat_ha",
[…]
