### 2025-03-19 d54f927
replace {fs} with base fns
no need to use an extra pkg
