+++ R/read_historical_national_estimates.R
@@ -19,7 +19,7 @@
 #' read_hist_nat_est()
 #'
 read_historical_national_estimates <- function() {
-  f <- fs::path(tempdir(), "fdp-beta-national-historical.csv")
+  f <- file.path(tempdir(), "fdp-beta-national-historical.csv")
 
   .retry_download(
     "https://www.agriculture.gov.au/sites/default/files/documents/fdp-national-historical.csv",
