+++ tests/testthat/test-inspect_cache.R
@@ -1,8 +1,8 @@
-withr::local_envvar(R_USER_CACHE_DIR = fs::path(tempdir()))
+withr::local_envvar(R_USER_CACHE_DIR = file.path(tempdir()))
 # check message with no files present ----
 
 test_that("inspect_cache() works w/ no files", {
-  fs::dir_create(fs::path(.find_user_cache()), recurse = TRUE)
+  dir.create(file.path(.find_user_cache()), recurse = TRUE)
   cli_out <- function() {
     return(cli::cli_inform(c(
       "There do not appear to be any files cached for {.pkg {{read.abares}}}."
@@ -15,11 +15,11 @@ test_that("inspect_cache() works w/ no files", {
 # Now create a file to check when files are present ----
 
 test_that("inspect_cache() works, recursive = FALSE", {
-  test_file <- fs::path(.find_user_cache(), "test.R")
-  fs::file_create(test_file)
+  test_file <- file.path(.find_user_cache(), "test.R")
+  file.create(test_file)
 
   f <- .find_user_cache()
-  f <- fs::dir_ls(f, full.names = TRUE)
+  f <- list.files(f, full.names = TRUE)
   expect_identical(
     inspect_cache() |>
       capture_output(),
@@ -28,11 +28,11 @@ test_that("inspect_cache() works, recursive = FALSE", {
 })
 
 test_that("inspect_cache() works, recurse = TRUE", {
-  test_file <- fs::path(.find_user_cache(), "test.R")
-  fs::file_create(test_file)
+  test_file <- file.path(.find_user_cache(), "test.R")
+  file.create(test_file)
 
   f <- .find_user_cache()
-  f <- fs::dir_ls(f, recurse = TRUE, full.names = TRUE)
+  f <- list.files(f, recurse = TRUE, full.names = TRUE)
   expect_identical(
     inspect_cache(recurse = TRUE) |>
       capture_output(),
