+++ tests/testthat/test-get_agfd.R
@@ -3,15 +3,15 @@ test_that("get_agfd, fixed = TRUE works", {
   skip_on_ci()
   x <- get_agfd()
 
-  agfd_nc_dir <- fs::path(
+  agfd_nc_dir <- file.path(
     .find_user_cache(),
     "historical_climate_prices_fixed"
   )
-  agfd_nc <- fs::dir_ls(agfd_nc_dir, full.names = TRUE)
+  agfd_nc <- list.files(agfd_nc_dir, full.names = TRUE)
 
   nc_files <- function(agfd_nc_dir) {
     cli::cli_h1("Locally Available ABARES AGFD NetCDF Files")
-    cli::cli_ul(basename(fs::dir_ls(agfd_nc_dir)))
+    cli::cli_ul(basename(list.files(agfd_nc_dir)))
     cli::cat_line()
   }
   print_out <- capture.output(nc_files)
@@ -22,9 +22,9 @@ test_that("get_agfd, fixed = TRUE works", {
     nc_files(agfd_nc_dir) |>
       capture_output()
   )
-  # cache fs::dir_created
+  # cache dir.created
   expect_true(dir.exists(
-    fs::path(.find_user_cache(), "historical_climate_prices_fixed")
+    file.path(.find_user_cache(), "historical_climate_prices_fixed")
   ))
 })
 
@@ -33,15 +33,15 @@ test_that("get_agfd, fixed = FALSE works", {
   skip_on_ci()
   x <- get_agfd(fixed_prices = FALSE)
 
-  agfd_nc_dir <- fs::path(
+  agfd_nc_dir <- file.path(
     .find_user_cache(),
     "historical_climate_prices_fixed"
   )
-  agfd_nc <- fs::dir_ls(agfd_nc_dir, full.names = TRUE)
+  agfd_nc <- list.files(agfd_nc_dir, full.names = TRUE)
 
   nc_files <- function(agfd_nc_dir) {
     cli::cli_h1("Locally Available ABARES AGFD NetCDF Files")
-    cli::cli_ul(basename(fs::dir_ls(agfd_nc_dir)))
+    cli::cli_ul(basename(list.files(agfd_nc_dir)))
     cli::cat_line()
   }
   print_out <- capture.output(nc_files)
@@ -52,9 +52,9 @@ test_that("get_agfd, fixed = FALSE works", {
     nc_files(agfd_nc_dir) |>
       capture_output()
   )
-  # cache fs::dir_created
+  # cache dir.created
   expect_true(dir.exists(
-    fs::path(.find_user_cache(), "historical_climate_and_prices")
+    file.path(.find_user_cache(), "historical_climate_and_prices")
   ))
 })
 
@@ -62,15 +62,15 @@ test_that("get_agfd, fixed = TRUE, no cache works", {
   skip_if_offline()
   x <- get_agfd(cache = FALSE)
 
-  agfd_nc_dir <- fs::path(
+  agfd_nc_dir <- file.path(
     .find_user_cache(),
     "historical_climate_prices_fixed"
   )
-  agfd_nc <- fs::dir_ls(agfd_nc_dir, full.names = TRUE)
+  agfd_nc <- list.files(agfd_nc_dir, full.names = TRUE)
 
   nc_files <- function(agfd_nc_dir) {
     cli::cli_h1("\nLocally Available ABARES AGFD NetCDF Files\n")
-    cli::cli_ul(basename(fs::dir_ls(agfd_nc_dir)))
+    cli::cli_ul(basename(list.files(agfd_nc_dir)))
     cat("\n")
   }
   print_out <- capture.output(nc_files)
@@ -81,9 +81,9 @@ test_that("get_agfd, fixed = TRUE, no cache works", {
     nc_files(agfd_nc_dir) |>
[…]

[…]
