### 2025-03-20 14e5363
get_agfd() passes tests
