### 2025-03-20 d1df646
clear_cache() passes tests
