+++ tests/testthat/test-inspect_cache.R
@@ -2,7 +2,7 @@ withr::local_envvar(R_USER_CACHE_DIR = file.path(tempdir()))
 # check message with no files present ----
 
 test_that("inspect_cache() works w/ no files", {
-  dir.create(file.path(.find_user_cache()), recurse = TRUE)
+  dir.create(file.path(.find_user_cache()), recursive = TRUE)
   cli_out <- function() {
     return(cli::cli_inform(c(
       "There do not appear to be any files cached for {.pkg {{read.abares}}}."
@@ -27,14 +27,14 @@ test_that("inspect_cache() works, recursive = FALSE", {
   )
 })
 
-test_that("inspect_cache() works, recurse = TRUE", {
+test_that("inspect_cache() works, recursive = TRUE", {
   test_file <- file.path(.find_user_cache(), "test.R")
   file.create(test_file)
 
   f <- .find_user_cache()
-  f <- list.files(f, recurse = TRUE, full.names = TRUE)
+  f <- list.files(f, recursive = TRUE, full.names = TRUE)
   expect_identical(
-    inspect_cache(recurse = TRUE) |>
+    inspect_cache(recursive = TRUE) |>
       capture_output(),
     f |> capture_output()
   )
