+++ tests/testthat/test-read_aagis_regions.R
@@ -1,3 +1,25 @@
+# with caching ----
+# sets up a custom cache environment in `tempdir()` just for testing
+withr::local_envvar(R_USER_CACHE_DIR = file.path(tempdir(), "abares.cache.20"))
+
+test_that("read_aagis_regions does cache", {
+  skip_if_offline()
+  skip_on_ci()
+  x <- read_aagis_regions(cache = TRUE)
+  expect_s3_class(x, "sf")
+  expect_true(file.exists(
+    file.path(.find_user_cache(), "aagis_regions_dir/aagis.gpkg")
+  ))
+  expect_false(file.exists(
+    file.path(
+      .find_user_cache(),
+      "aagis_regions_dir/aagis_asgs16v1_g5a.shp"
+    )
+  ))
+  file.remove(file.path(.find_user_cache(), "aagis_regions_dir/aagis.gpkg"))
+})
+
+withr::deferred_run()
 # sets up a custom cache environment in `tempdir()` just for testing
 withr::local_envvar(R_USER_CACHE_DIR = file.path(tempdir(), "abares.cache.1"))
 
@@ -17,65 +39,6 @@ test_that("read_aagis_regions skips downloading if still in tempdir()", {
   skip_if_offline()
   skip_on_ci()
   x <- .check_existing_aagis(cache = FALSE)
-  expect_s3_class(x, "sf")
-})
-withr::deferred_run()
-
-# with caching ----
-# sets up a custom cache environment in `tempdir()` just for testing
-withr::local_envvar(R_USER_CACHE_DIR = file.path(tempdir(), "abares.cache.1"))
-
-unlink(file.path(.find_user_cache(), "aagis_regions_dir/aagis.gpkg"))
-
-test_that("read_aagis_regions caches", {
-  skip_if_offline()
-  skip_on_ci()
-  x <- read_aagis_regions(cache = TRUE)
-  expect_s3_class(x, "sf")
-  expect_true(file.exists(
-    file.path(.find_user_cache(), "aagis_regions_dir/aagis.gpkg")
-  ))
-  expect_false(file.exists(
-    file.path(.find_user_cache(), "aagis_regions_dir/aagis_zip")
-  ))
-  expect_false(file.exists(file.path(
-    .find_user_cache(),
-    "aagis_asgs16v1_g5a.*"
-  )))
-})
-
-test_that("read_aagis_regions skips downloading if cache is available", {
-  skip_if_offline()
-  skip_on_ci()
-  x <- .check_existing_aagis(cache = TRUE)
-  expect_s3_class(x, "sf")
-})
-withr::deferred_run()
-
-# sets up a custom cache environment in `tempdir()` just for testing
-withr::local_envvar(R_USER_CACHE_DIR = file.path(tempdir(), "abares.cache.2"))
-
-
-test_that("read_aagis_regions does cache", {
-  skip_if_offline()
-  skip_on_ci()
-  x <- read_aagis_regions(cache = TRUE)
-  expect_s3_class(x, "sf")
-  expect_true(file.exists(
-    file.path(.find_user_cache(), "aagis_regions_dir/aagis.gpkg")
-  ))
-  expect_false(file.exists(
-    file.path(
-      .find_user_cache(),
-      "aagis_regions_dir/aagis_asgs16v1_g5a.shp"
-    )
-  ))
-})
-
-test_that("read_aagis_regions skips downloading if still in tempdir()", {
-  skip_if_offline()
-  skip_on_ci()
-  x <- .check_existing_aagis(cache = TRUE)
   expect_s3_class(x, "sf")
 })
 
