### 2025-03-21 1ddd3bd
fixes read_estimates_by_performance_category() writing to user-space
