### 2025-03-21 8853d69
check for proper inputs and add test
