+++ R/read_agfd_dt.R
@@ -25,6 +25,11 @@
 #' @export
 
 read_agfd_dt <- function(files) {
+  if (!inherits(files, "read.abares.agfd.nc.files")) {
+    cli::cli_abort(
+      "{.var files} must be a list of class 'read.abares.agfd.nc.files'"
+    )
+  }
   tnc_list <- lapply(files, tidync::tidync)
   names(tnc_list) <- basename(files)
   dt <- data.table::rbindlist(
