+++ R/read_agfd_terra.R
@@ -25,6 +25,11 @@
 #' @export
 
 read_agfd_terra <- function(files) {
+  if (!inherits(files, "read.abares.agfd.nc.files")) {
+    cli::cli_abort(
+      "{.var files} must be a list of class 'read.abares.agfd.nc.files'"
+    )
+  }
   r <- purrr::map(.x = files, .f = terra::rast)
   names(r) <- basename(files)
   return(r)
