+++ tests/testthat/test-read_agfd_stars.R
@@ -91,3 +91,7 @@ test_that("read_agfd_stars() returns a stars object", {
     )
   )
 })
+
+test_that("read_agfd_stars() fails if the input is not a proper object", {
+  expect_error(read_agfd_stars(list(list.files(tempdir()))))
+})
