+++ tests/testthat/test-read_agfd_tidync.R
@@ -45,3 +45,7 @@ test_that("read_agfd_tidync() returns a tidync object", {
     )
   )
 })
+
+test_that("read_agfd_tidync() fails if the input is not a proper object", {
+  expect_error(read_agfd_tidync(list(list.files(tempdir()))))
+})
