### 2025-03-21 #667 @adamhsparks
Ok! @obrl-soil, I'm very sorry about that. I've now inspected, detected, debugged and refactored and it's now passing all tests and has improved code-coverage, 98%, along with better guard rails.

I _hope_ it's good to go now. This has been a problematic package, to say the least and it shouldn't be that complicated!
