### 2025-03-21 #667 @adamhsparks
I did move the slow tests back to "tests/testthat" because it was just too much overhead and I ran into several issues I'd missed during regular checks that were picked up with these tests. It only takes a few minutes or less to run normal tests after the first run, so I find it acceptable.
