### 2025-03-24 #667 @obrl-soil
Getting there I think! Most of the interactive examples seem fine now.

I am still getting some odd behaviour when running `devtools::test()` - multiple `abares.cache` folders are being created (`abares.cache.1`, `abares.cache.2`, and `abares.cache.20` in the same tmp folder) and I don't recall that happening previously. Haven't had any problems _downloading_ the data at least...

Interactively, `get_soil_thickness()` has started erroring again with a similar issue:
```
 <- get_soil_thickness()
Error in list_files(file, ignore_missing, TRUE, verbosity) :
  File '' does not exist.
In addition: Warning messages:
1: In dir.create(dirname(soil_thick_adf_file), recursive = TRUE) :
  'C:\Users\leobr\AppData\Local\R\cache\R\read.abares\soil_thickness_dir' already exists
2: In file.rename(file.path(download_dir, "staiar9cl__05911a01eg_geo___/"),  :
  cannot rename file 'C:/Users/leobr/AppData/Local/R/cache/R/read.abares/staiar9cl__05911a01eg_geo___' to 'C:/Users/leobr/AppData/Local/R/cache/R/read.abares/soil_thickness_dir', reason 'Access is denied'
```

I think your `dir.create()` call in `get_soil_thickness.R.104` maybe needs to be conditional?
