### 2025-03-24 #667 @adamhsparks
Thanks, @mpadge. I'd looked at using {fs} and opted not to for fewer dependencies.

@obrl-soil, I'll get back to you later, I'll follow Mark's advice and just use {fs}, but that will take some refactoring to do. I did it once and got mostly there but had some issues and so I rolled back but seems I should have stuck with it.
