+++ tests/testthat/test-read_agfd_dt.R
@@ -106,5 +106,5 @@ test_that("read_agfd_dt returns a data.table object", {
   )
 })
 test_that("read_agfd_dt() fails if the input is not a proper object", {
-  expect_error(read_agfd_dt(list(list.files(tempdir()))))
+  expect_error(read_agfd_dt(list(fs::dir_ls(tempdir()))))
 })
