+++ R/read_soil_thickness_terra.R
@@ -9,14 +9,14 @@
 #' @source <https://anrdl-integration-web-catalog-saxfirxkxt.s3-ap-southeast-2.amazonaws.com/warehouse/staiar9cl__059/staiar9cl__05911a01eg_geo___.zip>
 
 #'
-#' @returns a [terra::rast] object of the 'Soil Thickness for Australian Areas of
-#'  Intensive Agriculture of Layer 1'.
+#' @returns a [terra::rast] object of the 'Soil Thickness for Australian Areas
+#'  of Intensive Agriculture of Layer 1'.
 #'
 #' @examplesIf interactive()
 #' x <- get_soil_thickness(cache = TRUE) |>
 #'   read_soil_thickness_terra()
 #'
-#' # terra::plot() is reexported for convience
+#' # terra::plot() is reexported for convenience
 #' plot(x)
 #'
 #' @family soil_thickness
