### 2025-03-25 b346285
complete rewrite of downloading and file cache handling
