### 2025-03-25 bfde6c5
use internal .check_class()
