### 2025-03-25 a857a8a
ensure that soil grid is returned on download
