### 2025-03-26 0db7c12
fix clear_cache() function bugs
