+++ man/dot-create_soil_thickness_list.Rd
@@ -7,11 +7,12 @@
 .create_soil_thickness_list(soil_dir)
 }
 \arguments{
-\item{dir}{File where files have been downloaded.}
+\item{dir}{File where files have been stored.}
 }
 \value{
 A \code{read.abares.soil.thickness} object, which is a named \code{list} with
-the fs::path_file of the resulting Esri Grid file and text file of metadata.
+the \link[fs:path_file]{fs::path_file} of the resulting Esri Grid file and text file of
+metadata.
 }
 \description{
 Create a object of read.abares.soil.thickness.files
