### 2025-03-26 5a5a1b0
fix bugs in get_agfd()
