### 2025-03-26 4430a3f
I hope this fixes the bugs in getting soil thickness data!
