### 2025-03-26 b888cb1
fix tests for inspect_cache()
