+++ tests/testthat/test-inspect_cache.R
@@ -1,8 +1,8 @@
-withr::local_envvar(R_USER_CACHE_DIR = fs::path_file(tempdir()))
+withr::local_envvar(R_USER_CACHE_DIR = tempdir())
 # check message with no files present ----
 
 test_that("inspect_cache() works w/ no files", {
-  fs::dir_create(fs::path_file(.find_user_cache()), recursive = TRUE)
+  fs::dir_create(.find_user_cache())
   cli_out <- function() {
     return(cli::cli_inform(c(
       "There do not appear to be any files cached for {.pkg {{read.abares}}}."
@@ -15,7 +15,7 @@ test_that("inspect_cache() works w/ no files", {
 # Now create a file to check when files are present ----
 
 test_that("inspect_cache() works, recursive = FALSE", {
-  test_file <- fs::path_file(.find_user_cache(), "test.R")
+  test_file <- fs::path(.find_user_cache(), "test.R")
   file.create(test_file)
 
   f <- .find_user_cache()
@@ -28,13 +28,13 @@ test_that("inspect_cache() works, recursive = FALSE", {
 })
 
 test_that("inspect_cache() works, recursive = TRUE", {
-  test_file <- fs::path_file(.find_user_cache(), "test.R")
+  test_file <- fs::path(.find_user_cache(), "test.R")
   file.create(test_file)
 
   f <- .find_user_cache()
-  f <- fs::dir_ls(f, recursive = TRUE, full.names = TRUE)
+  f <- fs::dir_ls(f, recurse = TRUE, full.names = TRUE)
   expect_identical(
-    inspect_cache(recursive = TRUE) |>
+    inspect_cache(recurse = TRUE) |>
       capture_output(),
     f |> capture_output()
   )
