### 2025-03-26 f31ed65
ensure cache is cleared for following tests
