+++ pkgdown/_pkgdown.yml
@@ -4,59 +4,58 @@ template:
   params:
     bootswatch: journal
 reference:
-- title: AAGIS Regions
-  desc: |
-    Fetch and read the AAGIS regions
-  contents: read_aagis_regions
-- title: AGFD
-- subtitle: Fetch
-  desc: |
-    Fetch the Australian Gridded Farm Data
-  contents: get_agfd
-- subtitle: Read
-  desc: |
-    Read the Australian Gridded Farm Data
-  contents:
-  - read_agfd_dt
-  - read_agfd_stars
-  - read_agfd_terra
-  - read_agfd_tidync
-- subtitle: Meta
-  desc: |
-    Information about the AGFD NetCDF files and contents
-  contents:
-  - print_agfd_nc_file_format
-- title: Estimates
-  desc: |
-    Fetch and read the ABARES Estimates data
-  contents:
-  - read_estimates_by_performance_category
-  - read_estimates_by_size
-  - read_historical_national_estimates
-  - read_historical_regional_estimates
-  - read_historical_state_estimates
-- title: Historical agricultural forecast database
-  desc: |
-    Fetch and read the historical agricultural forecast database
-  contents: read_historical_forecast_database
-- title: Cache
-  desc: |
-    Work with your local file cache
-  contents:
-  - clear_cache
-  - inspect_cache
-- title: Soil Thickness
-  desc: |
-    Soil Thickness for Australian areas of intensive agriculture of Layer 1 (A Horizon - top-soil) (derived from soil mapping)
-  contents:
-  - get_soil_thickness
-  - read_soil_thickness_stars
-  - read_soil_thickness_terra
-  - print_soil_thickness_metadata
-- title: Trade
-  desc: |
-    Fetch and read ABARES' Australian agricultural export data
-  contents:
-  - read_abares_trade
-  - read_abares_trade_regions
-
+  - title: AAGIS Regions
+    desc: |
+      Fetch and read the Abares agricultural and Grazing Industries Survey (AAGIS) regions
+    contents: read_aagis_regions
+  - title: AGFD
+  - subtitle: Fetch
+    desc: |
+      Fetch the Australian Gridded Farm Data (AGFD) NetCDF files
+    contents: get_agfd
+  - subtitle: Read
+    desc: |
+      Read the Australian Gridded Farm Data (AGFD) NetCDF files
+    contents:
+      - read_agfd_dt
+      - read_agfd_stars
+      - read_agfd_terra
+      - read_agfd_tidync
+  - subtitle: Meta
+    desc: |
+      Information about the AGFD NetCDF files and contents
+    contents:
+      - print_agfd_nc_file_format
+  - title: Estimates
+    desc: |
+      Fetch and read the ABARES Estimates data
+    contents:
+      - read_estimates_by_performance_category
+      - read_estimates_by_size
+      - read_historical_national_estimates
+      - read_historical_regional_estimates
+      - read_historical_state_estimates
+  - title: Historical agricultural forecast database
+    desc: |
+      Fetch and read the historical agricultural forecast database
+    contents: read_historical_forecast_database
+  - title: Cache
+    desc: |
+      Work with your local file cache
+    contents:
+      - clear_cache
+      - inspect_cache
+  - title: Soil Thickness
+    desc: |
+      Soil Thickness for Australian areas of intensive agriculture of Layer 1 (A Horizon - top-soil) (derived from soil mapping)
+    contents:
+      - get_soil_thickness
+      - read_soil_thickness_stars
+      - read_soil_thickness_terra
+      - print_soil_thickness_metadata
+  - title: Trade
+    desc: |
+      Fetch and read ABARES' Australian agricultural export data
+    contents:
+      - read_abares_trade
+      - read_abares_trade_regions
