### 2025-03-26 4b798e0
streamline caching and align with soil thickness functionality
