+++ R/read_aagis_regions.R
@@ -2,13 +2,13 @@
 #'
 #' Download, cache and import the Australian Agricultural and Grazing
 #'  Industries Survey (\acronym{AAGIS} regions geospatial shapefile. Upon
-#'  import, the geometries  are automatically corrected to fix invalid
+#'  import, the geometries are automatically corrected to fix invalid
 #'  geometries that are present in the original shapefile.
 #'
 #' @param cache Cache the \acronym{AAGIS} regions' geospatial file after
 #' downloading using `tools::R_user_dir("read.abares", "cache")` to identify the
 #' proper directory for storing user data in a cache for this package. Defaults
-#' to `TRUE`, caching the files locally as a Geopackage. If `FALSE`, this
+#' to `TRUE`, caching the files locally as a GeoPackage. If `FALSE`, this
 #' function uses `tempdir()` and the files are deleted upon closing of the
 #' active \R session.
 #'
@@ -19,7 +19,7 @@
 #'
 #' @returns An \CRANpkg{sf} object of the \acronym{AAGIS} regions.
 #'
-#' @family AGFD
+#' @family AAGIS
 #'
 #' @references <https://www.agriculture.gov.au/abares/research-topics/surveys/farm-definitions-methods#regions>
 #' @source <https://www.agriculture.gov.au/sites/default/files/documents/aagis_asgs16v1_g5a.shp_.zip>
@@ -27,46 +27,17 @@
 #' @export
 
 read_aagis_regions <- function(cache = TRUE) {
-  aagis <- .check_existing_aagis(cache)
-  if (is.null(aagis)) {
-    aagis <- .download_aagis_shp(cache)
+  aagis_regions_cache <- fs::path(.find_user_cache(), "aagis_regions_dir")
+  if (fs::dir_exists(aagis_regions_cache)) {
+    return(sf::st_read(
+      fs::path(
+        aagis_regions_cache,
+        "aagis.gpkg"
+      ),
+      quiet = TRUE
+    ))
   } else {
-    return(aagis)
-  }
-}
-
-#' Check for a pre-existing file before downloading
-#'
-#' Checks the user cache first, then `tempdir()` for the files before
-#' returning a `NULL` value. If `cache == TRUE` and the file is not in the user
-#' cache, but is in `tempdir()`, it is saved to the cache before being returned
-#' in the current session.
-#'
-#' @returns An \CRANpkg{sf} object of AAGIS regions.
-#' @dev
-#' @autoglobal
-
-.check_existing_aagis <- function(cache) {
-  # TODO: streamline this check function with the way that get_soil_thickness works
-  aagis_gpkg <- fs::path(
-    .find_user_cache(),
-    "aagis_regions_dir/aagis.gpkg"
-  )
-  tmp_shp <- fs::path(tempdir(), "aagis_asgs16v1_g5a.shp")
-
-  if (fs::file_exists(aagis_gpkg)) {
-    return(sf::st_read(aagis_gpkg, quiet = TRUE))
-  } else if (fs::file_exists(tmp_shp)) {
-    aagis_sf <- sf::st_read(tmp_shp, quiet = TRUE)
-    # From checking the unzipped file, some geometries are invalid, this corrects
-    aagis_sf <- sf::st_make_valid(aagis_sf)
-    if (cache) {
-      fs::dir_create(fs::path_dir(aagis_gpkg), recurse = TRUE)
-      sf::st_write(obj = aagis_sf, dsn = aagis_gpkg, quiet = TRUE)
-    }
-    return(aagis_sf)
[…]
