### 2025-03-26 452eef7
fixes bugs with read_aagis_regions()
