### 2025-03-26 fe83550
simplify tests and fix fs path issues
