### 2025-03-26 94752e7
return the class value that is required
