+++ R/read_agfd_tidync.R
@@ -25,7 +25,7 @@
 #' @export
 
 read_agfd_tidync <- function(files) {
-  .check_class(x = files, class = "read.abares.nc.files")
+  .check_class(x = files, class = "read.abares.agfd.nc.files")
   tnc <- purrr::map(files, tidync::tidync)
   names(tnc) <- basename(files)
   return(tnc)
