### 2025-03-26 #667 @adamhsparks
@obrl-soil, thank you so much for your patience with this. This has been a problematic package from the start but I think thanks to you, @maurolepore and @mpadge, it's improved markedly even before the full review.

I've now used {fs} for better cross-platform support as suggested by @mpadge. I also managed to find some of the bugs and simplify the code where it was overly complex and causing issues.

I _hope_ it's ready for you to perform editor checks on now. It's passing tests for me locally and code coverage is ~98% now.
