### 2025-03-26 #667 @obrl-soil
So close! (sorry D: ). The cache issue is solved, the only two errors I'm still seeing in `devtools::test()` are:

```
 library(devtools)
Loading required package: usethis
> test()
ℹ Testing read.abares
Starting 2 test processes
✔ | F W  S  OK | Context
✔ |          2 | clear_cache
✔ |          2 | check_class
✔ |         11 | get_soil_thickness [1.9s]
✔ |          3 | inspect_cache
✔ |          6 | read_aagis_regions [3.4s]
✔ |          5 | read_abares_trade [197.4s]
✔ |          3 | read_abares_trade_regions
✔ |          4 | read_agfd_dt [50.3s]
✔ |          5 | read_agfd_stars [69.9s]
✔ |         12 | get_agfd [335.4s]
✔ |          5 | read_agfd_terra [16.4s]
✔ |          3 | read_estimates_by_performance_category
✔ |          4 | read_agfd_tidync [6.3s]
✔ |          3 | read_estimates_by_size [1.2s]
✔ |          3 | read_historical_national_estimates
✔ |          3 | read_historical_forecast_database [2.8s]
✔ |          3 | read_historical_regional_estimates [2.3s]
✖ | 1        0 | read_soil_thickness_stars
─────────────────────────────────────────────────────────────────────────────────────────────────────────
Error (test-read_soil_thickness_stars.R:4:3): read_soil_thickness_stars returns a stars object
Error in `list_files(file, ignore_missing, TRUE, verbosity)`: File '' does not exist.
Backtrace:
     ▆
  1. ├─read.abares::read_soil_thickness_stars(get_soil_thickness(cache = TRUE)) at test-read_soil_thickness_stars.R:4:3
  2. │ └─read.abares:::.check_class(x = files, class = "read.abares.soil.thickness.files") at read.abares/R/read_soil_thickness_stars.R:24:3
  3. │   └─base::inherits(x, class) at read.abares/R/internal_functions.R:20:3
  4. └─read.abares::get_soil_thickness(cache = TRUE)
  5.   └─read.abares:::.create_soil_thickness_list(soil_thickness_cache) at read.abares/R/get_soil_thickness.R:38:5
  6.     └─readtext::readtext(fs::path(soil_dir, "ANZCW1202000149.txt")) at read.abares/R/get_soil_thickness.R:54:3
  7.       └─readtext:::list_files(...)
  8.         ├─base::unlist(...)
  9.         └─base::lapply(...)
 10.           └─readtext (local) FUN(X[[i]], ...)
 11.             └─readtext:::list_file(...)
 12.               └─readtext:::list_files(file, ignore_missing, TRUE, verbosity)
─────────────────────────────────────────────────────────────────────────────────────────────────────────
✖ | 1        0 | read_soil_thickness_terra
─────────────────────────────────────────────────────────────────────────────────────────────────────────
Error (test-read_soil_thickness_terra.R:4:3): read_soil_thickness_stars returns a terra object
Error in `list_files(file, ignore_missing, TRUE, verbosity)`: File '' does not exist.
Backtrace:
     ▆
  1. ├─read.abares::read_soil_thickness_terra(get_soil_thickness(cache = TRUE)) at test-read_soil_thickness_terra.R:4:3
  2. │ └─read.abares:::.check_class(x = files, class = "read.abares.soil.thickness.files") at read.abares/R/read_soil_thickness_terra.R:27:3
  3. │   └─base::inherits(x, class) at read.abares/R/internal_functions.R:20:3
  4. └─read.abares::get_soil_thickness(cache = TRUE)
  5.   └─read.abares:::.create_soil_thickness_list(soil_thickness_cache) at read.abares/R/get_soil_thickness.R:38:5
  6.     └─readtext::readtext(fs::path(soil_dir, "ANZCW1202000149.txt")) at read.abares/R/get_soil_thickness.R:54:3
  7.       └─readtext:::list_files(...)
  8.         ├─base::unlist(...)
  9.         └─base::lapply(...)
 10.           └─readtext (local) FUN(X[[i]], ...)
 11.             └─readtext:::list_file(...)
 12.               └─readtext:::list_files(file, ignore_missing, TRUE, verbosity)
─────────────────────────────────────────────────────────────────────────────────────────────────────────
✔ |          2 | retry_download [1.2s]
✔ |          3 | read_historical_state_estimates [2.8s]

══ Results ══════════════════════════════════════════════════════════════════════════════════════════════
Duration: 352.7 s

```
