### 2025-03-26 8122110
trying to simplify/fix get_agfd() caching
