### 2025-03-26 #667 @obrl-soil
Curiouser and curiouser... I woke up with a feeling that the cache might have leftover junk from the previous testing work in it so:
- I manually deleted the cache, closed out of any RStudio projects, loaded the library, ran `get_soil_thickness()`  and `x <- get_agfd(cache = TRUE) |>  read_agfd_dt()` and had no problems.
- I went back into the package project cloned from GH, ran `devtools::test()`, and got new and different errors. All of them read like errors in the test suite, not the source code, but I can't see any obvious culprit in your recent commits.

<summary> <b>devtools::test() output:</b> </summary>

```
> test()
ℹ Testing read.abares
Starting 2 test processes
✔ | F W  S  OK | Context
✔ |          2 | clear_cache
✔ |          2 | check_class
✔ |         11 | get_soil_thickness [2.2s]
✔ |          3 | inspect_cache
✔ |          6 | read_aagis_regions [4.9s]
✔ |          5 | read_abares_trade [192.2s]
✔ |          3 | read_abares_trade_regions
✖ | 2        2 | read_agfd_dt
─────────────────────────────────────────────────────────────────────────────────────────────────────────
Failure (test-read_agfd_dt.R:8:3): read_agfd_dt returns a data.table object
Names of `x` ('lat', 'lon') don't match 'id', 'farmno', 'R_total_hat_ha', 'C_total_hat_ha', 'FBP_fci_hat_ha', 'FBP_fbp_hat_ha', 'A_wheat_hat_ha', 'H_wheat_dot_hat', 'A_barley_hat_ha', 'H_barley_dot_hat', 'A_sorghum_hat_ha', 'H_sorghum_dot_hat', 'A_oilseeds_hat_ha', 'H_oilseeds_dot_hat', 'R_wheat_hat_ha', 'R_sorghum_hat_ha', 'R_oilseeds_hat_ha', 'R_barley_hat_ha', 'Q_wheat_hat_ha', 'Q_barley_hat_ha', 'Q_sorghum_hat_ha', 'Q_oilseeds_hat_ha', 'S_wheat_cl_hat_ha', 'S_sheep_cl_hat_ha', 'S_sheep_births_hat_ha', 'S_sheep_deaths_hat_ha', 'S_beef_cl_hat_ha', 'S_beef_births_hat_ha', 'S_beef_deaths_hat_ha', 'Q_beef_hat_ha', 'Q_sheep_hat_ha', 'Q_lamb_hat_ha', 'R_beef_hat_ha', 'R_sheep_hat_ha', 'R_lamb_hat_ha', 'C_fodder_hat_ha', 'C_fert_hat_ha', 'C_fuel_hat_ha', 'C_chem_hat_ha', 'A_total_cropped_ha', 'FBP_pfe_hat_ha', 'farmland_per_cell', 'lon', 'lat'

Failure (test-read_agfd_dt.R:58:3): read_agfd_dt returns a data.table object
lapply(x, typeof) (`actual`) not identical to list(...) (`expected`).

`actual` is length 2
`expected` is length 44

    names(actual) | names(expected)
[1] "lat"         - "id"               [1]
[2] "lon"         - "farmno"           [2]
                  - "R_total_hat_ha"   [3]
                  - "C_total_hat_ha"   [4]
                  - "FBP_fci_hat_ha"   [5]
                  - "FBP_fbp_hat_ha"   [6]
                  - "A_wheat_hat_ha"   [7]
                  - "H_wheat_dot_hat"  [8]
                  - "A_barley_hat_ha"  [9]
                  - "H_barley_dot_hat" [10]
... ...             ...                and 34 more ...

`actual$id` is absent
`expected$id` is a character vector ('character')

`actual$farmno` is absent
`expected$farmno` is a character vector ('double')

`actual$R_total_hat_ha` is absent
`expected$R_total_hat_ha` is a character vector ('double')

`actual$C_total_hat_ha` is absent
`expected$C_total_hat_ha` is a character vector ('double')

`actual$FBP_fci_hat_ha` is absent
`expected$FBP_fci_hat_ha` is a character vector ('double')

`actual$FBP_fbp_hat_ha` is absent
`expected$FBP_fbp_hat_ha` is a character vector ('double')

`actual$A_wheat_hat_ha` is absent
`expected$A_wheat_hat_ha` is a character vector ('double')

`actual$H_wheat_dot_hat` is absent
`expected$H_wheat_dot_hat` is a character vector ('double')

`actual$A_barley_hat_ha` is absent
`expected$A_barley_hat_ha` is a character vector ('double')

`actual$H_barley_dot_hat` is absent
`expected$H_barley_dot_hat` is a character vector ('double')

`actual$A_sorghum_hat_ha` is absent
`expected$A_sorghum_hat_ha` is a character vector ('double')

`actual$H_sorghum_dot_hat` is absent
`expected$H_sorghum_dot_hat` is a character vector ('double')

`actual$A_oilseeds_hat_ha` is absent
`expected$A_oilseeds_hat_ha` is a character vector ('double')

`actual$H_oilseeds_dot_hat` is absent
`expected$H_oilseeds_dot_hat` is a character vector ('double')

`actual$R_wheat_hat_ha` is absent
`expected$R_wheat_hat_ha` is a character vector ('double')

`actual$R_sorghum_hat_ha` is absent
`expected$R_sorghum_hat_ha` is a character vector ('double')

`actual$R_oilseeds_hat_ha` is absent
`expected$R_oilseeds_hat_ha` is a character vector ('double')

`actual$R_barley_hat_ha` is absent
`expected$R_barley_hat_ha` is a character vector ('double')

`actual$Q_wheat_hat_ha` is absent
`expected$Q_wheat_hat_ha` is a character vector ('double')

`actual$Q_barley_hat_ha` is absent
`expected$Q_barley_hat_ha` is a character vector ('double')

`actual$Q_sorghum_hat_ha` is absent
`expected$Q_sorghum_hat_ha` is a character vector ('double')

`actual$Q_oilseeds_hat_ha` is absent
`expected$Q_oilseeds_hat_ha` is a character vector ('double')

`actual$S_wheat_cl_hat_ha` is absent
`expected$S_wheat_cl_hat_ha` is a character vector ('double')

`actual$S_sheep_cl_hat_ha` is absent
`expected$S_sheep_cl_hat_ha` is a character vector ('double')

`actual$S_sheep_births_hat_ha` is absent
`expected$S_sheep_births_hat_ha` is a character vector ('double')

`actual$S_sheep_deaths_hat_ha` is absent
`expected$S_sheep_deaths_hat_ha` is a character vector ('double')

`actual$S_beef_cl_hat_ha` is absent
`expected$S_beef_cl_hat_ha` is a character vector ('double')

`actual$S_beef_births_hat_ha` is absent
`expected$S_beef_births_hat_ha` is a character vector ('double')

`actual$S_beef_deaths_hat_ha` is absent
`expected$S_beef_deaths_hat_ha` is a character vector ('double')

`actual$Q_beef_hat_ha` is absent
`expected$Q_beef_hat_ha` is a character vector ('double')

`actual$Q_sheep_hat_ha` is absent
`expected$Q_sheep_hat_ha` is a character vector ('double')

`actual$Q_lamb_hat_ha` is absent
`expected$Q_lamb_hat_ha` is a character vector ('double')

`actual$R_beef_hat_ha` is absent
`expected$R_beef_hat_ha` is a character vector ('double')

`actual$R_sheep_hat_ha` is absent
`expected$R_sheep_hat_ha` is a character vector ('double')

`actual$R_lamb_hat_ha` is absent
`expected$R_lamb_hat_ha` is a character vector ('double')

`actual$C_fodder_hat_ha` is absent
`expected$C_fodder_hat_ha` is a character vector ('double')

`actual$C_fert_hat_ha` is absent
`expected$C_fert_hat_ha` is a character vector ('double')

`actual$C_fuel_hat_ha` is absent
`expected$C_fuel_hat_ha` is a character vector ('double')

`actual$C_chem_hat_ha` is absent
`expected$C_chem_hat_ha` is a character vector ('double')

`actual$A_total_cropped_ha` is absent
`expected$A_total_cropped_ha` is a character vector ('double')

`actual$FBP_pfe_hat_ha` is absent
`expected$FBP_pfe_hat_ha` is a character vector ('double')

`actual$farmland_per_cell` is absent
`expected$farmland_per_cell` is a character vector ('double')
─────────────────────────────────────────────────────────────────────────────────────────────────────────
✖ | 1        1 | read_agfd_stars
─────────────────────────────────────────────────────────────────────────────────────────────────────────
Error (test-read_agfd_stars.R:4:3): read_agfd_stars() returns a stars object
Error in `if (nchar(x) < 1) stop("NetCDF source cannot be empty string")`: missing value where TRUE/FALSE needed
Backtrace:
    ▆
 1. └─read.abares::read_agfd_stars(get_agfd()) at test-read_agfd_stars.R:4:3
 2.   └─stars::read_ncdf(files[1], var = var) at read.abares/R/read_agfd_stars.R:78:3
 3.     ├─stars:::.fix_meta(ncmeta::nc_meta(x))
 4.     ├─ncmeta::nc_meta(x)
 5.     └─ncmeta:::nc_meta.character(x)
─────────────────────────────────────────────────────────────────────────────────────────────────────────
✖ | 1        2 | read_agfd_terra
─────────────────────────────────────────────────────────────────────────────────────────────────────────
Error (test-read_agfd_terra.R:8:3): read_agfd_terra() returns a terra object
<subscriptOutOfBoundsError/error/condition>
Error in `x[[1]]`: subscript out of bounds
Backtrace:
    ▆
 1. └─testthat::expect_s4_class(x[[1]], "SpatRaster") at test-read_agfd_terra.R:8:3
 2.   └─testthat::quasi_label(enquo(object), arg = "object")
 3.     └─rlang::eval_bare(expr, quo_get_env(quo))
─────────────────────────────────────────────────────────────────────────────────────────────────────────
✖ | 1        2 | read_agfd_tidync
─────────────────────────────────────────────────────────────────────────────────────────────────────────
Error (test-read_agfd_tidync.R:8:3): read_agfd_tidync() returns a tidync object
<subscriptOutOfBoundsError/error/condition>
Error in `x[[1]]`: subscript out of bounds
Backtrace:
    ▆
 1. └─testthat::expect_s3_class(x[[1]], "tidync") at test-read_agfd_tidync.R:8:3
 2.   └─testthat::quasi_label(enquo(object), arg = "object")
 3.     └─rlang::eval_bare(expr, quo_get_env(quo))
─────────────────────────────────────────────────────────────────────────────────────────────────────────
✔ |          3 | read_estimates_by_performance_category
✔ |          3 | read_estimates_by_size [1.7s]
✔ |          3 | read_historical_forecast_database [1.4s]
✔ |          3 | read_historical_national_estimates
✔ |          3 | read_historical_regional_estimates [2.2s]
✔ |          3 | read_historical_state_estimates [2.5s]
✔ |          2 | read_soil_thickness_stars
✔ |          2 | read_soil_thickness_terra
✔ |          2 | retry_download
✔ |         12 | get_agfd [908.6s]

══ Results ══════════════════════════════════════════════════════════════════════════════════════════════
Duration: 931.3 s

── Failed tests ─────────────────────────────────────────────────────────────────────────────────────────
Failure (test-read_agfd_dt.R:8:3): read_agfd_dt returns a data.table object
Names of `x` ('lat', 'lon') don't match 'id', 'farmno', 'R_total_hat_ha', 'C_total_hat_ha', 'FBP_fci_hat_ha', 'FBP_fbp_hat_ha', 'A_wheat_hat_ha', 'H_wheat_dot_hat', 'A_barley_hat_ha', 'H_barley_dot_hat', 'A_sorghum_hat_ha', 'H_sorghum_dot_hat', 'A_oilseeds_hat_ha', 'H_oilseeds_dot_hat', 'R_wheat_hat_ha', 'R_sorghum_hat_ha', 'R_oilseeds_hat_ha', 'R_barley_hat_ha', 'Q_wheat_hat_ha', 'Q_barley_hat_ha', 'Q_sorghum_hat_ha', 'Q_oilseeds_hat_ha', 'S_wheat_cl_hat_ha', 'S_sheep_cl_hat_ha', 'S_sheep_births_hat_ha', 'S_sheep_deaths_hat_ha', 'S_beef_cl_hat_ha', 'S_beef_births_hat_ha', 'S_beef_deaths_hat_ha', 'Q_beef_hat_ha', 'Q_sheep_hat_ha', 'Q_lamb_hat_ha', 'R_beef_hat_ha', 'R_sheep_hat_ha', 'R_lamb_hat_ha', 'C_fodder_hat_ha', 'C_fert_hat_ha', 'C_fuel_hat_ha', 'C_chem_hat_ha', 'A_total_cropped_ha', 'FBP_pfe_hat_ha', 'farmland_per_cell', 'lon', 'lat'

Failure (test-read_agfd_dt.R:58:3): read_agfd_dt returns a data.table object
lapply(x, typeof) (`actual`) not identical to list(...) (`expected`).

`actual` is length 2
`expected` is length 44

    names(actual) | names(expected)
[1] "lat"         - "id"               [1]
[2] "lon"         - "farmno"           [2]
                  - "R_total_hat_ha"   [3]
                  - "C_total_hat_ha"   [4]
                  - "FBP_fci_hat_ha"   [5]
                  - "FBP_fbp_hat_ha"   [6]
                  - "A_wheat_hat_ha"   [7]
                  - "H_wheat_dot_hat"  [8]
                  - "A_barley_hat_ha"  [9]
                  - "H_barley_dot_hat" [10]
... ...             ...                and 34 more ...

`actual$id` is absent
`expected$id` is a character vector ('character')

`actual$farmno` is absent
`expected$farmno` is a character vector ('double')

`actual$R_total_hat_ha` is absent
`expected$R_total_hat_ha` is a character vector ('double')

`actual$C_total_hat_ha` is absent
`expected$C_total_hat_ha` is a character vector ('double')

`actual$FBP_fci_hat_ha` is absent
`expected$FBP_fci_hat_ha` is a character vector ('double')

`actual$FBP_fbp_hat_ha` is absent
`expected$FBP_fbp_hat_ha` is a character vector ('double')

`actual$A_wheat_hat_ha` is absent
`expected$A_wheat_hat_ha` is a character vector ('double')

`actual$H_wheat_dot_hat` is absent
`expected$H_wheat_dot_hat` is a character vector ('double')

`actual$A_barley_hat_ha` is absent
`expected$A_barley_hat_ha` is a character vector ('double')

`actual$H_barley_dot_hat` is absent
`expected$H_barley_dot_hat` is a character vector ('double')

`actual$A_sorghum_hat_ha` is absent
`expected$A_sorghum_hat_ha` is a character vector ('double')

`actual$H_sorghum_dot_hat` is absent
`expected$H_sorghum_dot_hat` is a character vector ('double')

`actual$A_oilseeds_hat_ha` is absent
`expected$A_oilseeds_hat_ha` is a character vector ('double')

`actual$H_oilseeds_dot_hat` is absent
`expected$H_oilseeds_dot_hat` is a character vector ('double')

`actual$R_wheat_hat_ha` is absent
`expected$R_wheat_hat_ha` is a character vector ('double')

`actual$R_sorghum_hat_ha` is absent
`expected$R_sorghum_hat_ha` is a character vector ('double')

`actual$R_oilseeds_hat_ha` is absent
`expected$R_oilseeds_hat_ha` is a character vector ('double')

`actual$R_barley_hat_ha` is absent
`expected$R_barley_hat_ha` is a character vector ('double')

`actual$Q_wheat_hat_ha` is absent
`expected$Q_wheat_hat_ha` is a character vector ('double')

`actual$Q_barley_hat_ha` is absent
`expected$Q_barley_hat_ha` is a character vector ('double')

`actual$Q_sorghum_hat_ha` is absent
`expected$Q_sorghum_hat_ha` is a character vector ('double')

`actual$Q_oilseeds_hat_ha` is absent
`expected$Q_oilseeds_hat_ha` is a character vector ('double')

`actual$S_wheat_cl_hat_ha` is absent
`expected$S_wheat_cl_hat_ha` is a character vector ('double')

`actual$S_sheep_cl_hat_ha` is absent
`expected$S_sheep_cl_hat_ha` is a character vector ('double')

`actual$S_sheep_births_hat_ha` is absent
`expected$S_sheep_births_hat_ha` is a character vector ('double')

`actual$S_sheep_deaths_hat_ha` is absent
`expected$S_sheep_deaths_hat_ha` is a character vector ('double')

`actual$S_beef_cl_hat_ha` is absent
`expected$S_beef_cl_hat_ha` is a character vector ('double')

`actual$S_beef_births_hat_ha` is absent
`expected$S_beef_births_hat_ha` is a character vector ('double')

`actual$S_beef_deaths_hat_ha` is absent
`expected$S_beef_deaths_hat_ha` is a character vector ('double')

`actual$Q_beef_hat_ha` is absent
`expected$Q_beef_hat_ha` is a character vector ('double')

`actual$Q_sheep_hat_ha` is absent
`expected$Q_sheep_hat_ha` is a character vector ('double')

`actual$Q_lamb_hat_ha` is absent
`expected$Q_lamb_hat_ha` is a character vector ('double')

`actual$R_beef_hat_ha` is absent
`expected$R_beef_hat_ha` is a character vector ('double')

`actual$R_sheep_hat_ha` is absent
`expected$R_sheep_hat_ha` is a character vector ('double')

`actual$R_lamb_hat_ha` is absent
`expected$R_lamb_hat_ha` is a character vector ('double')

`actual$C_fodder_hat_ha` is absent
`expected$C_fodder_hat_ha` is a character vector ('double')

`actual$C_fert_hat_ha` is absent
`expected$C_fert_hat_ha` is a character vector ('double')

`actual$C_fuel_hat_ha` is absent
`expected$C_fuel_hat_ha` is a character vector ('double')

`actual$C_chem_hat_ha` is absent
`expected$C_chem_hat_ha` is a character vector ('double')

`actual$A_total_cropped_ha` is absent
`expected$A_total_cropped_ha` is a character vector ('double')

`actual$FBP_pfe_hat_ha` is absent
`expected$FBP_pfe_hat_ha` is a character vector ('double')

`actual$farmland_per_cell` is absent
`expected$farmland_per_cell` is a character vector ('double')

Error (test-read_agfd_stars.R:4:3): read_agfd_stars() returns a stars object
Error in `if (nchar(x) < 1) stop("NetCDF source cannot be empty string")`: missing value where TRUE/FALSE needed
Backtrace:
    ▆
 1. └─read.abares::read_agfd_stars(get_agfd()) at test-read_agfd_stars.R:4:3
 2.   └─stars::read_ncdf(files[1], var = var) at read.abares/R/read_agfd_stars.R:78:3
 3.     ├─stars:::.fix_meta(ncmeta::nc_meta(x))
 4.     ├─ncmeta::nc_meta(x)
 5.     └─ncmeta:::nc_meta.character(x)

Error (test-read_agfd_terra.R:8:3): read_agfd_terra() returns a terra object
<subscriptOutOfBoundsError/error/condition>
Error in `x[[1]]`: subscript out of bounds
Backtrace:
    ▆
 1. └─testthat::expect_s4_class(x[[1]], "SpatRaster") at test-read_agfd_terra.R:8:3
 2.   └─testthat::quasi_label(enquo(object), arg = "object")
 3.     └─rlang::eval_bare(expr, quo_get_env(quo))

Error (test-read_agfd_tidync.R:8:3): read_agfd_tidync() returns a tidync object
<subscriptOutOfBoundsError/error/condition>
Error in `x[[1]]`: subscript out of bounds
Backtrace:
    ▆
 1. └─testthat::expect_s3_class(x[[1]], "tidync") at test-read_agfd_tidync.R:8:3
 2.   └─testthat::quasi_label(enquo(object), arg = "object")
 3.     └─rlang::eval_bare(expr, quo_get_env(quo))

[ FAIL 5 | WARN 0 | SKIP 0 | PASS 75 ]

```

</p></details>
