### 2025-03-27 #667 @adamhsparks
I uh. I have no words. Maybe I'll try a different version of R in a clean environment locally.
