### 2025-03-27 #667 @adamhsparks
something isn't right creating the cache, somewhere. I just ran it and it checked out on my work computer. I cleared the cache and now it's failing again with the same errors as last night on my personal computer.
