### 2025-03-27 e1e7942
fixes several issues with caching in `get_agfd()`
