### 2025-03-27 bf0e968
consolidate and fix agfd tests
