### 2025-03-27 6d3d8c3
aggregate soil thickness tests for speed
