### 2025-03-28 #667 @adamhsparks
Ok, @obrl-soil, I have made another attempt to wrangle this thing into submission. It may be passing tests now. I've ensured that any caching occurs in `tempdir()` and that it's created and cleaned up for each test that uses it to avoid polluting the environment leading to the issues that we were seeing.

Can you please perform another editor check? 🤞
