### 2025-03-28 #667 @adamhsparks
Code coverage is now [99.89%](https://app.codecov.io/gh/adamhsparks/read.abares), the bot reports it as too low because I skip most tests on CI due to connectivity issues or time it takes to run the tests.
