### 2025-03-29 #667 @obrl-soil
```

🐝 Your tests are the bee's knees 🐝
```
Good stuff!

---

Readme page:
- Justification: consider adding a little info to account for international users and students new to the field - start with a few sentences about who DAFF/ABARES are and what they do in the Aus ag space. Why would you go to this provider for this kind of data rather than somewhere else? "There's nowhere else" is fine as an answer, but its good to make that clear if that's the case.
- Flow: recommend starting with clear statements about what the package is intended to help with, then how it provides data, and only talk about what's currently not included and how to request changes at the end of the intro.

Vignette:
- The title should be more clear - 'Advanced uses for read.abares' or similar. At present the simple title may suggest that its an overall package intro vignette, and its a bit more specific than that.
- Totally optional: the `head()` calls to the larger tables could be replaced with `DT()` usage for prettier outputs.

Function documentation:
- There are a few places where roxygen tags can maybe be improved, e.g. using `\doi{}` for referencing papers like Hughes et al 2022, `\href{}{}` or `\url{}` for other URLs, `\code{}` in general rather than backticks for code-highlighting, using `\code{\link{<function}}` to crossref to other functions in the same package etc.
- Example code for `get_agfd()` and friends needs to be consistent - ideally create a persistent object with `x <- ...` and include `head()` and `plot()` example calls in either all or none
- the soil thickness family of functions might be better named as topsoil thickness, as the dataset only estimates the A horizon thickness in mm

Functionality:
- The `plot()` call for the `terra` version of the soil thickness dataset is interpreting the data as categorical, hence the unwieldy legend. This comes from the raster attribute table embedded in the underlying FileGDB dataset, which doesn't contain any extra useful information. You could maybe remove it in `read_soil_thickness_terra()` with `levels(x) <- NULL`, or recast the data to a RAT-less TIF in the cache when `get_soil_thickness()` is run.

other:
- `covr::package_coverage()` is reporting low coverage for me locally, so not sure if we need to work on replicating your own results posted above?
- `covr::package_coverage(type = 'all')` caught a typo in your example for internal function `.retry_download()`; the function call in the example is missing the leading `.` (see `internal_functions.R:43`)

I'll start looking for reviewers :)
