+++ man/read_aagis_regions.Rd
@@ -22,7 +22,7 @@ An \CRANpkg{sf} object of the \acronym{AAGIS} regions.
 }
 \description{
 Download, cache and import the Australian Agricultural and Grazing
-Industries Survey (\acronym{AAGIS} regions geospatial shapefile. Upon
+Industries Survey (\acronym{AAGIS}) regions geospatial shapefile. Upon
 import, the geometries are automatically corrected to fix invalid
 geometries that are present in the original shapefile.
 }
