### 2025-03-29 dd2da73
address @obrl-soil's excellent suggestions for improving the README
