+++ README.Rmd
@@ -22,10 +22,20 @@ knitr::opts_chunk$set(
 [![codecov](https://codecov.io/gh/adamhsparks/read.abares/graph/badge.svg?token=1ZPgFWJ0B2)](https://codecov.io/gh/adamhsparks/read.abares)
 <!-- badges: end -->
 
-An R package for automated downloading and ingestion of data from the Australian Bureau of Agricultural and Resource Economics and Sciences.
-Not all ABARES data are serviced.
-The list is hand-picked to be reasonably useful and maintainable, _i.e._, frequently updated values are not included in this, _e.g._, [Australian crop reports](https://daff.ent.sirsidynix.net.au/client/en_AU/ABARES/search/results?te=ASSET&st=PD#).
-However, if there is a data set that you feel would be useful to be serviced by {read.abares}, please feel free to [open an issue](https://github.com/adamhsparks/read.abares/issues/new) with details about the data set or better yet, open a pull request!
+An R package for automated downloading and ingestion of data from the Australian Bureau of Agricultural and Resource Economics and Sciences (ABARES).
+
+## About ABARES
+
+ABARES is the research arm of the Australian federal Government Department of Agriculture, Fisheries and Forestry (DAFF).
+ABARES' main role is "to provide "professionally independent data, research, analysis and advice that informs public and private decisions affecting Australian agriculture, fisheries and forestry"^[1].
+
+The data provided by ABARES is extensive and varied, and includes data on agricultural production, trade, and forecasts, as well as spatial data on farm locations and top-soil thickness.
+ABARES data are not available from other sources but are mostly available under Creative Commons Licences for reuse.
+
+## About {read.abares}: why this package?
+
+ABARES makes several data sets freely available as spreadsheets in Microsoft Excel or CSV file formats and zip archives of geospatial data as NetCDF or shape files.
+{read.abares} facilitates downloading, caching and importing these files in your R session as a {[data.table](https://cran.r-project.org/package=data.table)} object for spreadsheet data, {[sf](https://cran.r-project.org/package=sf)} for shape file data and your choice of the class of the resulting object(s) for the NetCDF data.
 
 Data serviced include:
 
@@ -38,8 +48,6 @@ Data serviced include:
   - the trade data and;
   - the trade region data.
 
-The files are freely available as CSV files, zip archives of NetCDF files or a zip archives of geospatial shape files.
-{read.abares} facilitates downloading, caching and importing these files in your R session with your choice of the class of the resulting object(s).
 
 ## Get Started
 
@@ -55,24 +63,36 @@ pak::pak("adamhsparks/read.abares")
 
 ## Features
 
-### Caching
+### Standardised Column Names and Orders
+
+ABARES spreadhsheet data are not always consistent in their column names or orders.
+{read.abares} standardises the column names and orders for the data sets it services to help you do your work more efficiently.
+
+### Automated Repairing of Geospatial Data
+
+The Australian Agricultural and Grazing Industries Survey (AAGIS) region mapping files report geometry errors that can be repaired using the `sf::st_make_valid()` function.
+{read.abaes} automatically repairs these geometries for you when you import the data.
+
+### File Caching
 
 {read.abares} supports caching the files using `tools::R_user_dir(package = "read.abares", which = "cache")` to save the files in a standardised location across platforms so you don't have to worry about where the files went or if they're still there.
-When requesting the files, {read.abares} will first check if they are available locally either in cached or temporary storage.
+When requesting the files {read.abares} will first check if they are available locally either in cached or temporary storage.
[…]
