+++ README.md
@@ -1,100 +1,151 @@
----
-output: github_document
----
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
-
-
 
 # {read.abares}: Simple downloading and importing of ABARES Data <img src="man/figures/logo.png" align="right"/>
 
 <!-- badges: start -->
-[![Lifecycle: stable](https://img.shields.io/badge/lifecycle-stable-green.svg)](https://lifecycle.r-lib.org/articles/stages.html#stable)
-[![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
+
+[![Lifecycle:
+stable](https://img.shields.io/badge/lifecycle-stable-green.svg)](https://lifecycle.r-lib.org/articles/stages.html#stable)
+[![Project Status: Active – The project has reached a stable, usable
+state and is being actively
+developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
 [![R-CMD-check](https://github.com/adamhsparks/read.abares/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/adamhsparks/read.abares/actions/workflows/R-CMD-check.yaml)
 [![codecov](https://codecov.io/gh/adamhsparks/read.abares/graph/badge.svg?token=1ZPgFWJ0B2)](https://codecov.io/gh/adamhsparks/read.abares)
 <!-- badges: end -->
 
-An R package for automated downloading and ingestion of data from the Australian Bureau of Agricultural and Resource Economics and Sciences.
-Not all ABARES data are serviced.
-The list is hand-picked to be reasonably useful and maintainable, _i.e._, frequently updated values are not included in this, _e.g._, [Australian crop reports](https://daff.ent.sirsidynix.net.au/client/en_AU/ABARES/search/results?te=ASSET&st=PD#).
-However, if there is a data set that you feel would be useful to be serviced by {read.abares}, please feel free to [open an issue](https://github.com/adamhsparks/read.abares/issues/new) with details about the data set or better yet, open a pull request!
+An R package for automated downloading and ingestion of data from the
+Australian Bureau of Agricultural and Resource Economics and Sciences
+(ABARES).
+
+## About ABARES
+
+ABARES is the research arm of the Australian federal Government
+Department of Agriculture, Fisheries and Forestry (DAFF). ABARES’ main
+role is “to provide”professionally independent data, research, analysis
+and advice that informs public and private decisions affecting
+Australian agriculture, fisheries and forestry”[^1].
+
+The data provided by ABARES is extensive and varied, and includes data
+on agricultural production, trade, and forecasts, as well as spatial
+data on farm locations and top-soil thickness. ABARES data are not
+available from other sources but are mostly available under Creative
+Commons Licences for reuse.
+
+## About {read.abares}: why this package?
+
+ABARES makes several data sets freely available as spreadsheets in
+Microsoft Excel or CSV file formats and zip archives of geospatial data
+as NetCDF or shape files. {read.abares} facilitates downloading, caching
+and importing these files in your R session as a
+{[data.table](https://cran.r-project.org/package=data.table)} object for
+spreadsheet data, {[sf](https://cran.r-project.org/package=sf)} for
+shape file data and your choice of the class of the resulting object(s)
+for the NetCDF data.
 
 Data serviced include:
 
-- the [ABARES Estimates](https://www.agriculture.gov.au/abares/data/farm-data-portal#data-download);
-- the [Australian Gridded Farm Data (AGFD) set](https://www.agriculture.gov.au/abares/research-topics/surveys/farm-survey-data/australian-gridded-farm-data);
[…]
