### 2025-03-29 9844ce9
fix descriptions of multiclasses for geo data
