### 2025-03-29 4b1042e
clarify testing and caching
