### 2025-03-29 0d5de35
add data.table to AGFD data formats
