### 2025-03-29 4c69749
tidy up lists
