### 2025-03-29 e8fa0e8
edits for clarity and functionality
