### 2025-03-29 39321fb
state that columns are properly formatted for data types
