+++ README.md
@@ -79,7 +79,10 @@ pak::pak("adamhsparks/read.abares")
 ABARES spreadhsheet data are not always consistent in their column names
 or orders. {read.abares} standardises the column names and orders and
 uses snake_case for all colnames with the first letter capitalised of
-every column to help you do your work more efficiently.
+every column to help you do your work more efficiently. Columns are
+formatted correctly for the data type, *e.g.*, dates are converted to
+`Date` class, and numbers are converted to `numeric` class where
+necessary.
 
 ### Automated Repairing of Geospatial Data
 
