+++ README.Rmd
@@ -67,7 +67,7 @@ pak::pak("adamhsparks/read.abares")
 
 ABARES spreadhsheet data are not always consistent in their column names or orders.
 {read.abares} standardises the column names and orders and uses snake_case for all colnames with the first letter capitalised of every column to help you do your work more efficiently.
-Columns are formatted correctly for the data type, _e.g._, dates are converted to `Date` class, and numbers are converted to `numeric` class where necessary.
+Columns are formatted correctly for the data type, _e.g._, dates are converted to `Date` class, and numbers are converted to `numeric` class where necessary, etc.
 
 ### Automated Repairing of Geospatial Data
 
