### 2025-03-29 e39ffbe
use \doi{}
