+++ R/get_agfd.R
@@ -157,7 +157,7 @@
 #'
 #' *Australian gridded farm data*, Australian Bureau of Agricultural and
 #'  Resource Economics and Sciences, Canberra, July 2024, DOI:
-#'  [10.25814/7n6z-ev41](https://doi.org/10.25814/7n6z-ev41).
+#'  \doi{https://doi.org/10.25814/7n6z-ev41}.
 #'  [CC BY 4.0](https://creativecommons.org/licenses/by/4.0/legalcode).
 #'
 #' N. Hughes, W.Y. Soh, C. Boult, K. Lawson, *Defining drought from
