### 2025-03-29 bbf353d
fix another DOI tag and redocument
