### 2025-03-30 8f60481
fixes legend in soilthickness plot for terra
