### 2025-03-30 511730d
polish documentation and standardise internal fns to use {fs}
