+++ R/get_agfd.R
@@ -13,7 +13,7 @@
 #'  These data have been produced by \acronym{ABARES} as part of the ongoing
 #'  Australian Agricultural Drought Indicator (\acronym{AADI}) project
 #'  (previously known as the Drought Early Warning System Project) and were
-#'  derived using \acronym{ABARES}'s
+#'  derived using \acronym{ABARES}
 #'  [*farmpredict*](https://www.agriculture.gov.au/abares/research-topics/climate/drought/farmpredict)
 #'  model, which in turn is based on ABARES Agricultural and Grazing Industries
 #'  Survey (\acronym{AAGIS}) data.\cr
