+++ man/read_agfd_tidync.Rd
@@ -10,7 +10,7 @@ read_agfd_tidync(files)
 \item{files}{A list of \acronym{AGFD} NetCDF files to import.}
 }
 \value{
-a \code{list} object of \CRANpkg{tidync} objects of the Australian
+A \code{list} object of \CRANpkg{tidync} objects of the Australian
 Gridded Farm Data with the file names as the  list's objects' names.
 }
 \description{
