+++ man/get_agfd.Rd
@@ -181,6 +181,9 @@ farmland_per_cell     \tab ha        \tab Indicative area of farmland in the gri
 
 \examples{
 \dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
+agfd <- get_agfd()
+
+agfd
 \dontshow{\}) # examplesIf}
 }
 \references{
