### 2025-03-30 7efff5e
fix more DOIs
