+++ README.Rmd
@@ -65,14 +65,14 @@ pak::pak("adamhsparks/read.abares")
 
 ### Standardised Column Names and Orders
 
-ABARES spreadhsheet data are not always consistent in their column names or orders.
+ABARES spreadsheet data are not always consistent in their column names or orders.
 {read.abares} standardises the column names and orders and uses snake_case for all colnames with the first letter capitalised of every column to help you do your work more efficiently.
 Columns are formatted correctly for the data type, _e.g._, dates are converted to `Date` class, and numbers are converted to `numeric` class where necessary, etc.
 
 ### Automated Repairing of Geospatial Data
 
 The Australian Agricultural and Grazing Industries Survey (AAGIS) region mapping files report geometry errors that can be repaired using the `sf::st_make_valid()` function.
-{read.abaes} automatically repairs these geometries for you when you import the data.
+{read.abares} automatically repairs these geometries for you when you import the data.
 
 ### File Caching
 
