### 2025-03-30 66c4e5d
fix a TODO and update documentation
