+++ tests/testthat/test-topsoil_thickness_functions.R
@@ -5,14 +5,14 @@ withr::local_envvar(R_USER_CACHE_DIR = tempdir())
 
 # without caching enabled ----
 
-test_that("get_soil_thickness doesn't cache", {
+test_that("get_topsoil_thickness doesn't cache", {
   skip_if_offline()
   skip_on_ci()
-  x <- get_soil_thickness(cache = FALSE)
-  expect_s3_class(x, c("read.abares.soil.thickness", "list"))
+  x <- get_topsoil_thickness(cache = FALSE)
+  expect_s3_class(x, c("read.abares.topsoil.thickness", "list"))
   expect_false(fs::dir_exists(fs::path(
     .find_user_cache(),
-    "soil_thickness_dir"
+    "topsoil_thickness_dir"
   )))
   # cleanup on the way out for the next test
   expect_message(clear_cache())
@@ -21,51 +21,51 @@ test_that("get_soil_thickness doesn't cache", {
 
 # with caching enabled after is was not initially enabled ----
 
-test_that("get_soil_thickness caches", {
+test_that("get_topsoil_thickness caches", {
   skip_if_offline()
   skip_on_ci()
-  x <- get_soil_thickness(cache = TRUE)
-  expect_s3_class(x, c("read.abares.soil.thickness", "list"))
+  x <- get_topsoil_thickness(cache = TRUE)
+  expect_s3_class(x, c("read.abares.topsoil.thickness", "list"))
   expect_true(fs::file_exists(fs::path(
     .find_user_cache(),
-    "soil_thickness_dir"
+    "topsoil_thickness_dir"
   )))
   expect_no_message(clear_cache())
 })
 
-test_that("get_soil_thickness does cache", {
+test_that("get_topsoil_thickness does cache", {
   skip_if_offline()
   skip_on_ci()
-  x <- get_soil_thickness(cache = TRUE)
-  expect_s3_class(x, c("read.abares.soil.thickness", "list"))
+  x <- get_topsoil_thickness(cache = TRUE)
+  expect_s3_class(x, c("read.abares.topsoil.thickness", "list"))
   expect_true(fs::file_exists(fs::path(
     .find_user_cache(),
-    "soil_thickness_dir"
+    "topsoil_thickness_dir"
   )))
   expect_true(fs::dir_exists(fs::path(
     .find_user_cache(),
-    "soil_thickness_dir"
+    "topsoil_thickness_dir"
   )))
 })
 
 # test reading with stars ----
 
-test_that("read_soil_thickness_stars returns a stars object", {
+test_that("read_topsoil_thickness_stars returns a stars object", {
   skip_if_offline()
   skip_on_ci()
-  x <- get_soil_thickness(cache = TRUE) |>
-    read_soil_thickness_stars()
+  x <- get_topsoil_thickness(cache = TRUE) |>
+    read_topsoil_thickness_stars()
   expect_s3_class(x, "stars")
   expect_named(x, "thpk_1")
 })
 
 # test reading with terra ----
 
-test_that("read_soil_thickness_stars returns a terra object", {
+test_that("read_topsoil_thickness_stars returns a terra object", {
   skip_if_offline()
   skip_on_ci()
-  x <- get_soil_thickness(cache = TRUE) |>
-    read_soil_thickness_terra()
+  x <- get_topsoil_thickness(cache = TRUE) |>
+    read_topsoil_thickness_terra()
   expect_s4_class(x, "SpatRaster")
   expect_named(x, "thpk_1")
 })
@@ -106,18 +106,18 @@ test_that("print.read.abares.thickness.files prints metadata", {
     )
     cli::cli_text(
       "To see the full metadata, call
-    {.fn print_soil_thickness_metadata} on a soil thickness object in your R
[…]
