+++ man/read_topsoil_thickness_stars.Rd
@@ -1,16 +1,16 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/read_soil_thickness_stars.R
-\name{read_soil_thickness_stars}
-\alias{read_soil_thickness_stars}
+\name{read_topsoil_thickness_stars}
+\alias{read_topsoil_thickness_stars}
 \title{Read a 'Soil Thickness for Australian Areas of Intensive Agriculture of Layer 1' file with stars}
 \source{
 \url{https://anrdl-integration-web-catalog-saxfirxkxt.s3-ap-southeast-2.amazonaws.com/warehouse/staiar9cl__059/staiar9cl__05911a01eg_geo___.zip}
 }
 \usage{
-read_soil_thickness_stars(files)
+read_topsoil_thickness_stars(files)
 }
 \arguments{
-\item{files}{An \pkg{read.abares} \code{read.abares.soil.thickness} object, a
+\item{files}{An \pkg{read.abares} \code{read.abares.topsoil.thickness} object, a
 \code{list} that contains the \acronym{ESRI} grid file to import.}
 }
 \value{
@@ -23,8 +23,8 @@ data as a \CRANpkg{stars} object.
 }
 \examples{
 \dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
-st_stars <- get_soil_thickness(cache = TRUE) |>
-  read_soil_thickness_stars()
+st_stars <- get_topsoil_thickness(cache = TRUE) |>
+  read_topsoil_thickness_stars()
 
 plot(st_stars)
 \dontshow{\}) # examplesIf}
@@ -33,9 +33,9 @@ plot(st_stars)
 \url{https://data.agriculture.gov.au/geonetwork/srv/eng/catalog.search#/metadata/faa9f157-8e17-4b23-b6a7-37eb7920ead6}
 }
 \seealso{
-Other soil_thickness: 
-\code{\link{get_soil_thickness}()},
-\code{\link{print_soil_thickness_metadata}()},
-\code{\link{read_soil_thickness_terra}()}
+Other topsoil_thickness: 
+\code{\link{get_topsoil_thickness}()},
+\code{\link{print_topsoil_thickness_metadata}()},
+\code{\link{read_topsoil_thickness_terra}()}
 }
-\concept{soil_thickness}
+\concept{topsoil_thickness}
