+++ man/dot-download_topsoil_thickness.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/get_soil_thickness.R
-\name{.download_soil_thickness}
-\alias{.download_soil_thickness}
+\name{.download_topsoil_thickness}
+\alias{.download_topsoil_thickness}
 \title{Downloads soil thickness data if not already found locally}
 \usage{
-.download_soil_thickness(cache)
+.download_topsoil_thickness(cache)
 }
 \arguments{
 \item{cache}{\code{Boolean} Cache the soil thickness data files after download

[…]
