### 2025-03-30 3dd7749
reknit vignette with updated functionality
