+++ vignettes/read.abares.Rmd
@@ -55,16 +55,16 @@ Check the file format information for the NetCDF files.
 
 ``` r
 print_agfd_nc_file_format()
-#> ──────────────────────────────────────────────────────────────────────────
-#> Each of the layers in simulation output data is represented as a 2D
-#> raster in NETCDF files, with the following grid format:
+#> ────────────────────────────────────────────────────────────────────────────────
+#> Each of the layers in simulation output data is represented as a 2D raster in
+#> NETCDF files, with the following grid format:
 #> CRS: EPSG:4326 - WGS 84 - Geographic
 #> Extent: 111.975 -44.525 156.275 -9.975
 #> Unit: Degrees
 #> Width: 886
 #> Height: 691
 #> Cell size: 0.05 degree x 0.05 degree
-#> ──────────────────────────────────────────────────────────────────────────
+#> ────────────────────────────────────────────────────────────────────────────────
 #> For further details, see the ABARES website,
 #> <https://www.agriculture.gov.au/abares/research-topics/surveys/farm-survey-data/australian-gridded-farm-data>
 ```
@@ -84,20 +84,20 @@ star <- get_agfd(cache = TRUE) |>
 head(star[[1]])
 #> stars object with 2 dimensions and 6 attributes
 #> attribute(s):
-#>                         Min.         1st Qu.         Median
-#> farmno          15612.000000 233091.50000000 329567.0000000
-#> R_total_hat_ha      2.954396      7.88312157     21.7520529
-#> C_total_hat_ha      1.304440      4.34079101      9.9449849
-#> FBP_fci_hat_ha   -143.759785      3.60529967     11.5796641
-#> FBP_fbp_hat_ha   -349.521639      3.36599833     11.5074294
-#> A_wheat_hat_ha      0.000000      0.04062786      0.1114289
-#>                           Mean        3rd Qu.           Max.   NA's
-#> farmno          324737.7187618 418508.5000000 669706.0000000 443899
-#> R_total_hat_ha     169.5139301    174.8553843   2415.7556059 443899
-#> C_total_hat_ha      93.2210542     95.7221857   1853.5385298 443899
-#> FBP_fci_hat_ha      76.2928759     77.6748501   1186.5830232 443899
-#> FBP_fbp_hat_ha      60.0750936     62.8596117   1240.6003218 443899
-#> A_wheat_hat_ha       0.1365683      0.2112845      0.5047761 565224
+#>                         Min.         1st Qu.         Median           Mean
+#> farmno          15612.000000 233091.50000000 329567.0000000 324737.7187618
+#> R_total_hat_ha      2.954396      7.88312157     21.7520529    169.5139301
+#> C_total_hat_ha      1.304440      4.34079101      9.9449849     93.2210542
+#> FBP_fci_hat_ha   -143.759785      3.60529967     11.5796641     76.2928759
+#> FBP_fbp_hat_ha   -349.521639      3.36599833     11.5074294     60.0750936
+#> A_wheat_hat_ha      0.000000      0.04062786      0.1114289      0.1365683
+#>                        3rd Qu.           Max.   NA's
+#> farmno          418508.5000000 669706.0000000 443899
+#> R_total_hat_ha     174.8553843   2415.7556059 443899
+#> C_total_hat_ha      95.7221857   1853.5385298 443899
+#> FBP_fci_hat_ha      77.6748501   1186.5830232 443899
+#> FBP_fbp_hat_ha      62.8596117   1240.6003218 443899
+#> A_wheat_hat_ha       0.2112845      0.5047761 565224
 #> dimension(s):
 #>     from  to         refsys              values x/y
[…]
