### 2025-03-30 145749c
try rejigging precompile.R to work as expected
