+++ vignettes/precompile.R
@@ -1,25 +1,17 @@
-
 # vignettes that depend on Internet access need to be precompiled
 library(knitr)
 library(here)
 
 # read.abares vignette
-knit(input = "vignettes/read.abares.Rmd.orig",
-     output = "vignettes/read.abares.Rmd")
-purl("vignettes/read.abares.Rmd.orig",
-     output = "vignettes/read.abares.R")
+knit(
+  input = "vignettes/read.abares.Rmd.orig",
+  output = "vignettes/read.abares.Rmd"
+)
+purl("vignettes/read.abares.Rmd.orig", output = "vignettes/read.abares.R")
 
-# move image files
-figs <-
-  fs::dir_ls(here("figure/"),
-             pattern = ".png$",
-             full.names = TRUE)
-file.copy(from = figs,
-          to = paste0(here("vignettes/"),
-                      basename(figs)),
-          overwrite = TRUE)
-file.remove(figs)
-file.remove(here("figure"))
+# Move figures into vignettes/ folder
+figs <- fs::dir_ls(glob = "vigfig-")
+fs::file_move(figs, fs::path("vignettes/", figs))
 
 # remove fs::path_file such that vignettes will build with figures
 ## read.abares vignette
@@ -36,9 +28,5 @@ build_vignettes()
 
 # move resource files (images) to /doc
 resources <-
-  fs::dir_ls(here("vignettes/"),
-             pattern = ".png$",
-             full.names = TRUE)
-file.copy(from = resources,
-          to = here("doc"),
-          overwrite =  TRUE)
+  fs::dir_ls(here("vignettes/"), glob = ".png$")
+file.copy(from = resources, to = here("doc"), overwrite = TRUE)
