### 2025-04-13 #667 @obrl-soil
Hey sorry I've been away on fieldwork - here's what I get with the same call

```
read.abares Coverage: 12.26%
R/get_soil_thickness.R: 0.00%
R/read_aagis_regions.R: 0.00%
R/read_abares_trade.R: 0.00%
R/read_abares_trade_regions.R: 0.00%
R/read_estimates_by_performance_category.R: 0.00%
R/read_estimates_by_size.R: 0.00%
R/read_historical_forecast_database.R: 0.00%
R/read_historical_national_estimates.R: 0.00%
R/read_historical_regional_estimates.R: 0.00%
R/read_historical_state_estimates.R: 0.00%
R/read_soil_thickness_stars.R: 0.00%
R/read_soil_thickness_terra.R: 0.00%
R/read_agfd_stars.R: 1.69%
R/read_agfd_dt.R: 8.33%
R/read_agfd_terra.R: 25.00%
R/read_agfd_tidync.R: 25.00%
R/get_agfd.R: 32.31%
R/clear_cache.R: 100.00%
R/inspect_cache.R: 100.00%
R/internal_functions.R: 100.00%
```

I don't think it matters if its getting checked properly on at least one machine, just a bit curious about what's going on with mine!

To answer your question about the tif vs grid issue - its a more widely accessible format and you get to store a single file object with embedded projection info instead of a folder of smaller files. It'll be pretty much the same size overall. I experimented with LERC compression but didn't see any real advantage since its already a fairly small dataset, so standard TIF or a COG version would work fine.
