### 2025-04-14 832a75a
fix (again) reading the topsoil as a terra object
