### 2025-04-14 #667 @adamhsparks
> Weird - maybe a covr bug then

{test that}, I think. I was just running tests alone
