### 2025-05-04 #667 @maurolepore
@adamhsparks and @obrl-soil I'm wrapping up my EiC rotation today and before I hand this over to the next EiC I want to thank you for your work and patience.
