### 2025-05-06 #667 @potterzot
- **Briefly describe any working relationship you have (had) with the package authors.** Adam Sparks [reviewed my rnassqs package](https://github.com/ropensci/software-review/issues/298) in 2019 on its initial acceptance to rOpenSci.

---

Great work on a package that provides access to a really neat dataset. Tests were very complete, documentation was extensive, and I learned some new things reading through the code.

It doesn't seem like there is a way to query subsets. It also seems like because file names must be directly coded, any changes / updates to the file paths will be package breaking, which is unfortunate. But in both cases I don't think there's anything that can be done on the package end to change this.

#### Specific comments

- [README.md]: I know `pak` is more prevalent these days, but wonder if including `remotes::install_github()` as an option would be useful for folks who don't use pak.
- In the vignette several commands took quite a long time to run. I think the same instructional point could be made with fewer regions/years. I can see that it isn't possible to query specific subsets of data, but before merging and plotting might be helpful to speed up the vignette. I encountered these slow points in particular:
    - [vignette/read.abares.Rmd] Line 80: The command `aagis_dat <- left_join(aagis_regions, aagis_region_data)` took 70 seconds on my laptop.
    - [vignette/read.abares.Rmd] Lines 95-102: The ggplot took a while to render
    - [vignette/read.abares.Rmd] Lines 95-102: This fetch command took at least 3 minutes and made my laptop lag: `star <- get_agfd(cache = TRUE) |> read_agfd_stars()`. Reducing the dimensions here would also make it so taht when a new user / this reviewer types `star` in the terminal they aren't swamped with a lot of printed data that has to be cancelled out of. The same is true for `read_agfd_terra`, `read_agfd_tidync`, and `read_agfd_dt`. If it can't be done via query,
- [DESCRIPTION] When running tests I received two warnings about needing to install `bit64` to properly display integer64 values. Maybe adding this package to `suggests`?
