### 2025-05-06 #667 @adamhsparks
@potterzot, just a quick note of many thanks.

Your comment on {pak} is timely. I've spent yesterday and today wrestling with it for other packages and it's coming up short. I've resorted to favouring {remotes} first with {pak} as an optional method for one of these packages and will implement the same here.

I'll have a detailed response with all the changes that follows.
