### 2025-05-30 #667 @mpaulacaldas
---

## Review Comments

Kudos on the package @adamhsparks! It gives access to many interesting data sets and leaves users a lot of flexibility on the geospatial libraries to use.

As you will see below, I had some issues with the downloads of the estimate datasets. I am happy to help you debug further as I imagine some of this might be due to my geographical location/locale.

### Vignette

- Like @potterzot, I find some of the operations in the vignette take a long time to render, notably `plot(aagis_regions)`.  Likewise, in the example in lines 71 to 103, I would suggest filtering the data _before_ the `left_join()`, to reduce the time of the `left_join()`.
- I think users of this package will likely rely quite heavily on the cache, so it may be useful to learn about `inspect_cache()` in the vignette.
- This is a minor note, but `library(read.abares)` is mentioned at the top of each section. Were you planning on having separate vignettes? If you have only one vignette, the library call becomes a bit redundant.

#### Failing downloads

Unfortunately, several functions don't work on my end. I didn't have a lot of time to debug (but I'm happy to help you out with it if you like), but it appears that all estimate downloads get blocked at the level of `.retry_download()`. The functions that don't work for me:

- `read_estimates_by_performance_category()`
- `read_estimates_by_size()`
- `read_historical_national_estimates()`
- `read_historical_state_estimates()`
- `read_historical_regional_estimates()`

Tests for `retry_download()` pass. I also don't have issues with the download of most other assets. I leave below my session info for reference.

<details><summary>Session info</summary>

```r
─ Session info ────────────────────────────────────────────────────────────────────
 os       macOS Sonoma 14.1
 system   aarch64, darwin20
 ui       RStudio
 collate  en_US.UTF-8
 ctype    en_US.UTF-8
 tz       Europe/Paris
 rstudio  2023.09.1+494 Desert Sunflower (desktop)
 pandoc   NA

─ Packages ────────────────────────────────────────────────────────────────────────
 package     * version date (UTC) lib source
 brio          1.1.4   2023-12-10 [1] CRAN (R 4.3.1)
 cachem        1.0.8   2023-05-01 [1] CRAN (R 4.3.0)
 cli           3.6.2   2023-12-11 [1] CRAN (R 4.3.1)
 codetools     0.2-19  2023-02-01 [2] CRAN (R 4.3.1)
 data.table    1.14.8  2023-02-17 [1] CRAN (R 4.3.0)
 devtools    * 2.4.5   2022-10-11 [1] CRAN (R 4.3.0)
 digest        0.6.35  2024-03-11 [1] CRAN (R 4.3.1)
 ellipsis      0.3.2   2021-04-29 [1] CRAN (R 4.3.0)
 fastmap       1.1.1   2023-02-24 [1] CRAN (R 4.3.0)
 fs            1.6.3   2023-07-20 [1] CRAN (R 4.3.0)
 glue          1.7.0   2024-01-09 [1] CRAN (R 4.3.1)
 htmltools     0.5.7   2023-11-03 [1] CRAN (R 4.3.1)
 htmlwidgets   1.6.4   2023-12-06 [1] CRAN (R 4.3.1)
 httpuv        1.6.14  2024-01-26 [1] CRAN (R 4.3.1)
 later         1.3.2   2023-12-06 [1] CRAN (R 4.3.1)
 lifecycle     1.0.4   2023-11-07 [1] CRAN (R 4.3.1)
 magrittr      2.0.3   2022-03-30 [1] CRAN (R 4.3.0)
 memoise       2.0.1   2021-11-26 [1] CRAN (R 4.3.0)
 mime          0.12    2021-09-28 [1] CRAN (R 4.3.0)
 miniUI        0.1.1.1 2018-05-18 [1] CRAN (R 4.3.0)
 pkgbuild      1.4.3   2023-12-10 [1] CRAN (R 4.3.1)
 pkgload       1.3.4   2024-01-16 [1] CRAN (R 4.3.1)
 profvis       0.3.8   2023-05-02 [1] CRAN (R 4.3.0)
 promises      1.2.1   2023-08-10 [1] CRAN (R 4.3.0)
 purrr         1.0.2   2023-08-10 [1] CRAN (R 4.3.0)
 R6            2.5.1   2021-08-19 [1] CRAN (R 4.3.0)
 Rcpp          1.0.12  2024-01-09 [1] CRAN (R 4.3.1)
 read.abares * 1.0.1   2025-05-18 [1] Github (adamhsparks/read.abares@832a75a)
 remotes       2.4.2.1 2023-07-18 [1] CRAN (R 4.3.0)
 rlang         1.1.3   2024-01-10 [1] CRAN (R 4.3.1)
 rstudioapi    0.16.0  2024-03-24 [1] CRAN (R 4.3.1)
 sessioninfo   1.2.2   2021-12-06 [1] CRAN (R 4.3.0)
 shiny         1.8.0   2023-11-17 [1] CRAN (R 4.3.1)
 stringi       1.8.3   2023-12-11 [1] CRAN (R 4.3.1)
 stringr       1.5.1   2023-11-14 [1] CRAN (R 4.3.1)
 terra         1.7-65  2023-12-15 [1] CRAN (R 4.3.1)
 testthat    * 3.2.1   2023-12-02 [1] CRAN (R 4.3.1)
 urlchecker    1.0.1   2021-11-30 [1] CRAN (R 4.3.0)
 usethis     * 2.2.2   2023-07-06 [1] CRAN (R 4.3.0)
 vctrs         0.6.5   2023-12-01 [1] CRAN (R 4.3.1)
 xtable        1.8-4   2019-04-21 [1] CRAN (R 4.3.0)

 [1] /Users/mariapaulacaldas/Library/R/arm64/4.3/library
 [2] /Library/Frameworks/R.framework/Versions/4.3-arm64/Resources/library

───────────────────────────────────────────────────────────────────────────────────
```

#### Error with data.table::fread()

`read_abares_trade()` also fails in my machine, but at the data.table level.
```r
trade <- read_abares_trade()
#> Error in data.table::fread(trade_zip) :
#>  embedded nul in string: 'PK\003\004\024\0\0\0\b\0-#> U\xa7Z\v\021\x82/\vgG\t\x9e\xa0iT\025\0\0\0ABARES_trade_data.csvĽ]\x93d\xb9q\xa6y\xbff\xfb\037\xd2\xfa\x8a\xb2\xadL\x83'
```

#### `inspect_cache()` doesn't work as expected

That is, it doesn't actually return the absolute paths of the files in the cache. This is what I get:

```r
p <- inspect_cache()
#>
#> ── Locally Available {read.abares} Cached Files ─────────────────────────────────────────
#> • aagis_regions_dir
#> • historical_climate_prices_fixed
#> • topsoil_thickness_dir
unclass(p)
#> [1] "cli-69638-30"
```

Looking at the code, I see this is because the function doesn't actually return `f` (the paths) after the cli messages. You could fix this by returning `invisible(f)` after the `cli` messages.

I would also suggest adding a unit test for the _value_ returned by `inspect_cache()`, as the current tests only focus on the cli output.

### Function documentation

- In the examples of `inspect_cache()`, it would be easier if you stick to either `fs` functions (e.g. `fs::path_file()` and `fs::path_dir()`) or to their base equivalent (`basename()` and `dirname()`), instead of mixing the two.

### Package API

Below are some "nice-to-have's", but please don't feel like you have to address them:

- As per the [rOpenSci guidelines](https://devguide.ropensci.org/pkg_building.html#packages-wrapping-web-resources-api-clients), it would be useful to give users the choice to override the user agent for the API requests preformed behind the scenes by `.retry_download()`. In addition of the user agent, I also quite like having the ability to override the `timeout`, so things can "fail fast" if I need them to.
- It would be nice let users choose the location of the cache directory or [to opt out of informative messages](https://devguide.ropensci.org/pkg_building.html#console-messages) via environment variables or global options.
