### 2025-06-05 #667 @adamhsparks
Thank you, @mpaulacaldas for this thorough review. @obrl-soil, I'll start working on this as I have time. I'm a bit busy with some "real" work currently, but shouldn't be long and addressing these comments (aside from ongoing download issues) should be rather straightforward.
