### 2025-12-14 #667 @adamhsparks
@econpotter and @mpaulacaldas, I'd like to acknowledge your contributions to the package as reviewers in the DESCRIPTION file's `authors@R` field. Is this something you would accept?

Also, I've just merged a big set of changes internally. It's more robust now and doesn't unzip files, except for the NetCDF files and therefore doesn't move files around as much and should be faster and easier to maintain.
